### 2026-01-22 #9885 @editorialbot
[PRE REVIEW]: BattMo - Battery Modelling Toolbox
**Submitting author:** @xavierr (<a href="http://orcid.org/0000-0002-4100-3035">Xavier Raynaud</a>)
**Repository:** https://github.com/BattMoTeam/BattMo
**Branch with paper.md** (empty if default branch): joss
**Version:** v04.0
