### 2026-01-22 #9885 @kyleniemeyer
@xavierr - I'm going to reject this submission for BattMo, since we have your original submission: https://github.com/openjournals/joss-reviews/issues/9884
