### 2026-01-25 b60d38e
more in only thermal example
