### 2026-01-26 f1fbe81
wip in runOnlyThermal
