### 2026-01-26 bf5a313
update in runOnlyThermal
