+++ Examples/Advanced/runOnlyThermal.m
@@ -1,17 +1,29 @@
-%% Decoupled electro-chemical and thermal simulation
+%% Comparison between coupled and uncoupled thermal simulation
+%
+% To compute the thermal response of a battery cell, one can either run a fully coupled simulation or run first an
+% isothermal simulation and then use the output to compute the thermal response in a thermal-only simulation.
+%
+% In the second case, we do not get any feedback from the thermal model to the electrochemical model.
+%
+% In this example, we compare the solutions obtained both for the voltage and for the temperature on a P4D model.
+%
+
 
 %% setup material property input
+%
+% We use a lithium-ion battery cell with NMC cathode and graphite anode
+% 
 jsonfilename = fullfile('ParameterData'        , ...
                         'BatteryCellParameters', ...
                         'LithiumIonBatteryCell', ...
                         'lithium_ion_battery_nmc_graphite.json');
 jsonstruct_material = parseBattmoJson(jsonfilename);
 
-jsonstruct_material.include_current_collectors = true;
-
 %% Setup geometry input
 %
-% We use a simple 3d-geometry
+% We use a simple 3d-geometry (see image below) with only one layer
+%
+
 jsonfilename = fullfile('Examples'     , ...
                         'JsonDataFiles', ...
                         'geometry3d.json');
@@ -19,60 +31,125 @@ jsonstruct_geometry = parseBattmoJson(jsonfilename);
 
 %% Setup  Control input
 %
+% We use a single discharge scenario
 jsonfilename = fullfile('Examples', 'JsonDataFiles', 'cc_discharge_control.json');
 jsonstruct_control = parseBattmoJson(jsonfilename);
 
-%% Setup full input
+%% 
+% We merge the structures into a single input structure
 %
+
 jsonstruct = mergeJsonStructs({jsonstruct_geometry , ...
                                jsonstruct_material , ...
                                jsonstruct_control}, 'warn', false);
 
+%% Modify input structure
+% We modify the input parameters manually to obtain higher temperature increases so that the difference between the two
+% simulation approaches get enhanced.
 
-jsonstruct.Electrolyte.ionicConductivity.functionName = 'modified_electrolyte_conductivity';
-jsonstruct.Electrolyte.ionicConductivity.functionName = 'modified_electrolyte_conductivity';
+%%
+% We change the heat transfer coefficient to low values
+%
 
-%% Run non thermal simulation
+jsonstruct.ThermalModel.externalHeatTransferCoefficientTab = 1e-1;
+jsonstruct.ThermalModel.externalHeatTransferCoefficient    = 1e-1;
 
-jsonstruct_isothermal = setJsonStructField(jsonstruct, 'use_thermal', false, 'handleMisMatch', 'quiet');
-output_isothermal = runBatteryJson(jsonstruct_isothermal);
+%%
+% We reduce by a constant factor the specific heat capacities of all the materials
+%
 
-states = output_isothermal.states;
-times  = output_isothermal.time;
+coef = 5e-2;
+locations = {{'NegativeElectrode', 'CurrentCollector', 'specificHeatCapacity'}, ...
+             {'NegativeElectrode', 'Coating', 'ActiveMaterial', 'specificHeatCapacity'}, ...
+             {'PositiveElectrode', 'CurrentCollector', 'specificHeatCapacity'}, ...
+             {'PositiveElectrode', 'Coating', 'ActiveMaterial', 'specificHeatCapacity'}, ...
+             {'Electrolyte', 'specificHeatCapacity'}};
 
-%% Setup model with thermal support
-
-output_fullycoupled = runBatteryJson(jsonstruct);
-model = output_fullycoupled.model;
-
[…]
