+++ Examples/Advanced/utils/modified_electrolyte_diffusion.m
@@ -1,15 +1,20 @@
 function D = modified_electrolyte_diffusion(c, T)
 %% modified version of electrolyte conductivity, used to enhance thermal coupling in an example
-
+%% original version is computeDiffusionCoefficient_Ai2020.m
+    
     % Calculate diffusion coefficients constant for the diffusion coefficient calculation
     cnst = [ -4.43, -54; 
              -0.22, 0.0 ];
 
     Tgi = [ 229; 5.0 ];
     
+    modified_coef1 = 1; % original is 1
+
     % Diffusion coefficient, [m^2 s^-1]
-    D = 1e-2 .* 10 .^ ( ( cnst(1,1) + cnst(1,2) ./ ( T - Tgi(1) - Tgi(2) .* c .* 1e-3) + cnst(2,1) .* ...
+    D = 10 .^ ( ( cnst(1,1) + modified_coef1*cnst(1,2) ./ ( T - Tgi(1) - Tgi(2) .* c .* 1e-3) + cnst(2,1) .* ...
                           c .* 1e-3) );
+
+    % D = max(D, 1e-15);
     
 end
 
