### 2026-01-26 a519888
added thermal-only example in test case
