+++ Tests/TestExamples/TestRunExamples.m
@@ -29,6 +29,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runElectrolyser'                                 , ...
                     'runConductorBlockExample'                        , ...
                     'runRegionBruggemanExample'                       , ...
+                    'runOnlyThermal'                                  , ...
                     'magnesium_script'                                , ...
                     'exampleControl_doc'                              , ...
                     'plotControlExamples_doc'                         , ...
