### 2026-01-26 1675c47
some more change in runOnlyThermal
