### 2026-01-26 44a4f33
fixed some typos in notebook
