+++ Examples/Advanced/runOnlyThermal.m
@@ -1,62 +1,49 @@
 %% Comparison between coupled and uncoupled thermal simulation
-%
-% To compute the thermal response of a battery cell, one can either run a fully coupled simulation or run first an
-% isothermal simulation and then use the output to compute the thermal response in a thermal-only simulation.
-%
-% In the second case, we do not get any feedback from the thermal model to the electrochemical model.
-%
-% In this example, we compare the solutions obtained both for the voltage and for the temperature on a P4D model.
-%
-
-
-%% setup material property input
-%
-% We use a lithium-ion battery cell with NMC cathode and graphite anode
+% To compute the thermal response of a battery cell, one can either run a fully 
+% coupled simulation or run first an isothermal simulation and then use the output 
+% to compute the thermal response in a thermal-only simulation.
 % 
+% In the second case, we do not get any feedback from the thermal model to the 
+% electrochemical model.
+% 
+% In this example, we compare the solutions obtained both for the voltage and 
+% for the temperature on a P4D model.
+%% setup material property input
+% We use a lithium-ion battery cell with NMC cathode and graphite anode
+
 jsonfilename = fullfile('ParameterData'        , ...
                         'BatteryCellParameters', ...
                         'LithiumIonBatteryCell', ...
                         'lithium_ion_battery_nmc_graphite.json');
 jsonstruct_material = parseBattmoJson(jsonfilename);
-
 %% Setup geometry input
-%
 % We use a simple 3d-geometry (see image below) with only one layer
-%
 
 jsonfilename = fullfile('Examples'     , ...
                         'JsonDataFiles', ...
                         'geometry3d.json');
 jsonstruct_geometry = parseBattmoJson(jsonfilename);
-
-%% Setup  Control input
-%
+%% Setup Control input
 % We use a single discharge scenario
+
 jsonfilename = fullfile('Examples', 'JsonDataFiles', 'cc_discharge_control.json');
 jsonstruct_control = parseBattmoJson(jsonfilename);
-
 %% 
 % We merge the structures into a single input structure
-%
 
 jsonstruct = mergeJsonStructs({jsonstruct_geometry , ...
                                jsonstruct_material , ...
                                jsonstruct_control}, 'warn', false);
-
 %% Modify input structure
-% We modify the input parameters manually to obtain higher temperature increases so that the difference between the two
-% simulation approaches get enhanced.
-
-%%
+% We modify the input parameters manually to obtain higher temperature increases 
+% so that the difference between the two simulation approaches get enhanced.
+% 
 % We change the heat transfer coefficient to low values
-%
 
 jsonstruct.ThermalModel.externalHeatTransferCoefficientTab = 1e-1;
 jsonstruct.ThermalModel.externalHeatTransferCoefficient    = 1e-1;
-
-%%
+%% 
 % We reduce by a constant factor the specific heat capacities of all the materials
-%
 
 coef = 5e-2;
 locations = {{'NegativeElectrode', 'CurrentCollector', 'specificHeatCapacity'}, ...
@@ -70,86 +57,68 @@ for iloc = 1 : numel(locations)
     val = getJsonStructField(jsonstruct, loc);
     jsonstruct = setJsonStructField(jsonstruct, loc, coef*val, 'handleMisMatch', 'quiet');
 end
-
-%%
-% We chane the lower cutoff voltage
-%
+%% 
+% We change the lower cutoff voltage
 
 jsonstruct.Control.lowerCutoffVoltage = 3.6;
-
-
 %% Run thermal simulation
 
 output_fullycoupled = runBatteryJson(jsonstruct);
-
-%% Grid mesh
-%
-% Now the model is setup, we can plot the grid
-%
+%% Plot mesh
+% Now the model is setup, we can plot the grid mesh
 
[…]
