### 2026-02-11 b19820c
Start working on cccv-style time series
