+++ Control/TCtestControlModel.m
@@ -1,0 +1,470 @@
+classdef TCtestControlModel < ControlModel
+
+    properties
+
+        timecontrol % time series struct
+
+        computeInput % function called to give update
+
+        % for clipping
+        lowerCutoffVoltage
+        upperCutoffVoltage
+        dIdtLimit
+        dEdtLimit
+
+        % This values are initiated depending on C/D rate values and battery model
+        ImaxCharge
+        ImaxDischarge
+
+        %
+        switchTolerances
+    end
+
+
+    methods
+
+        function model = TCtestControlModel(inputparams)
+
+            model = model@ControlModel(inputparams);
+
+            fdnames = {'timecontrol', ...
+                       'lowerCutoffVoltage', ...
+                       'upperCutoffVoltage', ...
+                       'dEdtLimit'         , ...
+                       'dIdtLimit'         , ...
+                       'switchTolerances'};
+
+            model = dispatchParams(model, inputparams, fdnames);
+
+            model.computeInput = @(t) model.computeInputFromTable(t);
+
+        end
+
+        function model = registerVarAndPropfuncNames(model)
+
+            model = registerVarAndPropfuncNames@ControlModel(model);
+
+            varnames = {};
+            % Control type : string that can take following value
+            % - CC_discharge1
+            % - CC_discharge2
+            % - CC_charge1
+            % - CV_charge2
+            varnames{end + 1} = 'ctrlType';
+
+            % control value that can be either a voltage, current or power (not implemented)
+            varnames{end + 1} = 'ctrlVal';
+
+            % Estimate of the time derivative for voltage and current
+            varnames{end + 1} = 'dIdt';
+            varnames{end + 1} = 'dEdt';
+
+            model = model.registerVarNames(varnames);
+
+            % The following variables are not used in the residual assembly
+
+            varnames = {'dIdt'          , ...
+                        'dEdt'};
+
+            model = model.setAsStaticVarNames(varnames);
+            model = model.setAsExtraVarNames(varnames);
+
+            % Register the functions
+
+            fn = @TCtestControlModel.updateControlEquation;
+            model = model.registerPropFunction({'controlEquation', fn, {'ctrlType', 'ctrlVal', 'E', 'I'}});
+
+            fn = @TCtestControlModel.updateDerivatives;
+            fn = {fn, @(prop) PropFunction.accumFuncCallSetupFn(prop)};
+            inputvarnames = {'E', 'I'};
+            model = model.registerPropFunction({'dIdt', fn, {'E', 'I'}});
+            model = model.registerPropFunction({'dEdt', fn, {'E', 'I'}});
+
+        end
+
+
+        function [ctrlVal, ctrlType] = computeInputFromTable(model, t)
+
+            keyboard;
+
+        % We could be more efficient here and keep track of previous index to avoid full search (the time spent for
+        % that is probabely negligeable compared to the rest.)
+            ind = find(t >= model.times, 1, 'last');
+
+            if t > model.times(end)
+                error('outside of time table')
+            elseif t == model.times(end)
+                ind = numel(model.values);
+            end
+
+            ctrlVal  = model.values(ind);
+            ctrlType = model.controltypes(ind);
+
+            assert(ctrlType == 1, 'only constant current implemented');
+
+            switch ctrlType
+              case 1
+                %ctrlType = 'constantCurrent';
+                ctrlType = 'cc_discharge1';
+              case 2
+                ctrlType = 'constantVoltage';
+              case 3
+                ctrlType = 'constantPower';
+              otherwise
+                error('ctrlType %g not recognized. It should be equal to 1, 2 or 3', ctrlType);
[…]
