+++ Battery/GenericBattery.m
@@ -226,7 +226,7 @@ classdef GenericBattery < BaseModel
             switch model.(ctrl).controlPolicy
               case {'CCDischarge', 'CCCharge', 'CC', 'timeControl', 'TCtest'}
                 model = model.registerPropFunction({{ctrl, 'ctrlVal'}, fn, inputnames});
-              case {'CCCV', 'Impedance'}
+              case {'CCCV', 'Impedance', 'CCCV2'}
                 % do nothing
               otherwise
                 error('controlPolicy not recognized');
@@ -530,7 +530,7 @@ classdef GenericBattery < BaseModel
 
               case "TCtest"
 
-                control = TCtest(inputparams);
+                control = TCtestControlModel(inputparams);
 
               case "Impedance"
 
@@ -558,6 +558,14 @@ classdef GenericBattery < BaseModel
                 control.ImaxCharge    = (C/hour)*CRate;
                 control.ImaxDischarge = (C/hour)*DRate;
 
+              case "CCCV2"
+
+                control = CcCvControlModel2(inputparams);
+                CRate = control.CRate;
+                DRate = control.DRate;
+                control.ImaxCharge    = (C/hour)*CRate;
+                control.ImaxDischarge = (C/hour)*DRate;
+
               case "powerControl"
 
                 control = PowerControlModel(inputparams);
@@ -568,7 +576,7 @@ classdef GenericBattery < BaseModel
 
               otherwise
 
-                error('Error controlPolicy not recognized', inputparams.controlPolicy);
+                error('Error controlPolicy %s not recognized', inputparams.controlPolicy);
             end
 
         end
@@ -921,7 +929,7 @@ classdef GenericBattery < BaseModel
                 initstate.(ctrl).ctrlType = 'constantCurrent';
                 initstate.(ctrl).I = 0;
 
-              case 'CCCV'
+              case {'CCCV', 'CCCV2'}
 
                 initstate.(ctrl).numberOfCycles = 0;
 
@@ -964,7 +972,7 @@ classdef GenericBattery < BaseModel
                 end
 
               otherwise
-                error('control policy not recognized');
+                error('control policy %s not recognized', model.(ctrl).controlPolicy);
             end
 
             initstate.time = 0;
@@ -973,13 +981,13 @@ classdef GenericBattery < BaseModel
 
         function state = addVariables(model, state)
 
-        % Given a state where only the primary variables are defined, this
-        % functions add all the additional variables that are computed in the assembly process and have some physical
-        % interpretation.
-        %
-        % To do so, we use getEquations function and sends dummy variable for state0, dt and drivingForces
+            % Given a state where only the primary variables are defined, this
+            % functions add all the additional variables that are computed in the assembly process and have some physical
+            % interpretation.
+            %
+            % To do so, we use getEquations function and sends dummy variable for state0, dt and drivingForces
 
-        % Values that need to be set to get the function getEquations running
+            % Values that need to be set to get the function getEquations running
 
             dt = 1;
             state0 = state;
@@ -1166,11 +1174,11 @@ classdef GenericBattery < BaseModel
 
             switch model.(ctrl).controlPolicy
 
-              case {'CCCV', 'powerControl', 'TCtest'}
+              case {'CCCV', 'powerControl', 'CCCV2'}
 
                 % nothing to do here
 
-              case {'timeControl'}
+              case {'timeControl', 'TCtest'}
 
                 [ctrlVal, ctrlType] = model.(ctrl).computeInput(state.time);
 
[…]
