+++ Battery/BatteryInputParams.m
@@ -94,6 +94,8 @@ classdef BatteryInputParams < InputParams
                 inputparams.(ctrl) = PowerControlModelInputParams(pick(ctrl));
               case 'timeControl'
                 inputparams.(ctrl) = TimeControlModelInputParams(pick(ctrl));
+              case 'TCtest'
+                inputparams.(ctrl) = TCtestControlModelInputParams(pick(ctrl));
               otherwise
                 error('controlPolicy %s not recognized', jsonstruct.(ctrl).controlPolicy);
             end
