### 2026-02-12 e42f201
simplify cccv2 by hnil
