+++ Control/CcCvControlModel2.m
@@ -123,124 +123,150 @@ classdef CcCvControlModel2 < ControlModel
 
             ctrlVal = interp1(model.times, model.values, t);
 
-            fprintf('Input %g s: %g A\n', t, ctrlVal);
+            % fprintf('Input %g s: %g A\n', t, ctrlVal);
 
         end
 
 
-        function state = updateControlState(model, state, state0, dt, t)
+        function state = updateControlState(model, state, state0, dt)
 
             state = updateControlState@ControlModel(model, state, state0, dt);
 
             state = model.updateDerivatives(state, state0, dt);
 
-            fprintf('Received %g A (%g V)\n', state.I, state.E);
+            Emin = model.lowerCutoffVoltage;
+            Emax = model.upperCutoffVoltage;
 
-            if ~isfield(state, 'ctrlType')
-                return
+            %ctrlType = 'current';
+            ctrlType = state.ctrlType;
+
+            if state.E <= Emin && strcmp(state.ctrlType, 'current')
+                ctrlType = 'Emin';
+                % state.E = Emin;
+                %keyboard;
             end
 
-            rsw00     = model.setupRegionSwitchFlags(state0, state0.ctrlType);
-            ctrlType0 = state0.ctrlType;
-
-            if state.E < model.lowerCutoffVoltage
-                keyboard;
+            if state.E > Emax && strcmp(state.ctrlType, 'current')
+                ctrlType = 'Emax';
+                state.E = Emax;
             end
 
-            if rsw00.beforeSwitchRegion
+            if strcmp(state.ctrlType, 'Emin') && state.I > state.ctrlVal
+                ctrlType = 'current';
+                state.I = state.ctrlVal;
+            end
 
-                % We have not entered the switching region in the time
-                % step. We are not going to change control in this
-                % step.
-                ctrlType = ctrlType0;
-
-            else
-
-                % We entered the switch region in previous time step.
-
-                currentCtrlType = state.ctrlType; % current control in the the Newton iteration
-                nextCtrlType0   = model.getNextCtrlType(ctrlType0); % next control that can occur afte the previous time step control (if it changes)
-
-                rsw0 = model.setupRegionSwitchFlags(state, ctrlType0);
-
-                switch currentCtrlType
-
-                  case ctrlType0
-
-                    % The control has not changed from previous time
-                    % step and we want to determine if we should
-                    % change it.
-
-                    if rsw0.afterSwitchRegion
-
-                        % We switch to a new control because we are no
-                        % longer in the acceptable region for the
-                        % current control
-                        ctrlType = nextCtrlType0;
-
-                    else
-
-                        ctrlType = ctrlType0;
-
-                    end
-
-                  case nextCtrlType0
-
-                    % We do not switch back to avoid oscillation. We
-                    % are anyway within the given tolerance for the
-                    % control so that we keep the control as it is.
-
-                    ctrlType = nextCtrlType0;
-
-                  otherwise
-
-                    error('control type not recognized');
-
-                end
-
+            if strcmp(state.ctrlType, 'Emax') && state.I < state.ctrlVal
+                ctrlType = 'current';
+                state.I = state.ctrlVal;
             end
 
             state.ctrlType = ctrlType;
-            state = model.updateValueFromControl(state);
+            % state = model.updateValueFromControl(state);
+
+
+            % fprintf('Received %g A (%g V)\n', state.I, state.E);
+
+            % if ~isfield(state, 'ctrlType')
[…]
