### 2026-02-12 46aaba1
clean up by removing unused functions
