+++ Control/CcCvControlModel2.m
@@ -123,8 +123,6 @@ classdef CcCvControlModel2 < ControlModel
 
             ctrlVal = interp1(model.times, model.values, t);
 
-            % fprintf('Input %g s: %g A\n', t, ctrlVal);
-
         end
 
 
@@ -162,116 +160,12 @@ classdef CcCvControlModel2 < ControlModel
             end
 
             state.ctrlType = ctrlType;
-            % state = model.updateValueFromControl(state);
 
-
-            % fprintf('Received %g A (%g V)\n', state.I, state.E);
-
-            % if ~isfield(state, 'ctrlType')
-            %     return
-            % end
-
-            % rsw00     = model.setupRegionSwitchFlags(state0, state0.ctrlType);
-            % ctrlType0 = state0.ctrlType;
-
-            % if state.E < model.lowerCutoffVoltage
-            %     keyboard;
-            % end
-
-            % if rsw00.beforeSwitchRegion
-
-            %     % We have not entered the switching region in the time
-            %     % step. We are not going to change control in this
-            %     % step.
-            %     ctrlType = ctrlType0;
-
-            % else
-
-            %     % We entered the switch region in previous time step.
-
-            %     currentCtrlType = state.ctrlType; % current control in the the Newton iteration
-            %     nextCtrlType0   = model.getNextCtrlType(ctrlType0); % next control that can occur afte the previous time step control (if it changes)
-
-            %     rsw0 = model.setupRegionSwitchFlags(state, ctrlType0);
-
-            %     switch currentCtrlType
-
-            %       case ctrlType0
-
-            %         % The control has not changed from previous time
-            %         % step and we want to determine if we should
-            %         % change it.
-
-            %         if rsw0.afterSwitchRegion
-
-            %             % We switch to a new control because we are no
-            %             % longer in the acceptable region for the
-            %             % current control
-            %             ctrlType = nextCtrlType0;
-
-            %         else
-
-            %             ctrlType = ctrlType0;
-
-            %         end
-
-            %       case nextCtrlType0
-
-            %         % We do not switch back to avoid oscillation. We
-            %         % are anyway within the given tolerance for the
-            %         % control so that we keep the control as it is.
-
-            %         ctrlType = nextCtrlType0;
-
-            %       otherwise
-
-            %         error('control type not recognized');
-
-            %     end
-
-            % end
-
-            % state.ctrlType = ctrlType;
-            % state = model.updateValueFromControl(state);
 
         end
 
-        % function state = updateValueFromControl(model, state)
-
-        %     % ImaxD = model.ImaxDischarge;
-        %     % ImaxC = model.ImaxCharge;
-        %     Emin  = model.lowerCutoffVoltage;
-        %     Emax  = model.upperCutoffVoltage;
-
-        %     switch state.ctrlType
-
-        %       case 'current'
-
-        %         %state.I = ImaxD;
-        %         state.I = state.ctrlVal;
-
-        %       case 'Emin'
-
-        %         % state.I = 0; % ?
-        %         state.E = Emin;
-
-        %       case 'Emax'
-
-        %         % state.I = 0; % ?
-        %         state.E = Emax;
-
-        %       otherwise
-
-        %         error('ctrlType %s not recognized.', state.ctrlType)
-
-        %     end
-
-        % end
-
         function state = updateControlEquation(model, state)
 
-            ImaxD = model.ImaxDischarge;
-            ImaxC = model.ImaxCharge;
             Emin  = model.lowerCutoffVoltage;
             Emax  = model.upperCutoffVoltage;
 
@@ -281,11 +175,8 @@ classdef CcCvControlModel2 < ControlModel
 
[…]
