### 2026-02-12 3a0203c
removing more unused stuff
