### 2026-02-12 c0f6c53
add CCCV2 to BatteryInputParams
