### 2026-02-13 39e3a2d
Use first current value as initial cond
