### 2026-02-13 47dd583
assert valid inputs
