+++ Examples/Basic/swellingTestComposite.m
@@ -79,8 +79,6 @@ jsonstruct.(ne).(am).diffusionModelType = 'full';
 % various parts of the simulator during the simulation. This class is
 % instantiated using the jsonstruct we just created:
 
-jsonstruct.(ne).use_swelling_material = true;
-
 inputparams = CompositeBatteryInputParams(jsonstruct);
 inputparams = inputparams.validateInputParams();
 
