### 2026-02-16 3c013d7
added reference volume fraction for swelling
