### 2026-02-16 93d1abc
clean-up in GenericBattery for swelling
