+++ Examples/Basic/swellingTest.m
@@ -13,7 +13,7 @@ mrstModule add ad-core mrst-gui mpfa agmg linearsolvers
 clear all
 close all
 
-mrstDebug(20);
+% mrstDebug(20);
 
 %%% Specifying the physical model
 % In this tutorial we will simulate a lithium-ion battery consisting of a 
@@ -89,8 +89,8 @@ inputparams = BatteryInputParams(jsonstruct);
 % model can be found here:
 % :class:`FullSolidDiffusionModelInputParams <Electrochemistry.FullSolidDiffusionModelInputParams>`.
 
-inputparams.(ne).(co).(am).(sd).N = 2;
-inputparams.(pe).(co).(am).(sd).N = 2;
+inputparams.(ne).(co).(am).(sd).N = 5;
+inputparams.(pe).(co).(am).(sd).N = 5;
 
 %%% Setting up the geometry
 % Here, we setup the 1D computational mesh that will be used for the
@@ -99,9 +99,9 @@ inputparams.(pe).(co).(am).(sd).N = 2;
 % be found in the BattMo/Battery/BatteryGeometry folder.
 
 gen = BatteryGeneratorP2D();
-gen.sepnx = 2; % discretization number for negative current collector (default = 10)
-gen.nenx  = 2; % discretization number for negative active material (default = 10)
-gen.penx  = 2; % discretization number for separator (default = 10)
+gen.sepnx = 10; % discretization number for negative current collector (default = 10)
+gen.nenx  = 10; % discretization number for negative active material (default = 10)
+gen.penx  = 10; % discretization number for separator (default = 10)
 
 %%%
 % Now, we update the inputparams with the properties of the mesh. This function
@@ -166,7 +166,7 @@ cgit.printRootVariables
 % We set the total time scaled by the CRate in the model.
 % The CRate has been set by the json file. We can access it here:
 
-timestep.timeStepDuration = 100;
+timestep.numberOfTimeSteps = 100;
 
 step    = model.Control.setupScheduleStep(timestep);
 control = model.Control.setupScheduleControl();
@@ -233,20 +233,7 @@ initstate = model.setupInitialState(jsonstruct);
 nls = NonLinearSolver;
 nls.errorOnFailure = false;
 
-[wellSols, states, report] = simulateScheduleAD(initstate, model, schedule, 'OutputMinisteps', true, 'NonLinearSolver', nls); 
-
-return
-
-
-%%%
-% The outputs from the simulation are:
-% - wellSols: which provides the current and voltage of the battery at each 
-% timestep. (This naming convention is a hangover from MRST where we model
-% reservoir injection via injection wells).
-% - states: which contains the values of the primary variables in the model
-% at each timestep.
-% - reports: which contains technical information about the steps used in
-% the numerical solvers.
+[~, states, report] = simulateScheduleAD(initstate, model, schedule, 'OutputMinisteps', true, 'NonLinearSolver', nls); 
 
 
 %% Plotting the results
@@ -277,6 +264,7 @@ plot(T/hour, I)
 xlabel('time [hours]')
 ylabel('Cell Current [A]')
 
+
 %% Plot the overpotential of the negative electrode as a function of time
 
 for istate = 1 : numel(states)
@@ -285,330 +273,70 @@ end
 
[…]
