### 2026-02-17 6f3b9d8
added inputparams class
