+++ Electrochemistry/FullSolidDiffusionSwellingModel.m
@@ -1,9 +1,34 @@
 classdef FullSolidDiffusionSwellingModel < FullSolidDiffusionModel
+% @article{Chandrasekaran_2010,
+%   title =        {Analysis of Lithium Insertion/Deinsertion in a Silicon Electrode Particle at Room Temperature},
+%   volume =       157,
+%   ISSN =         {0013-4651},
+%   url =          {http://dx.doi.org/10.1149/1.3474225},
+%   DOI =          {10.1149/1.3474225},
+%   number =       10,
+%   journal =      {Journal of The Electrochemical Society},
+%   publisher =    {The Electrochemical Society},
+%   author =       {Chandrasekaran, Rajeswari and Magasinski, Alexandre and Yushin, Gleb and Fuller, Thomas F.},
+%   year =         2010,
+%   pages =        {A1139}
+% }
 
     properties
 
         molarVolumeSi = 1.2e-05;
-        molarVolumeLi = 8.8e-06;
+        % molarVolumeSi = 5e-05;
+        molarVolumeLi = 9e-6
+
+        referenceFillInLevel % Fill-in value which corresponds to the given radius
+
+        %% Computed at initialization
+
+        zeroFillInParticleRadius
+        zeroFillInVolumetricSurfaceArea
+        
+        %% Advanced parameters
+        % typically setup by swelling coating model
+        zeroFillInVolumeFraction
         
     end
 
@@ -13,17 +38,43 @@ classdef FullSolidDiffusionSwellingModel < FullSolidDiffusionModel
 
             model = model@FullSolidDiffusionModel(inputparams);
             
+            fdnames = {'referenceFillInLevel', ...
+                       'zeroFillInVolumeFraction'};
+
+            model = dispatchParams(model, inputparams, fdnames);
+
+            %% compute zero fill in radius from reference fill-in and given particule radius
+
+            compmodel = model;
+            compmodel = compmodel.registerVarAndPropfuncNames();
+            compmodel = compmodel.removePropFunction('x');
+            compmodel = compmodel.setupComputationalGraph();
+
+            clear state
+            state.x = compmodel.referenceFillInLevel;
+            state = compmodel.evalVarName(state, 'radiusElongation');
+
+            re = state.radiusElongation;
+
+            model.zeroFillInParticleRadius        = model.particleRadius/re;
+            model.zeroFillInVolumetricSurfaceArea = model.volumetricSurfaceArea*re;
+            
         end
 
         function model = registerVarAndPropfuncNames(model)
-
-            sd = 'SolidDiffusion';
 
             model = registerVarAndPropfuncNames@FullSolidDiffusionModel(model);
 
             varnames = {'radiusElongation', ...
                         'radius'          , ...
                         'intercalationFlux'};
+
+            % Particle fill-in level (between 0 and 1). Depends only on stoichiometry
+            % By definition, we have : x = (volumeFraction*concentration)/(volumeFraction_max*concentration_max)
+            % Due to the special expression for the volume fraction, it reduces to a function of the stoichiometry only,
+            % See paper in reference
+
+            varnames{end + 1} = 'x';
 
             model = model.registerVarNames(varnames);
 
@@ -37,8 +88,11 @@ classdef FullSolidDiffusionSwellingModel < FullSolidDiffusionModel
             fn = @FullSolidDiffusionSwellingModel.updateRadius;
             model = model.registerPropFunction({'radius', fn, {'radiusElongation'}});
 
+            fn = @FullSolidDiffusionSwellingModel.updateFillInLevel;
+            model = model.registerPropFunction({'x', fn, {'cAverage'}});
+            
             fn = @FullSolidDiffusionSwellingModel.updateRadiusElongation;
[…]
