### 2026-02-17 9bd565c
update in initialization in case of swelling material
