### 2026-02-17 ae79d1c
improved plots in swellingTest
