### 2026-02-17 7cc1a4c
changed parameter name to maximumTotalConcentration (swelling)
