### 2026-02-17 5d328d0
added check for no binder and additives for swelling
