### 2026-02-17 139ef68
some clean-up in FullSolidDiffusionSwellingModel
