### 2026-02-17 b84a74f
added support for both electrodes in swelling
