### 2026-02-17 30a802f
some fix in computation of molar of lithiated volume
