### 2026-02-17 96506cb
reviving stress support for swelling
