### 2026-02-18 ce30e4c
improved behaviour of OpenPropFunction in computation graph
