### 2026-02-18 fef4738
fix for hydraulic stress support
