+++ Electrochemistry/SwellingCoating.m
@@ -16,8 +16,14 @@ classdef SwellingCoating < Coating
         molarMass
         referenceFillInLevel % fill-in value which corresponds to the given volumeFraction of the coating
 
-        includeHydrostaticStress = false;
+        include_hydrostatic_stress
 
+        %% If include hydrostatic stress
+
+        youngModulus             % Young's modulus
+        poissonRatio             % Poisson's ratio
+        solutePartialMolarVolume % partial molar volume of the solute
+        
         %% Computed at initialization
         
         zeroFillInVolumeFraction
@@ -31,8 +37,12 @@ classdef SwellingCoating < Coating
         function model = SwellingCoating(inputparams)
 
             model = model@Coating(inputparams)
-            fdnames = {'molarMass', ...
-                       'referenceFillInLevel'};
+            fdnames = {'molarMass'                 , ...
+                       'referenceFillInLevel'      , ...
+                       'include_hydrostatic_stress', ...
+                       'youngModulus'              , ...
+                       'poissonRatio'              , ...
+                       'solutePartialMolarVolume'};
             
             model = dispatchParams(model, inputparams, fdnames);
 
@@ -92,8 +102,8 @@ classdef SwellingCoating < Coating
 
             model = model.registerVarNames(varnames);
 
-            if model.includeHydrostaticStress
-                model = model.registerVarName('hydrostaticStress');
+            if model.include_hydrostatic_stress
+                model = model.registerVarName({am, 'hydrostaticStress'});
             end
             
             fn = @SwellingCoating.updateVolumetricSurfaceArea;
@@ -105,14 +115,14 @@ classdef SwellingCoating < Coating
             fn =  @SwellingCoating.updateSolidIntercalationFlux;
             model = model.registerPropFunction({{am, sd, 'intercalationFlux'}, fn, {{am, itf, 'intercalationFlux'}}});
 
-            if model.includeHydrostaticStress
+            if model.include_hydrostatic_stress
                 
                 fn = @SwellingCoating.updateHydrostaticStress;
-                model = model.registerPropFunction({'hydrostaticStress', fn, {{am, sd, 'cAverage'}, {am, itf, 'cElectrodeSurface'}}});
+                model = model.registerPropFunction({{am, 'hydrostaticStress'}, fn, {{am, sd, 'cAverage'}, {am, itf, 'cElectrodeSurface'}}});
 
-                fn  = @SwellingCoating.updateReactionRate;
-                inputnames = {{am, itf, 'T'}, {am, itf, 'j0'}, {am, itf, 'eta'}, 'hydrostaticStress'};
-                model = model.registerPropFunction({{am, itf, 'intercalationFlux'}, fn, inputnames});
+                fn  = @SwellingCoating.updateEta;
+                inputnames = {{am, itf, 'phiElectrode'}, {am, itf, 'phiElectrolyte'}, {am, itf, 'OCP'}, {am, 'hydrostaticStress'}};
+                model = model.registerPropFunction({{am, itf, 'eta'}, fn, inputnames});
                 
             end
             
@@ -181,26 +191,20 @@ classdef SwellingCoating < Coating
 
         end
 
-        function state = updateReactionRate(model, state)
+        function state = updateEta(model, state)
         % Same as in the interface class but uses the Butler Volmer
         % equation including stress
 
             am  = 'ActiveMaterial';
-            sd  = 'SolidDiffusion';
             itf = 'Interface';
-            
-            n     = model.(am).(itf).numberOfElectronsTransferred;
-            F     = model.(am).(itf).constants.F;
[…]
