### 2026-02-18 d87ca13
cleaning-up example
