+++ Examples/Basic/swellingTest.m
@@ -1,51 +1,31 @@
-% BattMo Tutorial
-% This tutorial explains how to setup and run a simulation in BattMo
+%% Silicon Swelling Example 
+%
 
-%%% Setting up the environment
-% BattMo uses functionality from :mod:`MRST <MRSTBattMo>`. This functionality 
-% is collected into modules where each module contains code for doing 
-% specific things. To use this functionality we must add these modules to 
-% the matlab path by running:
+%% Material properties
+% We load the json structure for the material properties
+jsonfilename = fullfile('ParameterData'        , ...
+                        'BatteryCellParameters', ...
+                        'LithiumIonBatteryCell', ...
+                        'lithium_ion_battery_nmc_silicon.json');
+jsonstruct_material = parseBattmoJson(jsonfilename);
 
+%% Geometry
+% We load the json structure for the geometrical properties
 
-mrstModule add ad-core mrst-gui mpfa agmg linearsolvers
+jsonfilename = fullfile('Examples', 'JsonDataFiles', 'geometry1d.json');
+jsonstruct_geometry = parseBattmoJson(jsonfilename);
 
-clear all
-close all
+%% Control
+% We load the json structure for the geometrical properties
 
-% mrstDebug(20);
+jsonfilename = fullfile('Examples', 'JsonDataFiles', 'cc_discharge_control.json');
+jsonstruct_control = parseBattmoJson(jsonfilename);
 
-%%% Specifying the physical model
-% In this tutorial we will simulate a lithium-ion battery consisting of a 
-% negative electrode, a positive electrode and an electrolyte. *BattMo* 
-% comes with some pre-defined models which can be loaded from JSON files.
-% Here we will load the basic lithium-ion model JSON file which comes with
-% Battmo.
+jsonstruct = mergeJsonStructs({jsonstruct_material, ...
+                               jsonstruct_geometry, ...
+                               jsonstruct_control});
 
-fname = fullfile('ParameterData'        , ...
-                 'BatteryCellParameters', ...
-                 'LithiumIonBatteryCell', ...
-                 'lithium_ion_battery_nmc_silicon.json');
-
-jsonstruct = parseBattmoJson(fname);
-
-%%%
-% The parseBattmoJson function parses the JSON input and creates a matlab
-% structure containing the same fields as the JSON input. This structure 
-% can be changed to setup the model in the way that we want. 
-
-%%%
-% In this instance we will exclude temperature effects by setting
-% use_thermal to false.
-
-jsonstruct.use_thermal = false;
-
-%%%
-% We will also not use current collectors in this example:
-
-jsonstruct.include_current_collectors = false;
-
-%%%
+%%
 % The structure created in the jsonstruct follows the same hierarchy as the
 % fields in the JSON input file. These can be referenced by name in the
 % jsonstruct. To make life easier for ourselves we define some shorthand
@@ -62,181 +42,10 @@ sd      = 'SolidDiffusion';
 ctrl    = 'Control';
 cc      = 'CurrentCollector';
 
-%%%
-% Now we can set the diffusion model type for the active material (am) in the
-% positive (pe) and negative (ne) electrodes to 'full'.
+output = runBatteryJson(jsonstruct, 'runSimulation', true);
 
-jsonstruct.(ne).(co).(am).diffusionModelType = 'swelling';
-jsonstruct.(pe).(co).(am).diffusionModelType = 'full';
-
-%%%
-% To see which other types of diffusion model are available one can view 
-% :class:`ActiveMaterialInputParams <Electrochemistry.ActiveMaterialInputParams>.
-
-%%%
-% When running a simulation, *BattMo* requires that all model parameters
-% are stored in an instance of :class:`BatteryInputParams <Battery.BatteryInputParams>`. 
[…]
