+++ Examples/Basic/swellingTestComposite.m
@@ -1,686 +1,0 @@
-% BattMo Tutorial
-% This tutorial explains how to setup and run a simulation in BattMo
-
-%%% Setting up the environment
-% BattMo uses functionality from :mod:`MRST <MRSTBattMo>`. This functionality 
-% is collected into modules where each module contains code for doing 
-% specific things. To use this functionality we must add these modules to 
-% the matlab path by running:
-
-mrstModule add ad-core mrst-gui mpfa agmg linearsolvers
-
-clear all
-close all
-
-%%% Specifying the physical model
-% In this tutorial we will simulate a lithium-ion battery consisting of a 
-% negative electrode, a positive electrode and an electrolyte. *BattMo* 
-% comes with some pre-defined models which can be loaded from JSON files.
-% Here we will load the basic lithium-ion model JSON file which comes with
-% Battmo.
-
-fname = fullfile('ParameterData','BatteryCellParameters',...
-    'LithiumIonBatteryCell','lithium_ion_battery_nmc_silicon_graphite_Swelling.json');
-jsonstruct = parseBattmoJson(fname);
-
-%%%
-% The parseBattmoJson function parses the JSON input and creates a matlab
-% structure containing the same fields as the JSON input. This structure 
-% can be changed to setup the model in the way that we want. 
-
-%%%
-% In this instance we will exclude temperature effects by setting
-% use_thermal to false.
-
-jsonstruct.use_thermal = false;
-
-%%%
-% We will also not use current collectors in this example:
-
-jsonstruct.include_current_collectors = false;
-
-%%%
-% Our model will simulate diffusion so we set use_particle_diffusion to
-% true:
-
-jsonstruct.use_particle_diffusion = true;
-
-%%%
-% The structure created in the jsonstruct follows the same hierarchy as the
-% fields in the JSON input file. These can be referenced by name in the
-% jsonstruct. To make life easier for ourselves we define some shorthand
-% names for various parts of the structure.
-
-ne      = 'NegativeElectrode';
-pe      = 'PositiveElectrode';
-elyte   = 'Electrolyte';
-thermal = 'ThermalModel';
-am      = 'ActiveMaterial';
-itf     = 'Interface';
-sd      = 'SolidDiffusion';
-ctrl    = 'Control';
-cc      = 'CurrentCollector';
-
-%%%
-% Now we can set the diffusion model type for the active material (am) in the
-% positive (pe) and negative (ne) electrodes to 'full'.
-
-jsonstruct.(pe).(am).diffusionModelType = 'full';
-jsonstruct.(ne).(am).diffusionModelType = 'full';
-
-%%%
-% To see which other types of diffusion model are available one can view 
-% :class:`ActiveMaterialInputParams <Electrochemistry.ActiveMaterialInputParams>.
-
-%%%
-% When running a simulation, *BattMo* requires that all model parameters
-% are stored in an instance of :class:`BatteryInputParams <Battery.BatteryInputParams>`. 
-% This class is used to initialize the simulation and is accessed by
-% various parts of the simulator during the simulation. This class is
-% instantiated using the jsonstruct we just created:
-
-inputparams = CompositeBatteryInputParams(jsonstruct);
-inputparams = inputparams.validateInputParams();
-
-inputparams.scenario = 'discharge';
-
-%%%
-% It is also possible to update the properties of this inputparams in a
[…]
