### 2026-02-18 f073da5
added swelling example in test suite
