+++ Tests/TestExamples/TestRunExamples.m
@@ -34,6 +34,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'plotControlExamples_doc'                         , ...
                     'plotGeometries_doc'                              , ...
                     'runExample3D_doc'                                , ...
+                    'ruSwellingExample'                               , ...
                     'runSafariSEI'                                    , ...
                     'runBolaySEI'                                     , ...
                     'runBolaySEI0'                                    , ...
