+++ Battery/BatteryGeometry/BatteryGenerator1DSwelling.m
@@ -1,247 +1,0 @@
-classdef BatteryGenerator1DSwelling < BatteryGenerator
-% setup 1D grid 
-    properties
-
-        %
-        % vector of component lengths
-        %
-        % - x(1) : length of negative current collector (default = 25 micro meter)
-        % - x(2) : length of negative active material (default = 64 micro meter)
-        % - x(3) : length of separator (default = 15 micro meter)
-        % - x(4) : length of positive active material (default = 57 micro meter)
-        % - x(5) : length of positive current collector (default = 15 micro meter)
-        % 
-        xlength = 1e-6*[25; 25; 25; 71; 15];
-
-        sepnx  = 10; % discretization number for negative current collector (default = 10)
-        nenx   = 10; % discretization number for negative active material (default = 10)
-        penx   = 10; % discretization number for separator (default = 10)
-        nenr   = 10; % discretization for solid diffusion (default = 10)
-        penr   = 10; % discretization for solid diffusion (default = 10)
-        ccnenx = 10; % discretization number for positive current collector (default = 10)
-        ccpenx = 10; % discretization number for positive active material (default = 10)
-
-        %
-        % refinement factor (can be used to easily increase discretization refinement)
-        % see applyResolutionFactors method
-        %
-        fac = 1;
-
-        % boolean : true if grid for current collectors should be included
-        include_current_collectors
-        % boolean : true if grid for thermal model should be included
-        use_thermal
-
-        % Face area in the transversal direction (default = 1)
-        faceArea = 1; 
-        
-    end
-    
-    methods
-        
-        function gen = BatteryGenerator1DSwelling()
-            gen = gen@BatteryGenerator();  
-        end
-            
-        function [inputparams, gen] = updateBatteryInputParams(gen, inputparams)
-
-            gen.include_current_collectors = inputparams.include_current_collectors;
-            gen.use_thermal = inputparams.use_thermal;
-            inputparams = gen.setupBatteryInputParams(inputparams, []);
-
-            % We define some shorthand names for simplicity.
-            ne      = 'NegativeElectrode';
-            pe      = 'PositiveElectrode';
-            elyte   = 'Electrolyte';
-            sep     = 'Separator';
-            thermal = 'ThermalModel';
-            am      = 'ActiveMaterial';
-            cc      = 'CurrentCollector';
-
-            % we update all the grids to adjust grid to faceArea
-            inputparams.G               = gen.adjustGridToFaceArea(inputparams.G);
-            inputparams.(elyte).G       = gen.adjustGridToFaceArea(inputparams.(elyte).G);
-            inputparams.(elyte).(sep).G = gen.adjustGridToFaceArea(inputparams.(elyte).(sep).G);
-            
-            eldes = {ne, pe};
-            for ielde = 1 : numel(eldes)
-                elde = eldes{ielde};
-                inputparams.(elde).(am).G = gen.adjustGridToFaceArea(inputparams.(elde).(am).G);
-                if gen.include_current_collectors
-                    inputparams.(elde).(cc).G = gen.adjustGridToFaceArea(inputparams.(elde).(cc).G);
-                end
-            end
-            
-            if gen.use_thermal
-                inputparams.(thermal).G = gen.adjustGridToFaceArea(inputparams.(thermal).G);                
-            end
-            
-        end
-
-        function [inputparams, gen] = setupGrid(gen, inputparams, ~)
-            
-            sepnx  = gen.sepnx;
-            nenx   = gen.nenx;
-            penx   = gen.penx;
-            ccnenx = gen.ccnenx;
-            ccpenx = gen.ccpenx;
-
[…]
