+++ Tests/TestExamples/TestRunExamples.m
@@ -35,7 +35,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'plotControlExamples_doc'                         , ...
                     'plotGeometries_doc'                              , ...
                     'runExample3D_doc'                                , ...
-                    'ruSwellingExample'                               , ...
+                    'runSwellingExample'                              , ...
                     'runSafariSEI'                                    , ...
                     'runBolaySEI'                                     , ...
                     'runBolaySEI0'                                    , ...
