### 2026-02-19 7b88b06
more runSwellingExample
