+++ Examples/Advanced/Swelling/runSwellingExample.m
@@ -1,5 +1,25 @@
 %% Silicon Swelling Example 
 %
+% In this example, we illustrate the implementation of Silicon swelling in BattMo
+%
+% Silicon swelling model in BattMo includes
+%
+% * Modified solid diffusion equation due to changing particle radius
+% * Volume fraction change and impact of the effective values of transport parameters such as diffusion and conductivity in electrolyte and electrode
+% * Effective convective flux
+% * Modified Butler-Volmer equation due to hydrostatic stress
+%
+% We do not include the propagation of mechanical stress.
+%
+% The modeling equations are taken from the following references
+%
+% * Shweta Dhillon et al. “Modelling capacity fade in silicon-graphite composite electrodes for lithium-ion batteries”. In: Electrochimica Acta 377 (May 2021).
+% * Rajeswari Chandrasekaran and Thomas Fuller. “Analysis of the Lithium-Ion Insertion Silicon Composite Electrode/Separator/Lithium Foil Cell”. In: ECS Meeting Abstracts MA2011-02.7 (Aug. 2011).
+% * Rajeswari Chandrasekaran et al. “Analysis of Lithium Insertion/Deinsertion in a Silicon Electrode Particle at Room Temperature”. In: Journal of The Electrochemical Society 157.10 (2010).
+% * Yang-Tse Cheng and Mark W Verbrugge. “Evolution of stress within a spherical insertion electrode particle under potentiostatic and galvanostatic operation”. In: Journal of Power Sources 190.2 (2009).
+% * Haoliang Li et al. “Effects of stress dependent electrochemical reaction on voltage hysteresis of lithium ion batteries”. In: Applied Mathematics and Mechanics 39.10 (2018).
+%
+% 
 
 %% Material properties
 % We load the json structure for the material properties
@@ -10,13 +30,15 @@ jsonfilename = fullfile('ParameterData'        , ...
 jsonstruct_material = parseBattmoJson(jsonfilename);
 
 %% Geometry
-% We load the json structure for the geometrical properties
+% We load the json structure for the geometrical properties. We consider a P2D model.
+% 
 
 jsonfilename = fullfile('Examples', 'JsonDataFiles', 'geometry1d.json');
 jsonstruct_geometry = parseBattmoJson(jsonfilename);
 
 %% Control
 % We load the json structure for the geometrical properties
+%
 
 jsonfilename = fullfile('Examples', 'JsonDataFiles', 'cc_discharge_control.json');
 jsonstruct_control = parseBattmoJson(jsonfilename);
@@ -25,22 +47,6 @@ jsonstruct = mergeJsonStructs({jsonstruct_material, ...
                                jsonstruct_geometry, ...
                                jsonstruct_control});
 
-%%
-% The structure created in the jsonstruct follows the same hierarchy as the
-% fields in the JSON input file. These can be referenced by name in the
-% jsonstruct. To make life easier for ourselves we define some shorthand
-% names for various parts of the structure.
-
-ne      = 'NegativeElectrode';
-pe      = 'PositiveElectrode';
-co      = 'Coating';
-elyte   = 'Electrolyte';
-thermal = 'ThermalModel';
-am      = 'ActiveMaterial';
-itf     = 'Interface';
-sd      = 'SolidDiffusion';
-ctrl    = 'Control';
-cc      = 'CurrentCollector';
 
 output = runBatteryJson(jsonstruct, 'runSimulation', true);
 
[…]
