### 2026-02-19 8149e75
added notebooks for runSwellingExample
