### 2026-02-19 eb993dd
removed unused function
