### 2026-02-19 40aea95
adding half cell support
