+++ Battery/HalfCell.m
@@ -1,0 +1,612 @@
+classdef HalfCell < BaseModel
+% 
+% Used when at least one of the two electrodes is a swelling Material.
+% It functions as the Battery class but implements a new equation necessary because of the particle swelling :
+% the volume Conservation equation 
+%
+    properties
+
+        Electrolyte
+        Separator
+        ActiveMaterial
+
+        Control
+        SOC
+        use_thermal
+
+        externalCoupling
+        
+        initT
+        
+        primaryVarNames
+        funcCallList
+        
+    end
+    
+    methods
+        
+        function model = HalfCell(inputparams)
+            
+            model = model@BaseModel()
+
+            model.Electrolyte    = Electrolyte(inputparams.Electrolyte);
+            model.ActiveMaterial = ActiveMaterial(inputparams.ActiveMaterial);
+            
+            model.Control        = IEswitchControlModel(inputparams.Control);
+
+            model.initT = 298;
+            
+        end
+
+        function state = addVariables(model, state)
+            
+        % Given a state where only the primary variables are defined, this
+        % functions add all the additional variables that are computed in the assembly process and have some physical
+        % interpretation.
+        %
+        % To do so, we use getEquations function and sends dummy variable for state0, dt and drivingForces 
+            
+        % Values that need to be set to get the function getEquations running
+            
+            dt = 1;
+            state0 = state;
+            model.Control = ControlModel([]);
+            model.Control.controlPolicy = 'None';
+            drivingForces = model.getValidDrivingForces();
+
+            % We call getEquations to update state
+            
+            [~, state] = getEquations(model, state0, state, dt, drivingForces, 'ResOnly', true);
+
+            % we set to empty the fields we know that are not meaningfull
+
+            ne      = 'NegativeElectrode';
+            pe      = 'PositiveElectrode';
+            am      = 'ActiveMaterial';
+            cc      = 'CurrentCollector';
+            elyte   = 'Electrolyte';
+            am      = 'ActiveMaterial';
+            itf     = 'Interface';
+            sd      = "SolidDiffusion";
+            thermal = 'ThermalModel';
+            ctrl    = 'Control';
+
+
+            switch model.(am).diffusionModelType
+              case 'full'
+                state.(am).(sd).massAccum = [];
+                state.(am).(sd).massCons = [];
+              case {'simple', 'interParticleOnly'}
+                state.(am).massAccum = [];
+                state.(am).massCons = [];                    
+              otherwise
+                error('diffusion model type not recognized');
+            end
+           
+           
+        
+            switch model.(am).diffusionModelType
+              case 'full'
+                state.(elde).(sd) = model.(am).(sd).updateAverageConcentration(state.(elde).(am).(sd));
+                state.(elde).(am) = model.(am).updateSOC(state.(elde).(am));
+                state.(elde).(am) = model.(am).updateAverageConcentration(state.(elde).(am));
+              case {'simple', 'interParticleOnly'}
+                % do nothing
+              otherwise
+                error('diffusion model type not recognized');
+            end
+            
+
+            if model.use_thermal
+                state = model.updateThermalIrreversibleReactionSourceTerms(state);
+                state = model.updateThermalReversibleReactionSourceTerms(state);
+            end
+            
+        end
+
+        function forces = getValidDrivingForces(model)
+            
+            forces = getValidDrivingForces@PhysicalModel(model);
+            
+            ctrl = 'Control';
+            switch model.(ctrl).controlPolicy
+              case 'CCCV'
[…]
