### 2026-02-19 60ece6e
update in notebook and in doc
