### 2026-02-26 18830d1
Fix habanero version for compatability with urllib3
