+++ Documentation/requirements.txt
@@ -16,7 +16,7 @@ docutils==0.21.2
 fastjsonschema==2.21.1
 fonttools==4.60.2
 h11==0.16.0
-habanero==2.0.0
+habanero>=2.0.0
 httpcore==1.0.9
 httpx==0.28.1
 idna==3.10
