### 2026-03-02 1fe6b90
updated parameter names for Lithium plating
