+++ Electrochemistry/LithiumPlatingLatz.m
@@ -1,14 +1,14 @@
 classdef LithiumPlatingLatz < BaseModel
 % Implementation of the model from Hein et al
 % @article{hein2020electrochemical,
-%   title={An electrochemical model of lithium plating and stripping in lithium ion batteries},
-%   author={Hein, Simon and Danner, Timo and Latz, Arnulf},
-%   journal={ACS Applied Energy Materials},
-%   volume={3},
-%   number={9},
-%   pages={8519--8531},
-%   year={2020},
-%   publisher={ACS Publications}
+%   title     = {An electrochemical model of lithium plating and stripping in lithium ion batteries},
+%   author    = {Hein, Simon and Danner, Timo and Latz, Arnulf},
+%   journal   = {ACS Applied Energy Materials},
+%   volume    = {3},
+%   number    = {9},
+%   pages     = {8519--8531},
+%   year      = {2020},
+%   publisher = {ACS Publications}
 % }
     
     properties
@@ -16,31 +16,31 @@ classdef LithiumPlatingLatz < BaseModel
         F = PhysicalConstants.F
         R = PhysicalConstants.R
 
-        alphaPl                                 % Symmetry factor for lithium plating/stripping reaction (typically ≈ 0.3)
-        alphaStr                                % Symmetry factor for stripping (may be 1-alphaPl)
-        alphaChInt                              % Symmetry factor for charge-neutral chemical intercalation of plated lithium
+        symmetryFactorPlating               % Symmetry factor for lithium plating/stripping reaction (typically ≈ 0.3)
+        symmetryFactorStripping             % Symmetry factor for stripping (may be 1-symmetryFactorPlating)
+        symmetryFactorChemicalIntercalation % Symmetry factor for charge-neutral chemical intercalation of plated lithium
 
-        kPl                                     % Reaction rate constant for lithium plating (N⁰₀ Plating)
-        kChInt                                  % Reaction rate constant for chemical intercalation of plated lithium
-        kInter                                  % Reaction rate constant for direct lithium-ion intercalation into graphite
+        reactionRatePlating                 % Reaction rate constant for lithium plating (N⁰₀ Plating)
+        reactionRateChemicalIntercalation   % Reaction rate constant for chemical intercalation of plated lithium
+        reactionRateDirectIntercalation     % Reaction rate constant for direct lithium-ion intercalation into graphite
 
-        nPl0                                    % Phenomenological parameter: minimum lithium amount needed to activate metal activity (see eqn (8))
-        nPlLimit                                % Limit amount of plated lithium corresponding to one monolayer on graphite surface (see eqn (26))
-        platedConcentrationRef                  % Reference concentration for plated lithium [mol/m^3]
-        c
-        volumetricSurfaceArea                   % Interfacial surface area between graphite and electrolyte per unit volume [m²/m³]
-        particleRadius                          % Radius of graphite particles, used in solid-state diffusion modeling
-        volumeFraction                          % Porosity
+        thresholdParameter                  % Phenomenological parameter: minimum lithium amount needed to activate metal activity (see eqn (8))
+        limitAmount                         % Limit amount of plated lithium corresponding to one monolayer on graphite surface (see eqn (26))
+        platedReferenceConcentration        % Reference concentration for plated lithium [mol/m^3]
+
+        volumetricSurfaceArea               % Interfacial surface area between graphite and electrolyte per unit volume [m²/m³]
+        particleRadius                      % Radius of graphite particles, used in solid-state diffusion modeling
+        volumeFraction                      % Porosity
         
-        platedLiMonolayerThickness              % Thickness of one monolayer of plated lithium around a particle [m]. See equation (S-3)
[…]
