+++ Examples/Advanced/Plating/runPlating2.m
@@ -1,161 +1,0 @@
-%% Multi-reactionRatePlating and reactionRateChemicalIntercalation simulation with lithium plating
-
-clear
-close all
-
-%% Define reactionRatePlating and reactionRateChemicalIntercalation values to test
-reactionRatePlating_values = logspace(-11, -9, 10);
-reactionRateChemicalIntercalation_values = logspace(-15, -9, 40);
-n_reactionRatePlating = length(reactionRatePlating_values);
-n_reactionRateChemicalIntercalation = length(reactionRateChemicalIntercalation_values);
-
-results = cell(n_reactionRatePlating, n_reactionRateChemicalIntercalation);
-line_styles = {'-', '--', ':'};
-
-for i = 1:n_reactionRatePlating
-    for j = 1:n_reactionRateChemicalIntercalation
-        reactionRatePlating = reactionRatePlating_values(i);
-        reactionRateChemicalIntercalation = reactionRateChemicalIntercalation_values(j);
-        fprintf('\nRunning simulation (%d/%d, %d/%d): reactionRatePlating = %.1e, reactionRateChemicalIntercalation = %.1e\n', ...
-            i, n_reactionRatePlating, j, n_reactionRateChemicalIntercalation, reactionRatePlating, reactionRateChemicalIntercalation);
-
-        jsonstruct = parseBattmoJson(fullfile('ParameterData','BatteryCellParameters','LithiumIonBatteryCell','lithium_ion_battery_nmc_graphite.json'));
-        jsonstruct_lithium_plating = parseBattmoJson(fullfile('Examples','Advanced','Plating','lithium_plating.json'));
-
-        jsonstruct.use_thermal = false;
-        jsonstruct.include_current_collectors = false;
-
-        ne = 'NegativeElectrode'; pe = 'PositiveElectrode';
-        co = 'Coating'; am = 'ActiveMaterial'; lp = 'LithiumPlating';
-
-        jsonstruct.(ne).(co).(am).diffusionModelType = 'full';
-        jsonstruct.(pe).(co).(am).diffusionModelType = 'full';
-        jsonstruct.(ne).(co).(am).useLithiumPlating = true;
-        jsonstruct.(ne).(co).(am).isRootSimulationModel = true;
-
-        jsonstruct_lithium_plating.LithiumPlating.reactionRatePlating = reactionRatePlating;
-        jsonstruct_lithium_plating.LithiumPlating.reactionRateChemicalIntercalation = reactionRateChemicalIntercalation;
-        jsonstruct.(ne).(co).(am).LithiumPlating = jsonstruct_lithium_plating.LithiumPlating;
-
-        jsonstruct.Control.controlPolicy = 'CCCharge';
-
-        inputparams = BatteryInputParams(jsonstruct);
-        inputparams = inputparams.(ne).(co).(am);
-        inputparams.Interface.openCircuitPotential.functionname = 'computeOCP_Graphite_Latz';
-
-        model = ActiveMaterial(inputparams);
-        model = model.setupForSimulation();
-
-        sd = 'SolidDiffusion'; itf = 'Interface';
-        cElectrolyte = 0.5 * mol / litre;
-        phiElectrolyte = 0;
-        T = 298;
-
-        cInit = model.(itf).guestStoichiometry0 * model.(itf).saturationConcentration;
-        N = model.(sd).N;
-        initState.(sd).c = cInit * ones(N, 1);
-        initState.(sd).cSurface = cInit;
-
-        initState.T = T;
-        initState.(itf).cElectrolyte = cElectrolyte;
-        initState.(itf).phiElectrolyte = phiElectrolyte;
-        initState = model.evalVarName(initState, {itf, 'OCP'});
-        OCP = initState.(itf).OCP;
[…]
