+++ Examples/Advanced/Plating/platingSumup.m
@@ -1,359 +1,0 @@
-%% Run stand-alone active material model with lithium plating
-% This example shows how to simulate a single particle of a silicon graphite
-% electrode, taking into account the plating phenomenon
-
-clear all
-close all
-
-%% Setup the properties of Li-ion battery materials and cell design
-
-jsonstruct = parseBattmoJson(fullfile('ParameterData','BatteryCellParameters','LithiumIonBatteryCell','lithium_ion_battery_nmc_graphite.json'));
-
-inputparams = BatteryInputParams(jsonstruct);
-
-ne      = 'NegativeElectrode';
-pe      = 'PositiveElectrode';
-lp      = 'LithiumPlating';
-elyte   = 'Electrolyte';
-thermal = 'ThermalModel';
-co      = 'Coating';
-am      = 'ActiveMaterial';
-itf     = 'Interface';
-sd      = 'SolidDiffusion';
-ctrl    = 'Control';
-cc      = 'CurrentCollector';
-
-jsonstruct.use_thermal = false;
-jsonstruct.include_current_collectors = false;
-
-jsonstruct.(ne).(co).(am).diffusionModelType = 'full';
-jsonstruct.(pe).(co).(am).diffusionModelType = 'full';
-
-jsonstruct.(ne).(co).(am).useLithiumPlating = true;
-
-% Flag for stand-alone model
-jsonstruct.(ne).(co).(am).isRootSimulationModel = true;
-
-jsonstruct_lithium_plating = parseBattmoJson(fullfile('Examples', 'Advanced', 'Plating', 'lithium_plating.json'));
-
-jsonstruct.(ne).(co).(am).LithiumPlating = jsonstruct_lithium_plating.LithiumPlating;
-
-scenario = 'charge';
-
-%% following is not used at particle level
-
-switch scenario
-  case 'charge'
-    jsonstruct.Control.controlPolicy = 'CCCharge';
-  case 'discharge'
-    jsonstruct.Control.controlPolicy = 'CCDischarge';
-  otherwise
-    error('scenario not recognized');
-end
-
-
-% Setup InputParams
-inputparams = BatteryInputParams(jsonstruct);
-inputparams = inputparams.(ne).(co).(am);
-
-%OCP is computed via a function described in the article (S-5)
-inputparams.Interface.openCircuitPotential.functionname = 'computeOCP_Graphite_Latz';
-
-%% Setup the model
-
-model = ActiveMaterial(inputparams);
-
-%% Equip model for simulation
-
-model = model.setupForSimulation();
-
-%% Setup initial state
-
-sd  = 'SolidDiffusion';
-itf = 'Interface';
-
-cElectrolyte   = 5e-1*mol/litre;
-phiElectrolyte = 0;
-T              = 298;
-
-switch scenario
-  case 'charge'
-    % cElectrodeInit = (model.(itf).guestStoichiometry0)*(model.(itf).saturationConcentration);
-    cElectrodeInit = 29.9*mol/litre;
-  case 'discharge'
-    cElectrodeInit = (model.(itf).guestStoichiometry100)*(model.(itf).saturationConcentration);
-  otherwise
-    error('scenario not recognized');
-end
-
-N = model.(sd).N;
-initState.(sd).c        = cElectrodeInit*ones(N, 1);
-initState.(sd).cSurface = cElectrodeInit;
-
-initState.T = T;
-initState.(itf).cElectrolyte   = cElectrolyte;
-initState.(itf).phiElectrolyte = phiElectrolyte;
-
-initState = model.evalVarName(initState, {itf, 'OCP'});
-OCP = initState.(itf).OCP;
-initState.E = OCP + phiElectrolyte;
-
-
-F = model.(itf).constants.F;
[…]

[…]
