### 2026-03-02 f958c1b
clean-up for Lithium plating
