### 2026-03-02 3c60ea6
added lithium plating notebook in documentation
