### 2026-03-02 bdd7b4e
implementing plating for a full cell
