### 2026-03-02 6b4331b
added runPlatingCell
