### 2026-03-03 cfad672
clean-up for lithium plating setup in GenericBattery
