### 2026-03-03 c3cd7df
fixes for Lithium plating for whole cell
