### 2026-03-03 52d99fa
update in runPlatingCell
