+++ Examples/Advanced/Plating/runPlatingCell.m
@@ -1,4 +1,4 @@
-%% Run stand-alone active material model with lithium plating
+%% Lithium plating example
 %
 % In This example, we illustrate the implementation of Lithium plating.
 %
@@ -6,7 +6,6 @@
 %
 % Hein, Danner and Latz "An Electrochemical Model of Lithium Plating and Stripping in Lithium Ion Batteries" In: ACS Applied Energy Materials (2022) http://dx.doi.org/10.1021/acsaem.0c01155
 %
-
 
 %% Parameter setup
 %
@@ -89,240 +88,108 @@ jsonstruct_geometry = parseBattmoJson(filename);
 
 jsonstruct = mergeJsonStructs({jsonstruct, jsonstruct_geometry});
 
+%%
+% We change control to charge control
+%
+
+jsonstruct.Control.controlPolicy      = 'CCCharge';
+jsonstruct.Control.CRate              = 1;
+jsonstruct.SOC                        = 0.01;
+jsonstruct.Control.upperCutoffVoltage = 4.5;
+jsonstruct.Control.useCVswitch        = true;
+
+%%
+% We run the simulation
+%
+
 output = runBatteryJson(jsonstruct, 'runSimulation', true);
 
-return
-
-%%
-% We equip the model for simulation. Again, this step is done automatically for more main stream models.
-%
-
-model = model.setupForSimulation();
-
-%% Setup simulation schedule
-%
-% We define the simulation schedule. Increasing Iref strengthens the effect.
-
-Iref = 7e-13;
-Imax = Iref;
-total = 6e-3*hour*(Iref/Imax); % total time of the charge
-n     = 400;
-dt    = total/n;
-step  = struct('val', dt*ones(n, 1), 'control', ones(n, 1));
-
-tup = 1*second*(Iref/Imax);
-
-srcfunc = @(time) rampupControl(time, tup, -Imax);
-
-control.src = srcfunc;
-
-schedule = struct('control', control, 'step', step);
-
-%% Setup initial state
-
-sd  = 'SolidDiffusion';
-itf = 'Interface';
-
-cElectrolyte   = 5e-1*mol/litre;
-phiElectrolyte = 0;
-T              = 298;
-cElectrodeInit = 30*mol/litre;
-
-[model, initstate] = setupPlatingInitialState(model, T, cElectrolyte, phiElectrolyte, cElectrodeInit, Imax);
-
-%% Run first simulation, particle charge
-%
-
-inputSim = struct('model'    , model   , ...
-                  'schedule' , schedule, ...
-                  'initstate', initstate);
-simsetup = SimulationSetup(inputSim);
-
-states = simsetup.run();
-
-chargeStates = states; % for later
-
-%% Setup discharge simulation
-% 
-%
-% We use the last state of the previous simulation to initialise the new one
-
-simsetup.initstate = states{end};
-
-%%
-% Setup schedule structure for discharge
-%
-
-cmin = (model.(itf).guestStoichiometry0)*(model.(itf).saturationConcentration);
-
-control.stopFunction = @(model, state, state0_inner) (state.(sd).cSurface <= cmin);
-control.src          = @(time) rampupControl(time, tup, Imax);
-
-simsetup.schedule = struct('control', control, 'step', step);
-
-%% Run second simulation, particle discharge
-%
-
-states = simsetup.run();
-
-%%
-% We concatenate the two phases
-
-dischargeStates = states;
-states = vertcat(chargeStates, dischargeStates);
 
 %% Plotting
+%
+close all
 
-ind = cellfun(@(state) ~isempty(state), states);
-states = states(ind);
+states = output.states;
+model  = output.model;
 
-time     = cellfun(@(state) state.time, states);
-cSurface = cellfun(@(state) state.(sd).cSurface, states);
[…]
