### 2026-03-03 dd61721
minor fix in Lithium Plating
