### 2026-03-03 094a238
added cutoff for Lithium plating in generic battery
