+++ Tests/TestExamples/TestRunExamples.m
@@ -39,7 +39,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runSafariSEI'                                    , ...
                     'runBolaySEI'                                     , ...
                     'runBolaySEI0'                                    , ...
-                    'runPlating'                                      , ...
+                    'runSPMplating'                                      , ...
                     'runExampleStokes'                                , ...
                     'runProtonicMembrane'                             , ...
                     'runGasSupply'                                    , ...
