+++ Documentation/plating.rst
@@ -1,0 +1,27 @@
+===============
+Lithium Plating
+===============
+
+.. toctree::
+   :hidden:
+      
+   runSPMplating.ipynb
+   runCellPlating.ipynb
+   
+We illustrate in two notebooks the implementation of Lihium Plating. The governing equations are taken from :cite:`Hein_2020`
+
+This work has been mainly done by Victor Mota, who was an intern at SINTEF during Spring 2025. His report can be uploaded `here <_static/documents/InternshipReport_VictorMota.pdf>`_
+
+.. grid:: 1
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Lithum Plating - Particle <runSPMplating>`
+           
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Lithium Plating - Full Cell <runCellPlating>`
+           
+
