+++ Documentation/refs.bib
@@ -1,3 +1,18 @@
+@article{Hein_2020,
+  title =        {An Electrochemical Model of Lithium Plating and Stripping in Lithium Ion Batteries},
+  volume =       3,
+  ISSN =         {2574-0962},
+  url =          {http://dx.doi.org/10.1021/acsaem.0c01155},
+  DOI =          {10.1021/acsaem.0c01155},
+  number =       9,
+  journal =      {ACS Applied Energy Materials},
+  publisher =    {American Chemical Society (ACS)},
+  author =       {Hein, Simon and Danner, Timo and Latz, Arnulf},
+  year =         2020,
+  month =        jul,
+  pages =        {8519–8531}
+}
+
 @article{V_llestad_2019,
   title =        {Mixed proton and electron conducting double perovskite anodes for stable and efficient tubular proton
   ceramic electrolysers},
