### 2026-03-03 0e9aab9
added runCellPlating in test suite
