+++ Tests/TestExamples/TestRunExamples.m
@@ -39,7 +39,8 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runSafariSEI'                                    , ...
                     'runBolaySEI'                                     , ...
                     'runBolaySEI0'                                    , ...
-                    'runSPMplating'                                      , ...
+                    'runSPMplating'                                   , ...
+                    'runCellPlating'                                  , ...
                     'runExampleStokes'                                , ...
                     'runProtonicMembrane'                             , ...
                     'runGasSupply'                                    , ...
