+++ _sources/runCellPlating.nblink.txt
@@ -1,0 +1,3 @@
+{
+  "path" : "../Examples/Advanced/Plating/notebooks/runCellPlating.ipynb"
+}
