+++ Documentation/refs.bib
@@ -1,3 +1,43 @@
+@article{li2018effects,
+  title={Effects of stress dependent electrochemical reaction on voltage hysteresis of lithium ion batteries},
+  author={Li, Haoliang and Song, Yicheng and Lu, Bo and Zhang, Junqian},
+  journal={Applied Mathematics and Mechanics},
+  volume={39},
+  number={10},
+  pages={1453--1464},
+  year={2018},
+  publisher={Springer}
+}
+
+@article{Chandrasekaran_2011,
+  title =        {Analysis of the Lithium-Ion Insertion Silicon Composite Electrode/Separator/Lithium Foil Cell},
+  volume =       {MA2011-02},
+  ISSN =         {2151-2043},
+  url =          {http://dx.doi.org/10.1149/ma2011-02/7/389},
+  DOI =          {10.1149/ma2011-02/7/389},
+  number =       7,
+  journal =      {ECS Meeting Abstracts},
+  publisher =    {The Electrochemical Society},
+  author =       {Chandrasekaran, Rajeswari and Fuller, Thomas},
+  year =         2011,
+  month =        aug,
+  pages =        {389–389}
+}
+
+@article{Chandrasekaran_2010,
+  title =        {Analysis of Lithium Insertion/Deinsertion in a Silicon Electrode Particle at Room Temperature},
+  volume =       157,
+  ISSN =         {0013-4651},
+  url =          {http://dx.doi.org/10.1149/1.3474225},
+  DOI =          {10.1149/1.3474225},
+  number =       10,
+  journal =      {Journal of The Electrochemical Society},
+  publisher =    {The Electrochemical Society},
+  author =       {Chandrasekaran, Rajeswari and Magasinski, Alexandre and Yushin, Gleb and Fuller, Thomas F.},
+  year =         2010,
+  pages =        {A1139}
+}
+
 @article{Hein_2020,
   title =        {An Electrochemical Model of Lithium Plating and Stripping in Lithium Ion Batteries},
   volume =       3,
