+++ Control/TimeControlModel.m
@@ -3,38 +3,17 @@ classdef TimeControlModel < ControlModel
 
     properties
 
-        inputtype % type used to give the time control input. It can be
-                  % - table
-                  % - function
+        value % function setup (see Utilities/JsonSchemas/Function.schema.json) that returns the value for the control,
+              % either current or voltage, depending on the returned value of the type function
+        type % function setup (see Utilities/JsonSchemas/Function.schema.json) that returns the value for the control
+             % type,
+             % - 1 for current
+             % - 2 for voltage
 
-        %% property used in case of table input
+        %% helper properties
         
-        times        % Array with time value (should include at the end the end time, so that length(times) = length(durations) + 1)
-        durations    % Array with duration value
-        values       % Array with control value
-        controltypes % Array with control type. The convention is
-                     % - 1 for current
-                     % - 2 for voltage
-
-        
-        %% property used in case of function input
-
-        controlValueFunction % function of time for the control value
-        controlTypeFunction  % function of time for the type control value (1 : current, 2 : voltage)
-        
-        %% Helpers 
-
-        usetable    % true if table is used
-        usefunction % true if we use a matlab function
-
-        computeInput % function called to give update
-        
-        use_durations   % Setup when usetable is true
-        
-        % Function handlers instantiated from controlValueFunction controlTypeFunction  
-
-        controlValueFunc 
-        controlTypeFunc
+        getValue % function object (see Utilities/FunctionInterface/Function.m) which is setup from value property
+        getType % function object (see Utilities/FunctionInterface/Function.m) which is setup from type property
         
     end
 
@@ -44,39 +23,13 @@ classdef TimeControlModel < ControlModel
             
             model = model@ControlModel(inputparams);
 
-            fdnames = {'inputtype'           , ...
-                       'times'               , ...
-                       'durations'           , ...
-                       'values'              , ...
-                       'controltypes'        , ...
-                       'controlValueFunction', ...
-                       'controlTypeFunction'};
-
+            fdnames = {'value', ...
+                       'type'};
+            
             model = dispatchParams(model, inputparams, fdnames);
 
-            switch model.inputtype
-              case 'table'
-                model.usetable = true;
-              case 'function'
-                model.usefunction = true;
-              otherwise
-                error('input type not recognized');
-            end
-            
-            if model.usetable
-                
-                model.computeInput = @(t) model.computeInputFromTable(t);
-                
-            end
-
-            if model.usefunction
-
-                model.controlValueFunc = setupFunction(model.controlValueFunction);
-                model.controlTypeFunc  = setupFunction(model.controlTypeFunction);
-                
-                model.computeInput = @(t) model.computeInputFromFunction(t);
-
-            end
+            model.getValue = setupFunction(model.value);
+            model.getType  = setupFunction(model.type);
             
         end
         
@@ -100,119 +53,96 @@ classdef TimeControlModel < ControlModel
         end
 
         
-        <<<<<<< HEAD
         function [ctrlVal, ctrlType] = computeInput(model, t)
 
             ctrlVal  = model.getValue.eval(t);
             ctrlType = model.getType.eval(t);
-            =======
-            function [ctrlVal, ctrlType] = computeInputFromFunction(model, t)
+            
+        end
 
-                ctrlVal  = model.controlValueFunc(t);
-                ctrlType = model.controlTypeFunc(t);
+        function state = updateControlEquation(model, state)
+            
+            E        = state.E;
+            I        = state.I;            
+            ctrlVal  = state.ctrlVal;
+            ctrlType = state.ctrlType;
 
-                switch ctrlType
-                  case 1
-                    ctrlType = 'constantCurrent';
-                  case 2
[…]
