### 2026-03-04 c615262
fix in TimeControlModel after merging
