### 2026-03-04 a9c5c47
clean-up in documentation for the controls
