+++ Examples/Advanced/runGenericStepControl1.m
@@ -1,10 +1,36 @@
-% Clear the workspace and close open figures
-clear
-close all
+%% Example using generic control input
+%
 
-jsonstruct_material = parseBattmoJson(fullfile('ParameterData', 'BatteryCellParameters', 'LithiumIonBatteryCell', 'lithium_ion_battery_nmc_graphite.json'));
+%% input setup
+%
+
+%%
+% We load material parameter input
+jsonstruct_material = parseBattmoJson(fullfile('ParameterData', ...
+                                               'BatteryCellParameters', ...
+                                               'LithiumIonBatteryCell', ...
+                                               'lithium_ion_battery_nmc_graphite.json'));
+
+%%
+% We load geometry parameter input
+
+
 jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
-jsonstruct_control  = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'debugging_flat_generic_step_control.json'));
+
+%%
+% We load the generic control parameter we want to use
+
+jsonstruct_control  = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'generic_step_control_example1.json'));
+
+%%
+% We print it to command window
+
+viewJsonStruct(jsonstruct_control)
+
+%%
+% We remove the fields that are not relevant for this example. This is done to simplify the input parameters and focus
+% on the relevant aspects of the model.
+%
 
 jsonstruct_material = removeJsonStructFields(jsonstruct_material              , ...
                                              {'Control', 'DRate'}             , ...
@@ -13,65 +39,29 @@ jsonstruct_material = removeJsonStructFields(jsonstruct_material              , 
                                              {'Control', 'rampupTime'}        , ...
                                              {'Control', 'lowerCutoffVoltage'});
 
+%%
+% We merge all the input structures
+% 
 jsonstruct = mergeJsonStructs({jsonstruct_material, ...
                                jsonstruct_geometry, ...
                                jsonstruct_control});
 
-% We define some shorthand names for simplicity.
-ne      = 'NegativeElectrode';
-pe      = 'PositiveElectrode';
-elyte   = 'Electrolyte';
-thermal = 'ThermalModel';
-co      = 'Coating';
-am      = 'ActiveMaterial';
-itf     = 'Interface';
-sd      = 'SolidDiffusion';
-ctrl    = 'Control';
-cc      = 'CurrentCollector';
 
+%%
+% We do not include thermal effects and the current collectors in this example to simplify the model.
+%
 jsonstruct.use_thermal = false;
 jsonstruct.include_current_collectors = false;
 
-inputparams = BatteryInputParams(jsonstruct);
-
-%% We setup the battery geometry ("bare" battery with no current collector).
-
-[inputparams, gen] = setupBatteryGridFromJson(inputparams, jsonstruct);
-
-%%  The Battery model is initialized by sending inputparams to the Battery class constructor
-
-model = GenericBattery(inputparams);
-
-% timestep.timeStepDuration = 100;
-
-% step    = model.Control.setupScheduleStep(timestep);
-% control = model.Control.setupScheduleControl();
-
-% This control is used to set up the schedule
-schedule = model.Control.setupSchedule();
-
-%% Setup the initial state of the model
-% The initial state of the model is setup using the model.setupInitialState() method.
-
-initstate = model.setupInitialState();
-
-model.verbose = true;
-
-solver = NonLinearSolver();
-solver.maxIterations = 10;
-
-%% Run the simulation
[…]
