### 2026-03-04 d4dc51f
some fix in runGenericStepControl1
