+++ Examples/Advanced/runGenericStepControlCycle.m
@@ -19,7 +19,21 @@ jsonstruct_material = parseBattmoJson(fullfile('ParameterData'        , ...
 
 jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
 
+%%
+% We load the generic control parameter we want to use
+
 jsonstruct_control  = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'generic_step_control_cycle_example.json'));
+
+%%
+% We print it to command window
+
+viewJsonStruct(jsonstruct_control)
+
+%%
+% We remove the fields that are not relevant for this example. This is done to simplify the input parameters and focus
+% on the relevant aspects of the model.
+%
+
 
 jsonstruct_material = removeJsonStructFields(jsonstruct_material              , ...
                                              {'Control', 'DRate'}             , ...
@@ -28,65 +42,35 @@ jsonstruct_material = removeJsonStructFields(jsonstruct_material              , 
                                              {'Control', 'rampupTime'}        , ...
                                              {'Control', 'lowerCutoffVoltage'});
 
+%%
+% We merge all the input structures
+% 
+
 jsonstruct = mergeJsonStructs({jsonstruct_material, ...
                                jsonstruct_geometry, ...
                                jsonstruct_control});
 
-% We define some shorthand names for simplicity.
-ne      = 'NegativeElectrode';
-pe      = 'PositiveElectrode';
-elyte   = 'Electrolyte';
-thermal = 'ThermalModel';
-co      = 'Coating';
-am      = 'ActiveMaterial';
-itf     = 'Interface';
-sd      = 'SolidDiffusion';
-ctrl    = 'Control';
-cc      = 'CurrentCollector';
 
-jsonstruct.use_thermal = false;
+%%
+% We do not include thermal effects and the current collectors in this example to simplify the model.
+%
+
+jsonstruct.use_thermal                = false;
 jsonstruct.include_current_collectors = false;
 
-inputparams = BatteryInputParams(jsonstruct);
+%%
+% We run the simulation
+%
 
-%% We setup the battery geometry ("bare" battery with no current collector).
+output = runBatteryJson(jsonstruct);
 
-[inputparams, gen] = setupBatteryGridFromJson(inputparams, jsonstruct);
+%% Plotting
+% 
 
-%%  The Battery model is initialized by sending inputparams to the Battery class constructor
+states = output.states;
 
-model = GenericBattery(inputparams);
-
-% timestep.timeStepDuration = 100;
-
-% step    = model.Control.setupScheduleStep(timestep);
-% control = model.Control.setupScheduleControl();
-
-% This control is used to set up the schedule
-schedule = model.Control.setupSchedule();
-
-%% Setup the initial state of the model
-% The initial state of the model is setup using the model.setupInitialState() method.
-
-initstate = model.setupInitialState();
-
-model.verbose = true;
-
-solver = NonLinearSolver();
-solver.maxIterations = 10;
-
-%% Run the simulation
-[~, states, report] = simulateScheduleAD(initstate, model, schedule, 'OutputMinisteps', true, 'NonLinearSolver', solver);
-% output = runBatteryJson(jsonstruct);
-
-%% Process output and recover the output voltage and current from the output states.
-ind = cellfun(@(x) not(isempty(x)), states);
-states = states(ind);
 E = cellfun(@(x) x.Control.E, states);
 I = cellfun(@(x) x.Control.I, states);
-T = cellfun(@(x) max(x.(thermal).T), states);
[…]
