+++ Documentation/controlinput.rst
@@ -8,10 +8,11 @@ Control models
       
    runControlExamples.nblink
    runTimeControlExample.nblink
-
+   runGenericStepControlSimple.nblink
+   runGenericStepControlCycle.nblink
+      
 The battery :battmo:`model <Battery#19>` contains a *Control* sub-model, see the overall model structure
 :ref:`description<architecture:BattMo Model Architecture>`. The Control model determine the control *values* and *type*.
-
 
 At a **given time**, there are in general two control types:
 
@@ -28,6 +29,8 @@ The most standard controls can be called from BattMo using the json interface:
 * Constant Current Charge (CCCharge)
 * Constant Current Constant Voltage (CCCV)
 
+We also have a generic control setup, see example below.
+
 The parameters for each of the model are described in the json schema :battmofile:`ControlModel.schema.json
 <Utilities/JsonSchemas/ControlModel.schema.json>`. 
    
@@ -42,4 +45,14 @@ The parameters for each of the model are described in the json schema :battmofil
       :padding: 2
 
       :ref:`Time Functional Control Examples <runTimeControlExample>`
-   
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Generic Control Simple Example <runGenericStepControlSimple>`
+              
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Generic Control Cycle Example <runGenericStepControlCycle>`
+              
