+++ _sources/controlinput.rst.txt
@@ -2,9 +2,17 @@
 Control models
 ==============
 
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+      
+   runControlExamples.nblink
+   runTimeControlExample.nblink
+   runGenericStepControlSimple.nblink
+   runGenericStepControlCycle.nblink
+      
 The battery :battmo:`model <Battery#19>` contains a *Control* sub-model, see the overall model structure
 :ref:`description<architecture:BattMo Model Architecture>`. The Control model determine the control *values* and *type*.
-
 
 At a **given time**, there are in general two control types:
 
@@ -15,193 +23,36 @@ Given a control type, the control *value* can change in time. For the most stand
 can be used. We plan to include there more control models. Below, we give some short explanations on how a control model
 can be implemented.
 
-
-Json input control interface
-============================
-
-
 The most standard controls can be called from BattMo using the json interface:
 
 * Constant Current Discharge (CCDischarge)
 * Constant Current Charge (CCCharge)
 * Constant Current Constant Voltage (CCCV)
 
+We also have a generic control setup, see example below.
+
 The parameters for each of the model are described in the json schema :battmofile:`ControlModel.schema.json
-<Utilities/JsonSchemas/ControlModel.schema.json>`. Let us review them by looking at examples. For all controls, we have
-a given :code:`CRate`. The property :code:`controlPolicy` is used to set the control model.
-
-**Constant Current Discharge**
-
-Before the simulation starts, the capacity of the battery is computed. The battery capacity value is used to compute the
-given input current from the CRate value. For a constant current discharge, we have a lower cutoff voltage where the
-simulation stops.  The parameter :code:`useCVswitch` can be used if we want to continue the simulation until a given
-fixed time, by switching to voltage control. At the start of the simulation, strong non-linearities in the equations may
-create difficulties to reach the activation potential. For this reason, there is the possibility to use a *rampup time*
-where the current is gradually increased from zero to the targeted given input current. 
-
-.. code:: json
-
-   "Control": {
-    "controlPolicy": "CCDischarge",
-    "rampupTime": 0.1,
-    "CRate": 1,
-    "lowerCutoffVoltage": 2.4
-  }
-
-For the dataset given :battmofile:`here
-<ParameterData/BatteryCellParameters/LithiumIonBatteryCell/lithium_ion_battery_nmc_graphite.json>`, we obtain the
-following result.
-
-.. figure:: img/controlExampleCCDischarge.png
-   :target: _images/controlExampleCCDischarge.png
-   :width: 70%
-   :align: center
-
-   Constant Current Discharge control
-      
-**Constant Current Charge**
-
-In the same way as the discharge case, the capacity of the battery is computed first to set the given input current from
-the CRate. When the upper cuttoff voltage limit is reached, the control type is switched to voltage control. Also in
-this case, we can include a rampup time. As an example,
-
-
-.. code:: json
-
-   "Control": {
-    "controlPolicy": "CCCharge",
-    "rampupTime": 0.1,
-    "CRate": 1,
-    "upperCutoffVoltage": 4.1,
-  }
-
-For the dataset given :battmofile:`here
-<ParameterData/BatteryCellParameters/LithiumIonBatteryCell/lithium_ion_battery_nmc_graphite.json>`, we obtain the
-following result.
-
-.. figure:: img/controlExampleCCCharge.png
-   :target: _images/controlExampleCCCharge.png
-   :width: 70%
-   :align: center
-
-   Constant Current Charge control
-
-**Constant Current Constant Charge**
-
-An example of input is
-
-.. code:: json
-
[…]
