+++ Documentation/battmointernals.rst
@@ -1,0 +1,30 @@
+================
+BattMo Internals
+================
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   modelinitialisation
+   architecture
+   Computational Graph <computationalGraph/graphdoc>
+
+.. grid:: 2
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`The Battery Simulation Model<modelinitialisation:The Battery Simulation Model>`
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`BattMo Model Architecture<architecture:BattMo Model Architecture>`
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Computational Graph<computationalGraph/graphdoc:Computation Graph Model Design ang Assembly>`
+                                
+
