+++ Documentation/inputparameters.rst
@@ -1,0 +1,42 @@
+================
+Input parameters
+================
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   json
+   mergejsonstruct
+   functioninterface
+   controlinput
+   parsets
+
+.. grid:: 2
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`JSON input specification<json:JSON input specification>`
+          
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Merging Parameters<mergejsonstruct:Merging parameters>`
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Functional Parameters<functioninterface:Functional parameters>`
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Control models<controlinput:Control models>`
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Parameter Sets <parsets:Parameter sets>`
+           
+           
