+++ Documentation/various.rst
@@ -1,0 +1,22 @@
+===================
+Various Information
+===================
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   units
+   Octave <octave>
+
+.. grid-item-card::
+   :padding: 2
+
+   :ref:`Units <units:Units>`
+        
+.. grid-item-card::
+   :padding: 2
+
+   :ref:`Octave Support <octave:Note on Octave Support>`
+           
+
