+++ _sources/intermediate.rst.txt
@@ -1,96 +1,0 @@
-==================
-Intermediate Usage
-==================
-
-.. toctree::
-   :maxdepth: 2
-   :hidden:
-
-   mergejsonstruct
-   functioninterface
-   modelinitialisation
-   controlinput
-   soliddiffusion
-   parsets
-   units
-   runThermalExample.nblink
-   runOnlyThermal.nblink
-   runSiliconGraphiteBattery.ipynb
-   runSwellingExample.ipynb
-   plating
-   optimisation
-   Octave <octave>
-
-.. grid:: 2
-
-   .. grid-item-card::
-      :padding: 2
-      
-      :ref:`Merging Parameters<mergejsonstruct:Merging parameters>`
-
-   .. grid-item-card::
-      :padding: 2
-      
-      :ref:`Functional Parameters<functioninterface:Functional parameters>`
-                                
-   .. grid-item-card::
-      :padding: 2
-      
-      :ref:`The Battery Simulation Model<modelinitialisation:The Battery Simulation Model>`
-
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Control models <controlinput:Control models>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Solid Diffusion Models <soliddiffusion:Solid Diffusion Models>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Parameter Sets <parsets:Parameter sets>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Units <units:Units>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Thermal Simulation <runThermalExample>`
-
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Iso-thermal and Fully-coupled Simulation <runOnlyThermal>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Composite Material <runSiliconGraphiteBattery>`
-
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Silicon Swelling <runSwellingExample>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Lithium Plating <plating:Lithium Plating>`
-
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Optimisation <optimisation:Optimization>`
-           
-   .. grid-item-card::
-      :padding: 2
-
-      :ref:`Octave Support <octave:Note on Octave Support>`
-           
-
