+++ _sources/batteryphysics.rst.txt
@@ -1,0 +1,55 @@
+===============
+Battery Physics
+===============
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   runThermalExample.nblink
+   runOnlyThermal.nblink
+   runBolaySEI.ipynb
+   runSiliconGraphiteBattery.ipynb
+   plating
+   runSwellingExample.ipynb
+   soliddiffusion
+
+.. grid:: 2
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Thermal Simulation <runThermalExample>`
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Iso-thermal and Fully-coupled Simulation <runOnlyThermal>`
+           
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`SEI formation <runBolaySEI>`
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Composite Material <runSiliconGraphiteBattery>`
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Lithium Plating <plating:Lithium Plating>`
+
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Silicon Swelling <runSwellingExample>`
+           
+   .. grid-item-card::
+      :padding: 2
+
+      :ref:`Solid Diffusion Models <soliddiffusion:Solid Diffusion Models>`
+           
+
+
