+++ _sources/various.rst.txt
@@ -1,0 +1,30 @@
+===================
+Various Information
+===================
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   units
+   Octave <octave>
+   juliabridge
+
+.. grid:: 2
+      
+   .. grid-item-card::
+      :padding: 2
+   
+      :ref:`Units <units:Units>`
+           
+   .. grid-item-card::
+      :padding: 2
+   
+      :ref:`Octave Support <octave:Note on Octave Support>`
+   
+   .. grid-item-card::
+      :padding: 2
+   
+      :ref:`BattMo Julia Bridge <juliabridge:BattMo Julia bridge>`
+                   
+
