+++ _sources/index.rst.txt
@@ -6,21 +6,23 @@
 .. toctree::
    :hidden:
 
-   Installation <installation>
-   basicusage
-   tutorials
-   intermediate
-   modeling
-   architecture
-   json
-   geometryinput
-   juliabridge
-   Electrolyser simulation <publishedExamples/runElectrolyser>
-   protonicmembrane
-   app
-   Computational Graph <computationalGraph/graphdoc>
-   seealso
-   References <bibliography>
+   ..
+      Installation <installation>
+      basicusage
+      tutorials
+   inputparameters
+   ..
+      geometryinput
+      batteryphysics
+      optimisation
+      modeling
+      battmointernals
+      app
+      various
+      Electrolyser simulation <publishedExamples/runElectrolyser>
+      protonicmembrane
+      seealso
+      References <bibliography>
 
 
 Welcome
