### 2026-03-04 f2867ef
fix script name in test suite
