+++ Tests/TestExamples/TestRunExamples.m
@@ -45,8 +45,8 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runProtonicMembrane'                             , ...
                     'runGasSupply'                                    , ...
                     'runProtonicCell'                                 , ...
-                    'runGenericStepControl1'                          , ...
-                    'runGenericStepControl2'                          , ...
+                    'runGenericStepControlSimple'                     , ...
+                    'runGenericStepControlCycle'                      , ...
                     'runIntelligent'                                  , ...
                     'runSiGrLnmoScript'                               , ...
                     'runTestJsonInputs'                               , ...
