+++ Documentation/conf.py
@@ -127,6 +127,7 @@ author = 'Simon Clark'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
+# exclude_patterns = ['_build', '*.nblink', 'publishedExamples', 'pynbnotebooks']
 exclude_patterns = ['_build']
 
 # The reST default role (used for this markup: `text`) to use for all
