### 2026-03-05 415b369
added runJsonExamples in test suite
