+++ Examples/Documentation/runJsonExamples.m
@@ -1,0 +1,3 @@
+jsonstruct = parseBattmoJson('Examples/Documentation/jsonfiles/Example/simulation.json');
+
+output = runBatteryJson(jsonstruct);
