+++ Tests/TestExamples/TestRunExamples.m
@@ -10,6 +10,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runBatteryP3D'                                   , ...
                     'runBatteryP4D'                                   , ...
                     'runBatteryPrecondTestP2D'                        , ...
+                    'runJsonExamples'                                 , ...
                     'runChen2020'                                     , ...
                     'runCR'                                           , ...
                     'runGittTest'                                     , ...
