+++ Documentation/json.rst
@@ -1,4 +1,3 @@
-        
 ========================
 JSON input specification
 ========================
