+++ Documentation/jsonexample.rst
@@ -2,7 +2,16 @@
 My First Json Input
 ===================
 
-In this page, we present example of json inputs. The parameter names are hopefully chosen in a way that will be easily understood by battery modelers.
+In this page, we present example of json inputs. The parameter names haven hopefully chosen in a way that they will be
+easily understood by battery modelers. Their description is included in the json schemas, see some links include below,
+and this documentation :ref:`this page<json:JSON input specification>`.
+
+The json input follows the battery model hierarchy as described in :ref:`BattMo Model Architecture<architecture:BattMo
+Model Architecture>`. The input contains also simulation specific parameters, such as solver parameters and time
+stepping.
+
+The keywords :code:`isFile`, :code:`filename` and :code:`filenames` are recognized by the parser
+(:code:`parseBattmoJson`) and used to include inputs from other files, as shown in the next example
 
 Simulation
 ==========
@@ -10,11 +19,29 @@ Simulation
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/simulation.json
    :language: json
 
+The keyword :code:`isFile` is recognized. The value of :code:`filenames` contain an array with a list of the files that
+we want to include. Those files are also the ones that are presented in this page (all the files are in
+:battmofile:`this directory <Examples/Documentation/jsonfiles/Example/>`, which comes with your battmo installation)
+
+The path of the file is first tested as such but if no file is found, the parser will check the path that is obtained
+relative to your battmo installation path.
+
+You can actually run this illustration example using
+
+.. code:: matlab
+
+   jsonstruct = parseBattmoJson('Examples/Documentation/jsonfiles/Example/simulation.json');
+   output = runBatteryJson(jsonstruct);
+
 Battery
 =======
 
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/battery.json
    :language: json
+
+Here, we recognize the battery hierarchy where the sumodels are the negative and positive electrodes, the electrolyte,
+the separator and the thermal model, if thermal effects are included. The initial temperature and state of charge are
+also included here.
 
 Negative Electrode
 ==================
