+++ Documentation/jsonexample.rst
@@ -49,11 +49,18 @@ Negative Electrode
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/negativeElectrode.json
    :language: json
 
+The electrode consists of a current collector and a coating. They are fetched from files. Here, we show only the
+negative electrode.
+
 Coating
 =======
 
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/coating.json
    :language: json
+
+The coating consists of the active material, the binder and the conducting additive. The coating has also effective
+parameters that are set at this level, such as the :code:`effectiveDensity` and the :code:`bruggemanCoefficient`. The
+effective density is used to compute the porosity from the properties of the three coating components.
 
 Active Material
 ===============
@@ -61,29 +68,62 @@ Active Material
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/activeMaterial.json
    :language: json
 
+The submodels in the active material are the **interface** which collects the parameters related to the surface reaction
+between the electode and the electrolyte and the **solid diffusion** model which collects the the parameters related to
+the solid diffusion in the pseudo particle.
+
 Interface
 =========
 
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/interface.json
    :language: json
-                            
+
+We do not explain in detail here all the parameters. We hope the names are chosen so that they will be easily
+recognizable by battery modellers. Again, we refer the :ref:`json schema <json:Interface>` for a reference, see the
+:code:`description` key.
+
+The most important parameter of the electrochemical reaction is certainly the **open circuit potential**. The open
+circuit potential is a function. In this example, the function is implemented as a matlab function and we call
+:code:`named function` this type of format.
+
+Functional parameters enter the DFN model at several places and we have implemented a specific input for those which is
+explained in details in :ref:`Functional Parameters <functioninterface:Functional parameters>`
+              
 Solid Diffusion
 ===============
 
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/solidDiffusion.json
    :language: json
-                            
+
+The particle radius and the diffusion coefficient governs the time scale of the diffusion. The volumetric surface
+parameter is an important part of the volume averaging approach the DFN model uses, and it is not easy to estimate.
+
 Binder
 ======
 
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/binder.json
    :language: json
-                            
+
+The binder contains intrinsic parameters that are used to compute the effective coating material parameters for thermal
+and electronic conductivity but no other process are modeled at this level. 
+              
 Conducting Additive
 ===================
 
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/conductingAdditive.json
    :language: json
+
+As binder, conductive additive contains intrinsic parameters that are used to compute the effective coating material parameters for thermal
+and electronic conductivity but no other process are modeled at this level.
+              
+Electrolyte
+===========
+
+.. literalinclude:: ../Examples/Documentation/jsonfiles/Example/electrolyte.json
+   :language: json
+
+This is an other example where we see functional parameters, the ionic conductivity and the diffusion coefficient, which
+are functions of concentration and temperature.
 
 Current Collector
 =================
@@ -91,25 +131,16 @@ Current Collector
 .. literalinclude:: ../Examples/Documentation/jsonfiles/Example/currentCollector.json
    :language: json
 
-Positive Electrode
-==================
-
-Here, we provide all the parameter in a single json structure without linking to other files.
-
-.. literalinclude:: ../Examples/Documentation/jsonfiles/Example/positiveElectrode.json
-   :language: json
-
-Electrolyte
-===========
-
[…]
