+++ Documentation/various.rst
@@ -1,30 +1,39 @@
-===================
-Various Information
-===================
+======================
+Programming Interfaces
+======================
+
+BattMo has a `Julia version <https://battmo.org/BattMo.jl/dev/>`_ which does not support all the same features but runs
+faster. It is possible to setup a **julia bridge** use the julia solver from within matlab.
+
+An web app which uses the julia version is available for the user to test the functionalities directly in a web browser.
+
+BattMo should support Octave but all functionalities are not systematically tested.
 
 .. toctree::
    :maxdepth: 2
    :hidden:
 
-   units
-   Octave <octave>
    juliabridge
+   app
+   Octave <octave>
 
 .. grid:: 2
-      
+
    .. grid-item-card::
       :padding: 2
    
-      :ref:`Units <units:Units>`
+      :ref:`BattMo Julia Bridge <juliabridge:BattMo Julia bridge>`
+
+   .. grid-item-card::
+      :padding: 2
+   
+      :ref:`BattMoApp <app:BattMoApp>`
            
    .. grid-item-card::
       :padding: 2
    
       :ref:`Octave Support <octave:Note on Octave Support>`
    
-   .. grid-item-card::
-      :padding: 2
-   
-      :ref:`BattMo Julia Bridge <juliabridge:BattMo Julia bridge>`
+
                    
 
