+++ Documentation/index.rst
@@ -12,13 +12,11 @@
    inputparameters
    geometryinput
    batteryphysics
+   elchemsystems
    optimisation
    modeling
    battmointernals
-   app
-   various
-   Electrolyser simulation <publishedExamples/runElectrolyser>
-   protonicmembrane
+   programminginterfaces
    seealso
    References <bibliography>
 
