### 2026-03-06 00d3371
added script for documentation on exploring output
