+++ Examples/Documentation/exploreOutput.m
@@ -1,0 +1,7 @@
+%%
+% Explore BattMo output
+%
+
+jsonstruct = parseBattmoJson('Examples/Documentation/jsonfiles/explore_output_example.json');
+output = runBatteryJson(jsonstruct);
+
