### 2026-03-06 9339ccc
added explore output
