+++ Tests/TestExamples/TestRunExamples.m
@@ -5,6 +5,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
         % List of all the supported scripts
 
         filename = {'battMoTutorial'                                  , ...
+                    'exploreOutput'                                   , ...
                     'runBatteryMLP'                                   , ...
                     'runBatteryP2D'                                   , ...
                     'runBatteryP3D'                                   , ...
