### 2026-03-06 99d51db
update in a exploreOutput script
