### 2026-03-06 c68604f
added comment in conf.py
