+++ _sources/inputparameters.rst.txt
@@ -81,6 +81,11 @@ achieved using the function `mergeJsonStructs`, see :ref:`Merging Parameters<mer
    .. grid-item-card::
       :padding: 2
 
+      :ref:`Units <units:Units>`
+                      
+   .. grid-item-card::
+      :padding: 2
+
       :ref:`Parameter Sets <parsets:Parameter sets>`
            
            
