+++ _sources/output.rst.txt
@@ -1,0 +1,32 @@
+=============
+BattMo output
+=============
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   exploreOutput.nblink
+   staticspec
+
+Simulation output visualization and postprocessing are best illustrated in the :ref:`Tutorials <tutorials:Tutorials>`.
+
+Here, we provide a very short overview of the output structure returned by :code:`runBatteryJson`.
+.
+
+We give also an overview of simple routines to compute _static_ cell specifications. By static, we mean specifications
+that do not require any simulation such as mass, capacity or nominal energy.
+
+.. grid:: 2
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Simulation Output Structure<exploreOutput:Simulation Output Structure>`
+
+   .. grid-item-card::
+      :padding: 2
+      
+      :ref:`Cell Specifications<staticspec:Cell Specifications>`
+
+
