+++ Battery/GenericBattery.m
@@ -33,7 +33,7 @@ classdef GenericBattery < BaseModel
         use_thermal
         include_current_collectors
         include_swelling % true if swelling is included
-        
+
         % We write here the jsonstruct that has been process when constructing inputparams
         jsonstruct
 
@@ -156,7 +156,7 @@ classdef GenericBattery < BaseModel
                 error('controlPolicy not recognized');
             end
 
-            
+
             %% Temperature dispatch functions
             fn = @GenericBattery.updateTemperature;
 
@@ -240,7 +240,7 @@ classdef GenericBattery < BaseModel
                 model = model.registerPropFunction({{ctrl, 'ctrlType'}, fn, inputnames});
               case 'timeControl'
                 model = model.registerPropFunction({{ctrl, 'ctrlVal'}, fn, {'time'}});
-                model = model.registerPropFunction({{ctrl, 'ctrlType'}, fn, {'time'}});                
+                model = model.registerPropFunction({{ctrl, 'ctrlType'}, fn, {'time'}});
               case {'Generic'}
                 fn = @GenericBattery.updateControlTime;
                 model = model.registerPropFunction({{ctrl, 'time'}, fn, {'time'}});
@@ -264,9 +264,9 @@ classdef GenericBattery < BaseModel
                 for ielde = 1 : numel(eldes)
                     elde = eldes{ielde};
                     if model.(elde).coatingModelSetup.swelling
-                        inputNames = horzcat(inputNames, {{elde, co, 'volumeFraction'}});                        
+                        inputNames = horzcat(inputNames, {{elde, co, 'volumeFraction'}});
                     end
-                end                
+                end
                 model = model.registerPropFunction({{elyte,'volumeFraction'}, fn, inputNames});
 
                 fn = @GenericBattery.updateSwellingElectrolyteConvFlux;
@@ -276,11 +276,11 @@ classdef GenericBattery < BaseModel
                     if model.(elde).coatingModelSetup.swelling
                         inputNames = horzcat(inputNames, {{elde, co, am, sd, 'x'}});
                     end
-                end                
+                end
                 model = model.registerPropFunction({{elyte,'convFlux'}, fn, inputNames});
-                
+
             end
-            
+
             %% Function that update the Thermal Ohmic Terms
 
             if model.use_thermal
@@ -620,7 +620,7 @@ classdef GenericBattery < BaseModel
               case "Generic"
 
                 control = GenericControlModel(inputparams);
-                
+
               otherwise
 
                 error('Error controlPolicy not recognized');
@@ -755,7 +755,7 @@ classdef GenericBattery < BaseModel
             sep   = 'Separator';
 
             eldes = {ne, pe};
-            
+
             elyte_cells = zeros(model.G.getNumberOfCells(), 1);
             elyte_cells(inputparams.(elyte).G.mappings.cellmap) = (1 : inputparams.(elyte).G.getNumberOfCells())';
 
@@ -858,24 +858,24 @@ classdef GenericBattery < BaseModel
                         state.(amc).(sd).x = theta;
                         state = compmodel.evalVarName(state, {'volumeFraction'});
                         state = compmodel.evalVarName(state, {amc, sd, 'radiusElongation'});
-                        
+
                         re = state.(am).(sd).radiusElongation;
                         vf = state.volumeFraction;
 
                         N  = model.(elde).(co).(amc).(sd).N;
                         np = model.(elde).(co).(amc).(sd).np; % Note : we have by construction np = nc
-                        
+
                         c = model.(elde).(co).maximumTotalConcentration/vf*theta;
 
[…]
