### 2026-03-23 7a2f348
added conductivity in electrolyser
