### 2026-03-23 67134f0
fix in output of mergeJsonStructs
