+++ Utilities/JsonUtils/mergeJsonStructs.m
@@ -75,8 +75,15 @@ function jsonstruct = mergeJsonStructs(jsonstructs, varargin)
             elseif ~isstruct(jsonstruct.(fd2)) && ~isstruct(jsonstruct2.(fd2)) && isequal(jsonstruct.(fd2), jsonstruct2.(fd2))
                 % ok. Both are given but same values
             elseif opt.warn
-                varname = horzcat(tree, fd2);
-                fprintf('mergeJsonStructs: Parameter %s is assigned twice with different values %g and %g. Value from first jsonstruct is used.\n', strjoin(varname, '.'), jsonstruct.(fd2), jsonstruct2.(fd2));
+                [val1, isConverted]  = convertUnitBattMo(jsonstruct.(fd2));
+                [val2, isConverted2] = convertUnitBattMo(jsonstruct2.(fd2));
+                if isnumeric(val1) && isnumeric(val2)
+                    varname = horzcat(tree, fd2);
+                    fprintf('mergeJsonStructs: Parameter %s is assigned twice with different values %g and %g, relative difference is %g. Value from first jsonstruct is used.\n', strjoin(varname, '.'), val1, val2, abs(val1 - val2)/(abs(val1) + abs(val2)));
+                else
+                    varname = horzcat(tree, fd2);
+                    fprintf('mergeJsonStructs: Parameter %s is assigned twice with different values. Value from first jsonstruct is used.\n', strjoin(varname, '.'));
+                end
             end
         end
     end
