### 2026-03-24 5961280
runOnlyThermalExtended
