### 2026-03-25 596af48
sort distance in dependency list in ComputationalGraphInteractiveTool
