### 2026-03-25 f9ec814
work on adding exchange reaction at equilibrium conditions
