+++ Electrolyser/Models/Electrolyser.m
@@ -73,7 +73,9 @@ classdef Electrolyser < BaseModel
                 inputvarnames{end + 1} = {elde, exr, 'H2OexchangeRate'};
                 inputvarnames{end + 1} = {elde, ctl, 'inmrH2Osource'};
                 inputvarnames{end + 1} = {elde, ctl, 'inmrOHsource'};
-                inputvarnames{end + 1} = {elde, exr, 'OHexchangeRate'};
+                if ~model.(elde).useEquilibrium
+                    inputvarnames{end + 1} = {elde, exr, 'OHexchangeRate'};
+                end
             end
             model = model.registerPropFunction({{inm, 'H2OSource'}, fn, inputvarnames});
             model = model.registerPropFunction({{inm, 'OHsource'}, fn, inputvarnames});
@@ -87,12 +89,29 @@ classdef Electrolyser < BaseModel
                 for ilayer = 1 : numel(layers)
                     elde = eldes{ielde};
                     layer = layers{ilayer};
-                    model = model.registerPropFunction({{elde, layer, 'cOHinmr'}, fn, inputvarnames});
+                    switch layer
+                      case exr
+                        if ~model.(elde).useEquilibrium 
+                            model = model.registerPropFunction({{elde, layer, 'cOHinmr'}, fn, inputvarnames});
+                        end
+                      case ctl
+                        model = model.registerPropFunction({{elde, layer, 'cOHinmr'}, fn, inputvarnames});
+                      otherwise
+                        error('layer not recognized');
+                    end
                     model = model.registerPropFunction({{elde, layer, 'phiInmr'}, fn, inputvarnames});
                     model = model.registerPropFunction({{elde, layer, 'H2OaInmr'}, fn, inputvarnames});
                 end
             end
-
+            
+            fn = @Electrolyser.updateDispatchEquations;
+            for ielde = 1 : numel(eldes)
+                elde = eldes{ielde};
+                if model.(elde).useEquilibrium
+                    model = model.registerPropFunction({{elde, exr, 'dispatchEquation'}, fn, {{elde, exr, 'cOHinmr'}, {inm, 'cOH'}}});
+                end
+            end
+                
             fn = @Electrolyser.setupControl;
             fn = {fn, @PropFunction.drivingForceFuncCallSetupFn};
             inputvarnames = {{oer, ctl, 'I'}, ...
@@ -119,7 +138,8 @@ classdef Electrolyser < BaseModel
             ptl = 'PorousTransportLayer';
             ctl = 'CatalystLayer';
             inm = 'IonomerMembrane';
-
+            exr = 'ExchangeReaction';
+            
             con = model.con;
 
             pGas = 101325*Pascal; % pressure of active gas (O2 or H2)
@@ -141,6 +161,19 @@ classdef Electrolyser < BaseModel
 
                 elde = eldes{ielde};
 
+                if model.(elde).useEquilibrium
+                    
+                    % We need to setup OH initial concentration in ExchangeRate model
+                    coupterms = model.couplingTerms;
+                    coupnames = model.couplingNames;
+                    coupname  = sprintf('%s-%s', elde, inm);
+                    coupterm  = getCoupTerm(coupterms, coupname, coupnames);
+                    coupcells = coupterm.couplingcells;
+                    ncexr = size(coupcells(:, 1), 1); % same as size(coupcells(:, 2))
+                    state.(elde).(exr).cOHinmr = cOH*ones(ncexr, 1);
+                    
+                end
+                
                 % Compute density and partial molar volume using dedicated functions
[…]
