### 2026-03-25 3f94159
fixes for equilibrium exchange reaction
