### 2026-03-25 4a24af8
change in dispatch equilibrium equation
