### 2026-03-25 34eb549
input support for exchange reaction equilibrium
