### 2026-03-25 9294c78
fix nan initialization in a method in Electrolyser
