### 2026-03-25 6963647
DEBUG : runElectrolyser
