### 2026-03-26 6be1c84
Improve error message
