+++ Tests/TestExamples/TestRunExamples.m
@@ -13,6 +13,8 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runBatteryPrecondTestP2D'                        , ...
                     'runJsonDocExample'                               , ...
                     'runChen2020'                                     , ...
+                    'runImpedanceChen'                                , ...
+                    'runImpedanceScript'                              , ...
                     'runCR'                                           , ...
                     'runGittTest'                                     , ...
                     'runJellyRoll'                                    , ...
