### 2026-04-08 78c7a2b
some clean-up in impendance scripts
