+++ Documentation/batteryphysics.rst
@@ -13,7 +13,8 @@ Battery Physics
    plating
    runSwellingExample.ipynb
    soliddiffusion
-
+   runImpedanceScript.nblink
+   
 .. grid:: 2
 
    .. grid-item-card::
@@ -51,5 +52,9 @@ Battery Physics
 
       :ref:`Solid Diffusion Models <soliddiffusion:Solid Diffusion Models>`
            
+   .. grid-item-card::
+      :padding: 2
 
+      :ref:`Battery Impedance Computation <runImpedanceScript>`
+           
 
