### 2026-04-09 514a82a
Minor changes for adjoint usage
