+++ Documentation/JOSS/paper.bib
@@ -137,3 +137,43 @@ year={2025},
     title = {{GNU Octave}},
     url       = {https://octave.org}
     }
+
+@article{Hein2020,
+  title = {An Electrochemical Model of Lithium Plating and Stripping in Lithium Ion Batteries},
+  volume = {3},
+  ISSN = {2574-0962},
+  DOI = {10.1021/acsaem.0c01155},
+  number = {9},
+  journal = {ACS Applied Energy Materials},
+  publisher = {American Chemical Society (ACS)},
+  author = {Hein,  Simon and Danner,  Timo and Latz,  Arnulf},
+  year = {2020},
+  month = jul,
+  pages = {8519–8531}
+}
+
+@article{Safari2009,
+  title = {Multimodal Physics-Based Aging Model for Life Prediction of Li-Ion Batteries},
+  volume = {156},
+  ISSN = {0013-4651},
+  DOI = {10.1149/1.3043429},
+  number = {3},
+  journal = {Journal of The Electrochemical Society},
+  publisher = {The Electrochemical Society},
+  author = {Safari,  M. and Morcrette,  M. and Teyssot,  A. and Delacourt,  C.},
+  year = {2009},
+  pages = {A145}
+}
+
+@article{Bolay2022,
+  title = {Microstructure-resolved degradation simulation of lithium-ion batteries in space applications},
+  volume = {14},
+  ISSN = {2666-2485},
+  DOI = {10.1016/j.powera.2022.100083},
+  journal = {Journal of Power Sources Advances},
+  publisher = {Elsevier BV},
+  author = {Bolay,  Linda J. and Schmitt,  Tobias and Hein,  Simon and Mendoza-Hernandez,  Omar S. and Hosono,  Eiji and Asakura,  Daisuke and Kinoshita,  Koichi and Matsuda,  Hirofumi and Umeda,  Minoru and Sone,  Yoshitsugu and Latz,  Arnulf and Horstmann,  Birger},
+  year = {2022},
+  month = mar,
+  pages = {100083}
+}
