+++ Documentation/JOSS/paper.bib
@@ -177,3 +177,44 @@ year={2025},
   month = mar,
   pages = {100083}
 }
+
+@article{Planden2025,
+  title = {PyBOP: A Python package for battery model optimisation and parameterisation},
+  volume = {10},
+  ISSN = {2475-9066},
+  url = {http://dx.doi.org/10.21105/joss.07874},
+  DOI = {10.21105/joss.07874},
+  number = {116},
+  journal = {Journal of Open Source Software},
+  publisher = {The Open Journal},
+  author = {Planden,  Brady and Courtier,  Nicola E. and Robinson,  Martin and Khetarpal,  Agriya and Planella,  Ferran Brosa and Howey,  David A.},
+  year = {2025},
+  month = dec,
+  pages = {7874}
+}
+
+@article{fenics,
+  doi = {10.11588/ANS.2015.100.20553},
+  author = {Alnæs,  Martin and Blechta,  Jan and Hake,  Johan and Johansson,  August and Kehlet,  Benjamin and Logg,  Anders and Richardson,  Chris and Ring,  Johannes and Rognes,  Marie E and Wells,  Garth N},
+  language = {en},
+  title = {The FEniCS Project Version 1.5},
+  journal = {Archive of Numerical Software},
+  volume = {Vol 3},
+  publisher = {University Library Heidelberg},
+  year = {2015},
+}
+
+@article{Clark2026,
+  title = {Advanced Coupled Electrochemical-Thermal Simulations of 4680 Cylindrical Cells on a Pseudo-Four-Dimensional Grid},
+  volume = {173},
+  ISSN = {1945-7111},
+  url = {http://dx.doi.org/10.1149/1945-7111/ae4896},
+  DOI = {10.1149/1945-7111/ae4896},
+  number = {7},
+  journal = {Journal of The Electrochemical Society},
+  publisher = {The Electrochemical Society},
+  author = {Clark,  Simon and Raynaud,  Xavier and Møll Nilsen,  Halvor and Johansson,  August},
+  year = {2026},
+  month = apr,
+  pages = {070503}
+}
