### 2026-04-17 9300c3b
better ai statement
