+++ Documentation/JOSS/paper.bib
@@ -218,3 +218,13 @@ year={2025},
   month = apr,
   pages = {070503}
 }
+
+@misc{Schmitt2026,
+  doi = {10.48550/ARXIV.2601.10507},
+   author = {Schmitt,  Christina and Johansson,  August and Raynaud,  Xavier and Cedeño,  Eibar Joel Flores and Mugisa,  John and Sotta,  Dane and Martin,  Agathe and Schaeffer,  Nicolas and Debruyne,  Cédric and Reynier,  Yvan and Clark,  Simon and Kopljar,  Dennis},
+  keywords = {Materials Science (cond-mat.mtrl-sci),  Chemical Physics (physics.chem-ph),  FOS: Physical sciences,  FOS: Physical sciences},
+  title = {Comprehensive parameter and electrochemical dataset for a 1 Ah graphite/LNMO battery cell for physical modelling as a blueprint for data reporting in battery research},
+  publisher = {arXiv},
+  year = {2026},
+  copyright = {Creative Commons Attribution 4.0 International}
+}
