+++ Battery/EquivalentCircuitModel/exampleECM.m
@@ -7,6 +7,5 @@ cgit = model.cgit;
 clear state;
 state.I = 3;
 
-
 state = model.evalVarName(state, {'dSOCdt'});
 state = model.evalVarName(state, {'U'});
