+++ Battery/EquivalentCircuitModel/EquivalentCircuitModelInputParams.m
@@ -15,7 +15,8 @@ classdef EquivalentCircuitModelInputParams < InputParams
         initSOC
         initOverpotential2
         lowerVoltageCutoff
-        I 
+        I
+        totalTime
         
     end
 
