+++ Battery/EquivalentCircuitModel/exampleECM.m
@@ -1,10 +1,18 @@
+%% Equivalent Circuit Model
+%
+
+%% Load data set
 jsonstruct = parseBattmoJson('exampleECM.json');
 inputparams = EquivalentCircuitModelInputParams(jsonstruct);
+
+
+%% Setup model
 model = EquivalentCircuitModel(inputparams);
 
+%% run Simulation
 [t, U, I] = model.solve();
 
-%%
+%% Plotting
 
 tiledlayout(1, 2);
 nexttile
