### 2026-04-20 c279864
fixed capacity definition
