+++ Battery/EquivalentCircuitModel/exampleECM.m
@@ -1,21 +1,37 @@
-%% Equivalent Circuit Model
-%
+%% Equivalent Circuit Model Simulation
 
-%% Load data set
-jsonstruct = parseBattmoJson('exampleECM.json');
+% Load data set
+jsonstruct = jsondecode(fileread('exampleECM.json'));
 inputparams = EquivalentCircuitModelInputParams(jsonstruct);
 
-
-%% Setup model
+% Setup model
 model = EquivalentCircuitModel(inputparams);
 
-%% run Simulation
-[t, U, I] = model.solve();
+% Run Simulation
+[t, U, I, SOC] = model.solve();
 
-%% Plotting
 
-tiledlayout(1, 2);
+% Plotting
+figure; 
+tiledlayout(3, 1);
+
 nexttile
-plot(t, U)
+plot(t, U, 'LineWidth', 2)
+title('Tension de la Batterie')
+xlabel('Temps (s)')
+ylabel('Tension (V)')
+grid on
+
 nexttile
-plot(t, I)
+plot(t, I, 'LineWidth', 2, 'Color', 'r')
+title('Courant Appliqué')
+xlabel('Temps (s)')
+ylabel('Courant (A)')
+grid on
+
+nexttile
+plot(t, SOC, 'LineWidth', 2, 'Color', 'g')
+title('SOC')
+xlabel('Temps (s)')
+ylabel('SOC')
+grid on
