### 2026-04-20 c06f07a
new funtion create_parameters instead of json file to allow calculs
