+++ Battery/EquivalentCircuitModel/exampleECM.m
@@ -1,31 +1,9 @@
 %% Equivalent Circuit Model Simulation
 
 % Load data set
-jsonstruct = jsondecode(fileread('exampleECM.json'));
 
-R0_tot = [0.0074814,0.00375135,0.00199926,0.000808379,0.000529892;0.0074814,0.00375135,0.00185474,0.000738337,0.00049771;0.0074814,0.00373265,0.00151744,0.000673703,0.000468303;0.0074814,0.00342697,0.00130825,0.000635903,0.000448193;0.0074814,0.00272078,0.00117717,0.000611909,0.000432971;0.0069826,0.00229452,0.00109773,0.00058627,0.000423262;0.00612671,0.00206227,0.00105217,0.000571475,0.000417812;0.00536849,0.00193337,0.00101405,0.000562212,0.000409251;0.00480278,0.00185103,0.000983687,0.000549631,0.000410476;0.00440167,0.00178588,0.000957044,0.000543895,0.000405368;0.00411575,0.00172477,0.000934531,0.000533558,0.0003994;0.00390187,0.00167949,0.000914869,0.000526768,0.000397983;0.00374658,0.00164546,0.000905014,0.000521591,0.000396355;0.0036334,0.00160835,0.000895256,0.000517151,0.000393944;0.00353281,0.00158063,0.0008793,0.000516464,0.000391954;0.00344791,0.0015608,0.000869252,0.00051227,0.000391406;0.00338977,0.00154377,0.000859461,0.000509065,0.000390664;0.00333872,0.00152388,0.000857998,0.000507685,0.00038969;0.00329657,0.00150926,0.000852066,0.000501813,0.000385995;0.00326141,0.00149673,0.000846844,0.000505136,0.000388294;0.00322799,0.00149188,0.00084232,0.00050115,0.000384211;0.00320237,0.00148102,0.000839376,0.00050073,0.000386191;0.00318137,0.00147045,0.000836217,0.000500865,0.000386833;0.00315527,0.00146685,0.000833235,0.000497888,0.000386251;0.00311928,0.00145992,0.000832092,0.000498747,0.000382866;0.00310424,0.00145576,0.000832276,0.000494884,0.000386295;0.00309834,0.00144722,0.000829437,0.000496968,0.000386992;0.00307867,0.00144351,0.000828243,0.000495848,0.000387399;0.00306673,0.00144798,0.000830089,0.000500481,0.000387638;0.00305254,0.00144507,0.000828833,0.000496636,0.000382113;0.00304161,0.00144131,0.000827309,0.000497478,0.000382439;0.0030343,0.00143125,0.00082455,0.000492916,0.000389673;0.00302615,0.00142702,0.000818782,0.000500463,0.00038546;0.00302297,0.0014205,0.000814619,0.000491503,0.0003858;0.00301535,0.00141463,0.000814864,0.000492482,0.000383651;0.00300786,0.00140957,0.000816839,0.000493931,0.000382345;0.00299892,0.00140604,0.000813864,0.000493379,0.000386187;0.00299335,0.00140043,0.000808475,0.000490454,0.000383351;0.0029867,0.00139212,0.000804998,0.00048877,0.000377422;0.00298272,0.00138903,0.000800962,0.000487252,0.000380367;0.00298442,0.00138775,
[…]
