+++ _sources/pynbnotebooks/tutorial_2_changing_control_protocol_live.ipynb.txt
@@ -72,7 +72,7 @@
    "metadata": {},
    "source": [
     "\n",
-    "Here, we can see that the control policy follows the schema for a constant current discharge (CCDischarge). The three main control parameters in this schema are the CRate, lowerCutoffVoltage, and upperCutoffVoltage. Let's first try changing the protocol to discharge at a rate of C/10 to a lower cutoff voltage of 3 V.\n",
+    "Here, we can see that the control policy follows the schema for a constant current discharge (CCDischarge). The three main control parameters in this schema are the DRate, lowerCutoffVoltage, and upperCutoffVoltage. Let's first try changing the protocol to discharge at a rate of C/10 to a lower cutoff voltage of 3 V.\n",
     ""
    ]
   },
@@ -81,7 +81,7 @@
    "execution_count": 3,
    "metadata": {},
    "source": [
-    "jsonstruct.Control.CRate = 0.1;\n",
+    "jsonstruct.Control.DRate = 0.1;\n",
     "jsonstruct.Control.lowerCutoffVoltage = 3.0;"
    ],
    "outputs": []
@@ -131,7 +131,7 @@
    "execution_count": 5,
    "metadata": {},
    "source": [
-    "jsonstruct.TimeStepping.totalTime = (1./jsonstruct.Control.CRate) .* 3600 .* 1.1;"
+    "jsonstruct.TimeStepping.totalTime = (1./jsonstruct.Control.DRate) .* 3600 .* 1.1;"
    ],
    "outputs": []
   },
@@ -185,7 +185,7 @@
     {
      "data": {
       "text/html": [
[…]
