### 2026-04-27 a5786d1
Small change in runOnlyThermalExtended.m
