### 2026-04-29 77864fe
problem with omega loading fixed
