### 2026-04-29 4eb6435
change in evalProp in BaseModel
