+++ Electrolyser/Models/PlatiniumCatalystLayer.m
@@ -98,8 +98,8 @@ classdef PlatiniumCatalystLayer < CatalystLayer
         
         function state =  updateSources(model, state)
 
-        % Reaction in Electrolyte : 2*H2O + 2*e- <<-> H2 + 2(OH-)_elyte
-        % Reaction in Membrane :    2*H2O + 2*e- <<-> H2 + 2(OH-)_inmr
+        % Reaction in Electrolyte : H2O + e- <<-> (1/2)*H2 + (OH-)_elyte
+        % Reaction in Membrane :    H2O + e- <<-> (1/2)*H2 + (OH-)_inmr
         % Here, the direction of the reaction that is indicated by the repeated arrow symbol corresponds to a positive computed reaction rate
 
             F = model.constants.F;
