### 2026-05-04 5bff6e5
fitting with unitboxBFGS, limit: C1<10*C2
