+++ Battery/EquivalentCircuitModel/spectral_analysis/run_fitting_eis.m
@@ -1,15 +1,21 @@
 %% Fitting Script
 
 % initial values [R0, R1, C1, R2, C2]
- params0 = [8.2e-3, 1.5e-3, 1, 1.5e-2, 200];  %  handmade best ones
+% params0 = [8.2e-3, 1.5e-3, 1, 1.5e-2, 200];  %  handmade best ones
 
-% params0 = [6.2261e-3, 0.3e-3, 10000.1264, 0.00097, 2e5];
+params0 = [3e-3, 0.00260, 100, 0.00132, 1];
 
 
 [Z_re_exp, Z_im_exp, omega] = load_experimental_data();
 %lengthZ = length(Z_im_exp);
 
-scales = [1e-3, 1e-3, 1000, 1e-3, 10000];
+% synthetic data
+% opti_params = [0.00341, 0.00260, 5663.9, 0.00132, 0.9];
+% [Z_re_exp, Z_im_exp] = load_nyquist(opti_params, omega);
+
+%%
+
+scales = [10e-2, 10e-3, 15000, 10e-3, 5];
 
 params0_norm = params0 ./ scales; 
 
@@ -17,13 +23,21 @@ lb_true = [1e-5, 1e-5, 1e-5, 1e-5, 1e-5];
 lb_norm = lb_true ./ scales;
 ub_norm = [];
 
-delta = @(p_norm) compute_residuals(p_norm .* scales, Z_re_exp, Z_im_exp, omega);
+f_opt = @(p_norm) optifunc(p_norm, scales, Z_re_exp, Z_im_exp, omega) ;
 
-options = optimoptions('lsqnonlin', 'Display', 'iter', 'MaxFunctionEvaluations', 20000);
+% best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
+params0_norm = params0_norm(:);
 
-best_params_norm = lsqnonlin(delta, params0_norm, lb_norm, ub_norm, options);
+[min_value, best_params_norm, history] = unitBoxBFGS(...
+    params0_norm, ...                             
+    f_opt, ...           
+    'maximize', false, ...              
+    'linIneq', struct('A', [0, 0, -1, 0, 10], 'b', 0), ...      
+    'enforceFeasible', true ...         
+);
 
-best_params = best_params_norm .* scales;
+
+best_params = best_params_norm(:) .* scales(:);
 
 
 fprintf('\n=== PARAMETERS FOUND ===\n');
@@ -48,18 +62,18 @@ parameters.C2 = best_params(5);
 %%
 figure;
 subplot(3,1,1);
-plot(omega, Z_re_exp, 'ro', 'MarkerFaceColor', 'r');
+semilogx(omega, Z_re_exp, 'ro', 'MarkerFaceColor', 'r');
 hold on;
-plot(omega, Z_re_fit, 'bo', 'LineWidth', 2);        
+semilogx(omega, Z_re_fit, 'bo', 'LineWidth', 2);        
 legend('Expérimental', 'Modèle (Fitted)');
 title('Fitting results');
 xlabel('Omega');
 ylabel('Z_{re} '); 
 
 subplot(3,1,2);
-plot(omega, Z_im_exp, 'ro', 'MarkerFaceColor', 'r');
+semilogx(omega, Z_im_exp, 'ro', 'MarkerFaceColor', 'r');
 hold on;
-plot(omega, Z_im_fit, 'bo', 'LineWidth', 2);        
[…]
