+++ Battery/EquivalentCircuitModel/spectral_analysis/unitBoxBFGS.m
@@ -1,0 +1,578 @@
+function [v, u, history] = unitBoxBFGS(u0, f, varargin)
+% Iterative line search optimization using BFGS intended for scaled 
+% problems, i.e., 0<=u<=1 and f~O(1) 
+% 
+% SYNOPSIS:
+% [v, u, history] = unitBoxBFGS(u0, f, ...)
+%
+% PARAMETERS
+% u0    : inital guess nx1 vector with 0<= u0 <= 1 and feasible wrt
+%         any additional constraints 
+% f     : function handle s.t., [v,g] = f(u) returns
+%           v : objective value
+%           g : objective nx1 gradient vector
+% KEYWORD ARGUMENTS:
+%       'maximize'         : Boolean option (false will minimize objective)
+%                            Default: true
+%       'stepInit'         : Initial step (gradient scaling). If not provided 
+%                            (or set to nan), the following scaling will be used:
+%                            stepInit = maxInitialUpdate/max(|initial gradient|)
+%       'maxInitialUpdate' : as described above. Default: 0.05
+%       'history'          : Option for 'warm starting' based on previous
+%                            optimization (must have been run with option 
+%                            'outputHessian' = true)
+%   Stopping criteria options
+%       'gradTol'       : Absolute tollerance of inf-norm of projected gradient. 
+%                         Default: 1e-3
+%       'objChangeTol'  : Absolute objective update tollerance. Default: 5e-4
+%       'maxIt'         : Maximal number of iterations. Default: 25
+%   Line-search options
+%       'lineSearchMaxIt' : Maximal number of line-search iterations. Default: 5
+%       'stepIncreaseTol' : Maximal step increase factor between line
+%                           search iterations. Deafult: 10
+%       'wolfe1',         : Objective improvement condition. Default: 1e-4
+%       'wolfe2',         : Gradient reduction condition. Default: 0.9
+%   Hessian approximation options
+%       'useBFGS'        : If false pure gradient search will be used.
+%                          Deafult: true
+%       'limitedMemory'  : If false, full Hessian approximations will be
+%                          made (in general not recommended), otherwise
+%                          L-BFGS (see LimitedMemoryHessian.m).
+%                          Default: true
+%       'lbfgsNum'       : Number of vector-pairs stored for L-BFGS.
+%                        : Default: 5
+%       'lbfgsStrategy'  : 'static' or 'dynamic' (see LimitedMemoryHessian.m)
+%                          Default: 'dynamic'
+%   Linear constraints options
+%       'linEq'           : Linear equality constraints given as structure with
+%                           fields 'A' and 'b' to enforce A*u=b
+%       'linIneq'         : Linear inequality constraints *in addition to* the default 
+%                           box constraints (0 <= u <= 1). Given as structure with
+%                           fields 'A' and 'b' to enforce A*u<=b
+%       'enforceFeasible' : If true, will attempt to repair problematic 
+%                           constraint handling/violations by projection. A
+%                           NOTE: A non-feasible initial guess, will always be
+%                           attempted to be made feasible.
+%                           Default: false
+%   Plotting and output options
+%       'plotEvolution' : Plot progess of optimization in figure. 
+%                         Default: true
[…]
