+++ Battery/EquivalentCircuitModel/spectral_analysis/run_fitting_eis.m
@@ -43,15 +43,17 @@ params0_norm = params0_norm(:);
 [min_value, best_params_norm, history] = unitBoxBFGS(...
     params0_norm, ...                             
     f_opt, ...           
-    'maximize', false, ...              
-    'linIneq', struct('A', [0, 0, -1, 0, 2*scales(5)/scales(3)], 'b', 0), ...      
-    'enforceFeasible', true ...         
+    'maximize', false, ...                 
+    'enforceFeasible', true, ...     
+    'maxIt', 300, ...               
+    'objChangeTol', 1e-6, ...       
+    'gradTol', 1e-5 ...
 );
 
 
 best_params = best_params_norm(:) .* scales(:);
 
-[v_opt, g_opt] = optifunc(best_params, scales, Z_re_exp, Z_im_exp, omega)
+[v_opt, g_opt] = optifunc(best_params_norm, scales, Z_re_exp, Z_im_exp, omega)
 
 fprintf('\n=== FITTING SCORE ===\n');
 fprintf('Error : %e\n', v_opt);
