### 2026-05-05 67075ac
iterations and constraints
