+++ Battery/EquivalentCircuitModel/spectral_analysis/run_fitting_eis.m
@@ -43,7 +43,8 @@ params0_norm = params0_norm(:);
 [min_value, best_params_norm, history] = unitBoxBFGS(...
     params0_norm, ...                             
     f_opt, ...           
-    'maximize', false, ...                 
+    'maximize', false, ...      
+    'linIneq', struct('A', [0, 0, -1, 0, 2*scales(5)/scales(3)], 'b', 0), ...   
     'enforceFeasible', true, ...     
     'maxIt', 300, ...               
     'objChangeTol', 1e-6, ...       
