### 2026-05-05 b887e43
update in the stoichiometry coefficients
