+++ Electrolyser/Models/PlatiniumCatalystLayer.m
@@ -110,12 +110,12 @@ classdef PlatiniumCatalystLayer < CatalystLayer
 
             R = elyteR + inmrR;
             
-            state.activeGasSource = -R/F;
-            state.elyteH2Osource  = 2*elyteR/F;
-            state.inmrH2Osource   = 2*inmrR/F;
-            state.elyteOHsource   = -2*elyteR/F;
-            state.inmrOHsource    = -2*inmrR/F;
-            state.eSource         = -2*R.*vols; % in the term eSource, we multiply by the volumes, see unit ([A])
+            state.activeGasSource = -(1/2)*R/F;
+            state.elyteH2Osource  = elyteR/F;
+            state.inmrH2Osource   = inmrR/F;
+            state.elyteOHsource   = -elyteR/F;
+            state.inmrOHsource    = -inmrR/F;
+            state.eSource         = -R.*vols; % in the term eSource, we multiply by the volumes, see unit ([A])
             
         end
         
