### 2026-05-06 2a4c7bd
Experiment (disabled) use custom lower/upper cutoffs
