### 2026-05-06 6b3d582
classe fitting_eis
