+++ Battery/EquivalentCircuitModel/spectral_analysis/fitting_eis.m
@@ -1,0 +1,159 @@
+classdef fitting_eis 
+
+    properties
+       params0
+       Z_re_exp
+       Z_im_exp
+       omega
+       scales
+       lb_true 
+    end
+
+
+
+    methods        
+
+
+        function [min_value, best_params, history] = optimizationBFGS(params0, scales, Z_re_exp, Z_im_exp, omega)
+                
+            f_opt = @(p_norm) optifunc(p_norm, scales, Z_re_exp, Z_im_exp, omega) ;
+            
+            params0_norm = params0 ./ scales; 
+            % best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
+            params0_norm = params0_norm(:);
+            
+            [min_value, best_params_norm, history] = unitBoxBFGS(...
+                params0_norm, ...                             
+                f_opt, ...           
+                'maximize', false, ...      
+                'linIneq', struct('A', [0, 0, -1, 0, 2*scales(5)/scales(3)], 'b', 0), ...   
+                'enforceFeasible', true, ...     
+                'maxIt', 300, ...               
+                'objChangeTol', 1e-6, ...       
+                'gradTol', 1e-5 ...
+            );
+            best_params = best_params_norm(:) .* scales(:);
+
+        end
+
+        function plotresults(params0, scales, Z_re_exp, Z_im_exp, omega)
+
+            [min_value, best_params, history] = optimizationBFGS(params0, scales, Z_re_exp, Z_im_exp, omega);
+            parameters.R0 = best_params(1);
+            parameters.R1 = best_params(2);
+            parameters.C1 = best_params(3);
+            parameters.R2 = best_params(4);
+            parameters.C2 = best_params(5);
+            
+            [Z_re_fit, Z_im_fit] = load_nyquist(best_params, omega);  
+            
+            figure;
+            subplot(3,1,1);
+            semilogx(omega, Z_re_exp, 'ro', 'MarkerFaceColor', 'r');
+            hold on;
+            semilogx(omega, Z_re_fit, 'bo', 'LineWidth', 2);        
+            % legend('Expérimental', 'Modèle (Fitted)');
+            title('Fitting results');
+            xlabel('Omega');
+            ylabel('Z_{re} '); 
+            
+            subplot(3,1,2);
+            semilogx(omega, Z_im_exp, 'ro', 'MarkerFaceColor', 'r');
+            hold on;
+            semilogx(omega, Z_im_fit, 'bo', 'LineWidth', 2);        
+            % legend('Expérimental', 'Modèle (Fitted)');
+            title('Fitting results');
+            xlabel('Omega');
+            ylabel('-Z_{im} '); 
+            
+            subplot(3,1,3);
+            plot(Z_re_exp, Z_im_exp, 'ro', 'MarkerFaceColor', 'r');
+            hold on;
+            plot(Z_re_exp, Z_im_fit, 'bo', 'LineWidth', 2);        
+            % legend('Expérimental', 'Modèle (Fitted)');
+            title('Nyquist');
+            xlabel('Z_{re}');
+            ylabel('-Z_{im} '); 
+            axis equal;
+            
+            grid on;
+            
+            text_error = sprintf('Fitting error : %.2e', v_opt);
+            
+            subplot(3,1,3);
+            
[…]
