### 2026-05-06 d1974a3
added pressure term in potential
