+++ Electrolyser/Models/CatalystLayer.m
@@ -12,6 +12,8 @@ classdef CatalystLayer < BaseModel
         referenceExchangeCurrentDensity % Exchange current density
         standardEquilibriumPotential
         referencePotential
+        referencePressure % reference pressure 
+        
         E0eff
         species % species struct with field
         % - OH.chargeNumber           : Charge number
@@ -38,8 +40,9 @@ classdef CatalystLayer < BaseModel
             
             fdnames = { 'G'                              , ...
                         'referenceExchangeCurrentDensity', ...
-                        'standardEquilibriumPotential'    , ...
+                        'standardEquilibriumPotential'   , ...
                         'referencePotential'             , ...
+                        'referencePressure'              , ...
                         'species'                        , ...
                         'numberOfElectronsTransferred'   , ...
                         'chargeTransferCoefficient'      , ...
