+++ Electrolyser/Models/IridiumCatalystLayer.m
@@ -1,5 +1,4 @@
 classdef IridiumCatalystLayer < CatalystLayer
-
     
     methods
         
@@ -63,15 +62,18 @@ classdef IridiumCatalystLayer < CatalystLayer
             T    = state.T;
             cOH  = state.cOHelyte;
             H2Oa = state.H2OaElyte;
+            p    = state.pressureActiveGas;
             
-            E0  = model.E0eff;
-            c0  = model.species.OH.referenceConcentration;
-            con = model.constants;
-
+            E0   = model.E0eff;
+            c0   = model.species.OH.referenceConcentration;
+            con  = model.constants;
+            con  = model.constants;
+            pref = model.referencePressure;
+            
             F  = con.F;
             R  = con.R;
 
-            state.Eelyte = E0 + R*T/(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2));
+            state.Eelyte = E0 + R*T/(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2)).*(p/pref).^0.5;
 
         end
 
@@ -80,15 +82,17 @@ classdef IridiumCatalystLayer < CatalystLayer
             T    = state.T;
             cOH  = state.cOHinmr;
             H2Oa = state.H2OaInmr;
+            p    = state.pressureActiveGas;
             
-            E0  = model.E0eff;
-            c0  = model.species.OH.referenceConcentration;
-            con = model.constants;
-
+            E0   = model.E0eff;
+            c0   = model.species.OH.referenceConcentration;
+            con  = model.constants;
+            pref = model.referencePressure;
+            
             F  = con.F;
             R  = con.R;
 
-            state.Einmr = E0 + R*T./(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2));
+            state.Einmr = E0 + R*T./(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2)).*(p/pref).^0.5;
             
         end
 
