### 2026-05-06 f0b3932
started with phase field model
