### 2026-05-06 7595336
classe lmoins incohérente?
