+++ Battery/EquivalentCircuitModel/spectral_analysis/fitting_eis.m
@@ -1,4 +1,4 @@
-classdef fitting_eis 
+classdef fitting_eis < handle
 
     properties
        params0
@@ -6,19 +6,31 @@ classdef fitting_eis
        Z_im_exp
        omega
        scales
-       lb_true 
+       
+       best_params_found = [] 
+       fitting_error = []
     end
 
 
 
     methods        
 
+        function feis = fitting_eis(params0, scales, Z_re_exp, Z_im_exp, omega)
+            feis.params0 = params0(:);
+            feis.scales = scales(:);
+            feis.Z_re_exp = Z_re_exp(:);
+            feis.Z_im_exp = Z_im_exp(:);
+            feis.omega = omega(:);
+        end
 
-        function [min_value, best_params, history] = optimizationBFGS(params0, scales, Z_re_exp, Z_im_exp, omega)
-                
-            f_opt = @(p_norm) optifunc(p_norm, scales, Z_re_exp, Z_im_exp, omega) ;
+
+        function [min_value, best_params, history] = optimizationBFGS(feis)
+
             
-            params0_norm = params0 ./ scales; 
+                
+            f_opt = @(p_norm) feis.optifunc(p_norm.* feis.scales) ;
+            
+            params0_norm = feis.params0 ./ feis.scales; 
             % best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
             params0_norm = params0_norm(:);
             
@@ -26,50 +38,50 @@ classdef fitting_eis
                 params0_norm, ...                             
                 f_opt, ...           
                 'maximize', false, ...      
-                'linIneq', struct('A', [0, 0, -1, 0, 2*scales(5)/scales(3)], 'b', 0), ...   
+                'linIneq', struct('A', [0, 0, -1, 0, 2*feis.scales(5)/feis.scales(3)], 'b', 0), ...   
                 'enforceFeasible', true, ...     
                 'maxIt', 300, ...               
                 'objChangeTol', 1e-6, ...       
                 'gradTol', 1e-5 ...
             );
-            best_params = best_params_norm(:) .* scales(:);
+            best_params = best_params_norm(:) .* feis.scales(:);
+
+            feis.best_params_found = best_params;
+            [feis.fitting_error, ~] = feis.optifunc(best_params);
+            
 
         end
 
-        function plotresults(params0, scales, Z_re_exp, Z_im_exp, omega)
-
-            [min_value, best_params, history] = optimizationBFGS(params0, scales, Z_re_exp, Z_im_exp, omega);
-            parameters.R0 = best_params(1);
-            parameters.R1 = best_params(2);
-            parameters.C1 = best_params(3);
-            parameters.R2 = best_params(4);
-            parameters.C2 = best_params(5);
+        function plotresults(feis)
+            if isempty(feis.best_params_found)
+                feis.optimizationBFGS();
+            end
             
-            [Z_re_fit, Z_im_fit] = load_nyquist(best_params, omega);  
+            [Z_re_fit, Z_im_fit] = load_nyquist(feis.best_params_found, feis.omega);  
             
             figure;
             subplot(3,1,1);
-            semilogx(omega, Z_re_exp, 'ro', 'MarkerFaceColor', 'r');
+            semilogx(feis.omega, feis.Z_re_exp, 'ro', 'MarkerFaceColor', 'r');
             hold on;
-            semilogx(omega, Z_re_fit, 'bo', 'LineWidth', 2);        
[…]
