### 2026-05-06 68ae254
some changes runclassfitting
