### 2026-05-06 96f329a
some clean-up in FittingEIS
