+++ Battery/EquivalentCircuitModel/spectral_analysis/FittingEIS.m
@@ -1,14 +1,14 @@
-classdef FittingEIS < handle
+classdef FittingEIS
 
     properties
-       params0
-       Z_re_exp
-       Z_im_exp
-       omega
-       scales
-       
-       best_params_found = [] 
-       fitting_error = []
+        params0
+        Z_re_exp
+        Z_im_exp
+        omega
+        scales
+        
+        % best_params_found = [] 
+        % fitting_error = []
 
     end
 
@@ -23,10 +23,11 @@ classdef FittingEIS < handle
             feis.Z_re_exp = Z_re_exp;
             feis.Z_im_exp = Z_im_exp;
             feis.omega    = omega;
+            
         end
 
 
-        function [min_value, best_params, history] = optimizationBFGS(feis)
+        function [min_value, history, best_params, fitting_error] = optimizationBFGS(feis)
 
             f_opt = @(p_norm) feis.optifunc(p_norm.* feis.scales) ;
             
@@ -43,21 +44,16 @@ classdef FittingEIS < handle
                 'maxIt', 300, ...               
                 'objChangeTol', 1e-6, ...       
                 'gradTol', 1e-5 ...
-            );
+                                                                );
             best_params = best_params_norm(:) .* feis.scales(:);
 
-            feis.best_params_found = best_params;
-            [feis.fitting_error, ~] = feis.optifunc(best_params);
-            
+            fitting_error = feis.optifunc(best_params);
+
         end
 
-        function plotresults(feis)
+        function plotresults(feis, best_params, fitting_error)
 
-            if isempty(feis.best_params_found)
-                feis.optimizationBFGS();
-            end
-            
-            [Z_re_fit, Z_im_fit] = load_nyquist(feis.best_params_found, feis.omega);  
+            [Z_re_fit, Z_im_fit] = load_nyquist(best_params, feis.omega);  
             
             figure;
             subplot(3,1,1);
@@ -90,39 +86,33 @@ classdef FittingEIS < handle
             
             grid on;
             
-            text_error = sprintf('Fitting error : %.2e', feis.fitting_error);
+            text_error = sprintf('Fitting error : %.2e', fitting_error);
             
             subplot(3,1,3);
             
             text(0.05, 0.90, text_error, 'Units', 'normalized', ...
-                'BackgroundColor', 'white', ...   
-                'EdgeColor', 'black', ...        
-                'FontSize', 11, ...               
-                'FontWeight', 'bold');
+                 'BackgroundColor', 'white', ...   
+                 'EdgeColor', 'black', ...        
+                 'FontSize', 11, ...               
+                 'FontWeight', 'bold');
         end
 
-        function GiveResults(feis)
-            if isempty(feis.best_params_found)
-                feis.optimizationBFGS();
-            end
-
+        function printResults(feis, best_params, fitting_error)
 
             fprintf('\n=== FITTING SCORE ===\n');
-            fprintf('Error : %e\n', feis.fitting_error);
+            fprintf('Error : %e\n', fitting_error);
             
             fprintf('\n=== PARAMETERS FOUND ===\n');
-            fprintf('R0 = %.5f Ohms\n', feis.best_params_found(1));
-            fprintf('R1 = %.5f Ohms\n', feis.best_params_found(2));
-            fprintf('C1 = %.1f Farads\n', feis.best_params_found(3));
[…]
