### 2026-05-06 5318a92
class efficient
