+++ Battery/EquivalentCircuitModel/spectral_analysis/load_experimental_data.m
@@ -5,32 +5,51 @@ function [Z_real, Z_imag, omega] = load_experimental_data(filename)
     
     % Le délimiteur est la tabulation ('\t')
     opts = detectImportOptions(filename, 'Delimiter', '\t');
-    data = readtable(filename, opts);
-    
-    % --- 2. Identification des colonnes ---
-    % L'impédance réelle est dans Var26 et l'imaginaire dans Var27.
-    Z_real_raw = data.Var26; 
-    Z_imag_raw = data.Var27;
-    omega_raw = data.Var16;
-    % --- 3. Gestion des virgules ---
-    % On remplace les virgules par des points, puis on convertit le texte en nombres
-    Z_real = str2double(strrep(Z_real_raw, ',', '.'));
-    Z_imag = str2double(strrep(Z_imag_raw, ',', '.'));
-    omega = str2double(strrep(omega_raw, ',', '.'));
 
-    % --- 4. Nettoyage des données ---
-    % S'il y a des lignes de cyclage sans données EIS, on filtre les zéros ou NaN
+    opts.VariableNamesLine = 1;
+    opts.DataLine = 2;
+    opts.VariableNamingRule = 'preserve';
+    data = readtable(filename, opts);
+    colNames = data.Properties.VariableNames;
+    
+
+    idx_real = contains(colNames, 'Re(Z)', 'IgnoreCase', true);
+    idx_imag = contains(colNames, 'Im(Z)', 'IgnoreCase', true);
+    idx_freq = contains(colNames, 'freq', 'IgnoreCase', true);
+    
+    % Extraction des données brutes (peu importe où elles se trouvent)
+    Z_real_raw = data{:, idx_real}; 
+    Z_imag_raw = data{:, idx_imag};
+    freq_raw  = data{:, idx_freq};
+
+    if iscell(Z_real_raw) || isstring(Z_real_raw)
+        Z_real = str2double(strrep(Z_real_raw, ',', '.'));
+        Z_imag = str2double(strrep(Z_imag_raw, ',', '.'));
+        freq  = str2double(strrep(freq_raw, ',', '.'));
+    else
+        Z_real = Z_real_raw;
+        Z_imag = Z_imag_raw;
+        freq  = freq_raw;
+    end
+
     idx_valides = (Z_real ~= 0) & ~isnan(Z_real);
     Z_real = Z_real(idx_valides);
     Z_imag = Z_imag(idx_valides);
-    omega = omega(idx_valides);
+    freq = freq(idx_valides);
 
     Z_real = Z_real(8:60);
     Z_imag = Z_imag(8:60);
-    omega = 2*pi*omega(8:60);
+    omega = 2*pi*freq(8:60);
 
     Z_real = Z_real(:);
     Z_imag = Z_imag(:);
     omega = omega(:);
+%% Figure test
+    % figure;
+    % semilogx(omega, Z_real, 'ro', 'MarkerFaceColor', 'r');      
+    % % legend('Expérimental', 'Modèle (Fitted)');
+    % title('Fitting results');
+    % xlabel('Omega');
+    % ylabel('Z_{re} '); 
     
 end
