### 2026-05-08 cd6eed1
Bump mistune in /Documentation in the pip group across 1 directory
Bumps the pip group with 1 update in the /Documentation directory: [mistune](https://github.com/lepture/mistune).


Updates `mistune` from 3.1.3 to 3.2.1
- [Release notes](https://github.com/lepture/mistune/releases)
- [Changelog](https://github.com/lepture/mistune/blob/main/docs/changes.rst)
- [Commits](https://github.com/lepture/mistune/compare/v3.1.3...v3.2.1)

---
updated-dependencies:
- dependency-name: mistune
  dependency-version: 3.2.1
  dependency-type: direct:production
  dependency-group: pip
...

Signed-off-by: depe
[…]
