+++ Documentation/requirements.txt
@@ -32,7 +32,7 @@ kiwisolver==1.4.8
 latexcodec==3.0.0
 MarkupSafe==3.0.2
 matplotlib==3.10.1
-mistune==3.1.3
+mistune==3.2.1
 mypy-extensions==1.0.0
 nbclient==0.10.2
 nbconvert==7.17.1
