### 2026-05-08 2728853
added utility function setupBatterySimulation
