+++ Utilities/JsonUtils/setupBatterySimulation.m
@@ -1,0 +1,7 @@
+function simsetup = setupBatterySimulation(jsonstruct)
+
+    output = runBatteryJson(jsonstruct, 'runSimulation', false);
+    simsetup = output.simsetup;
+    
+end
+
