### 2026-05-08 cc78c96
send capacity to generic control model in battery setup
