### 2026-05-12 ff7c9a7
Comparison with Ank's experimental data
