### 2026-05-12 7e70a6c
fix pressure term in iridium catalyser
