+++ Electrolyser/Models/IridiumCatalystLayer.m
@@ -73,7 +73,7 @@ classdef IridiumCatalystLayer < CatalystLayer
             F  = con.F;
             R  = con.R;
 
-            state.Eelyte = E0 + R*T/(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2)).*(p/pref).^0.5;
+            state.Eelyte = E0 + R*T/(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2).*(p/pref).^0.5);
 
         end
 
@@ -92,7 +92,7 @@ classdef IridiumCatalystLayer < CatalystLayer
             F  = con.F;
             R  = con.R;
 
-            state.Einmr = E0 + R*T./(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2)).*(p/pref).^0.5;
+            state.Einmr = E0 + R*T./(2*F).*log(H2Oa.*(c0.^2).*(cOH.^-2).*(p/pref).^0.5);
             
         end
 
