### 2026-05-13 d3c3de3
first try to fit an ecm circuit with chen data
