### 2026-05-13 81ff691
try with lsqnonlin, it works
