+++ Battery/EquivalentCircuitModel/spectral_analysis/FittingEIS.m
@@ -48,16 +48,28 @@ classdef FittingEIS
             params0_norm = feis.params0 ./ feis.scales; 
             % best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
             params0_norm = params0_norm(:);
+           
+
+            A_custom = [0, 0, -1, 0, 2*feis.scales(5)/feis.scales(3)];
+            b_custom = 0;
             
+            limite_basse = 1e-5;
+            A_lim = -eye(5); 
+            b_lim = -limite_basse * ones(5, 1);
+            
+            A_total = [A_custom; A_lim];
+            b_total = [b_custom; b_lim];
+
             [min_value, best_params_norm, history] = unitBoxBFGS(...
                 params0_norm, ...                             
                 f_opt, ...           
                 'maximize', false, ...      
-                'linIneq', struct('A', [0, 0, -1, 0, 2*feis.scales(5)/feis.scales(3)], 'b', 0), ...   
+                'linIneq', struct('A', A_total, 'b', b_total), ...   
                 'enforceFeasible', true, ...     
                 'maxIt', 300, ...               
                 'objChangeTol', 1e-6, ...       
-                'gradTol', 1e-5 ...
+                'gradTol', 1e-5, ...
+                'lineSearchMaxIt', 100 ...
                                                                 );
             best_params = best_params_norm(:) .* feis.scales(:);
 
@@ -153,36 +165,120 @@ classdef FittingEIS
             v = sum(err_re.^2 + err_im.^2);
             
             if nargout > 1
-                
+
                 w  = feis.omega(:);
                 R0 = p(1);
                 R1 = p(2);
                 C1 = p(3);
                 R2 = p(4);
                 C2 = p(5);
-                
+
                 g_re_dR0 = -ones(size(w)); 
                 g_re_dR1 = -(1-(R1*C1.*w).^2) ./ (1+(R1*C1.*w).^2).^2;  
                 g_re_dC1 = (2*C1*R1^3.*w.^2) ./ (1+(R1*C1.*w).^2).^2;  
                 g_re_dR2 = -(1-(R2*C2.*w).^2) ./ (1+(R2*C2.*w).^2).^2;  
                 g_re_dC2 = (2*C2*R2^3.*w.^2) ./ (1+(R2*C2.*w).^2).^2;
-                
+
                 J_re = [g_re_dR0, g_re_dR1, g_re_dC1, g_re_dR2, g_re_dC2];
-                
+
                 g_im_dR0 = zeros(size(w));
                 g_im_dR1 = -(2*R1*C1.*w) ./ (1+(R1*C1.*w).^2).^2;  
                 g_im_dC1 = -(R1^2.*w - C1^2*R1^4.*w.^3) ./ (1+(R1*C1.*w).^2).^2;  
                 g_im_dR2 = -(2*R2*C2.*w) ./ (1+(R2*C2.*w).^2).^2;  
                 g_im_dC2 = -(R2^2.*w - C2^2*R2^4.*w.^3) ./ (1+(R2*C2.*w).^2).^2; 
-                
+
                 J_im = [g_im_dR0, g_im_dR1, g_im_dC1, g_im_dR2, g_im_dC2];
                 derr_re_dp = J_re ./ Module_Z; 
                 derr_im_dp = J_im ./ Module_Z;
-                
+
                 g_true = 2 * (derr_re_dp' * err_re) + 2 * (derr_im_dp' * err_im);
                 g_norm = g_true .* feis.scales(:);
                 
             end
+            
+
+            % finite difference method
+            % if nargout > 1
+            %     g_norm = zeros(5,1);
[…]
