### 2026-05-15 7aca7ac
added support to save results to file in SimulationSetup
