+++ Utilities/JsonUtils/runBatteryJson.m
@@ -73,11 +73,12 @@ function  output = runBatteryJson(jsonstruct, varargin)
 
     [model, nls, jsonstruct] = setupNonLinearSolverFromJson(model, jsonstruct);
 
-    simsetupInput = struct('model'          , model    , ...
-                           'initstate'      , initstate, ...
-                           'schedule'       , schedule , ...
-                           'NonLinearSolver', nls);
-
+    jsonstruct = setDefaultJsonStructField(jsonstruct, {'Output', 'saveOutput'}, false);
+    simsetupInput = struct('model'    , model    , ...
+                           'initstate', initstate, ...
+                           'schedule' , schedule , ...
+                           'Output'   , jsonstruct.Output);
+    
     simsetup = SimulationSetup(simsetupInput);
     
     model.verbose = opt.verbose;
