### 2026-05-15 13675e0
small refactoring for output in SimulationSetup
