+++ Documentation/requirements.txt
@@ -4,7 +4,6 @@ attrs==25.3.0
 autopep8==2.3.1
 babel==2.16.0
 beautifulsoup4==4.13.4
-black==26.3.1
 bleach==6.2.0
 certifi==2024.8.30
 charset-normalizer==3.4.0
