+++ Battery/EquivalentCircuitModel/spectral_analysis/runclassfitting.m
@@ -5,18 +5,18 @@
 % 
 % [Z_re_exp, Z_im_exp, omega] = load_experimental_data(filename);
 
-params0 = [3e-2, 1, 1e4, 1e-2, 3e4];
+params0 = [3e-2, 1, 3e3, 1e-1, 1e3];
 
 [Z_re_exp, Z_im_exp, omega] = load_chen_data();
 
-scales = [1e-1, 10, 1e5, 1e-1, 1e5];
+scales = [1e-1, 1e1, 1e5, 1e0, 1e5];
 
-feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega);
+% feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega);
+% 
+% [~, ~, best_params, fitting_error] = feis.optimizationBFGS();
 
-[~, ~, best_params, fitting_error] = feis.optimizationBFGS();
-
-% % Lancement avec LSQNONLIN
-% [min_value, best_params, fitting_error] = feis.optimizationLsqnonlin();
+% Lancement avec LSQNONLIN
+[min_value, best_params, fitting_error] = feis.optimizationLsqnonlin();
 
 
 % result with lsqnonlin:
