### 2026-05-22 a27714c
Bump the pip group across 1 directory with 2 updates
Bumps the pip group with 2 updates in the /Documentation directory: [idna](https://github.com/kjd/idna) and [urllib3](https://github.com/urllib3/urllib3).


Updates `idna` from 3.10 to 3.15
- [Release notes](https://github.com/kjd/idna/releases)
- [Changelog](https://github.com/kjd/idna/blob/master/HISTORY.md)
- [Commits](https://github.com/kjd/idna/compare/v3.10...v3.15)

Updates `urllib3` from 2.6.3 to 2.7.0
- [Release notes](https://github.com/urllib3/urllib3/releases)
- [Changelog](https://github.com/urllib3/urllib3/blob/main/CHANGES.rst
[…]
