+++ Documentation/requirements.txt
@@ -18,7 +18,7 @@ h11==0.16.0
 habanero>=2.0.0
 httpcore==1.0.9
 httpx==0.28.1
-idna==3.10
+idna==3.15
 imagesize==1.4.1
 jedi==0.19.2
 Jinja2==3.1.6
@@ -81,6 +81,6 @@ tornado==6.5.5
 tqdm==4.67.1
 traitlets==5.14.3
 typing_extensions==4.12.2
-urllib3==2.6.3
+urllib3==2.7.0
 webencodings==0.5.1
 yapf==0.43.0
