+++ Documentation/JOSS/paper.bib
@@ -29,7 +29,7 @@ number = {9},
 author = {Ciria Aylagas, Raul and Ganuza, Clara and Parra, Ruben and Yañez, Maria and Ayerbe, Elixabete},
 title = {{cideMOD: An Open Source Tool for Battery Cell Inhomogeneous Performance Understanding}},
 journal = {Journal of The Electrochemical Society},
-doi = {10.1149/1945-7111/ac91fb]}
+doi = {10.1149/1945-7111/ac91fb}
 }
 
 @article{Doyle1993ModelingCell,
