+++ Documentation/JOSS/paper.md
@@ -37,14 +37,14 @@ affiliations:
    index: 1
  - name: SINTEF Industry, Dept. of Sustainable Energy Technology, Norway
    index: 2
-date: 28 November 2025
+date: 22 May 2026
 bibliography: paper.bib
 ---
 <!-- To compile this file, after installing docker, from this directory, run : docker run --rm --volume $PWD:/data --user $(id -u):$(id -g) --env JOURNAL=joss openjournals/inara  -->
 # Summary
 
 This paper presents the Battery Modelling Toolbox (BattMo), a flexible finite volume continuum modelling framework in MATLAB$^®$ [@MATLAB] for
-simulating the performance of electro-chemical cells. BattMo can quickly setup and solve models for a variety of battery
+simulating the performance of electro-chemical cells. BattMo can quickly set up and solve models for a variety of battery
 chemistries, even considering 3D designs such as cylindrical and prismatic cells.
 
 The simulation input parameters, including the material parameters and geometric descriptions, are specified through
@@ -55,7 +55,7 @@ The Doyle-Fuller-Newman (DFN) [@Doyle1993ModelingCell] approach is used as a bas
 thermal simulations. It is possible to include degradation mechanisms such as SEI layer growth, and the use of composite
 material, such as a mixture of Silicon and graphite.
 
-The models are setup in a hierarchical way, for clarity and modularity. Each model corresponds to a computational graph,
+The models are set up in a hierarchical way, for clarity and modularity. Each model corresponds to a computational graph,
 which introduces a set of variables (the nodes) and functional relationship (the edges). This design enables the
 flexibility for changing and designing new models.
 
@@ -69,11 +69,17 @@ New high-performance electro-chemical systems such as Li-ion and post-Li-ion bat
 rigorous digital workflows can help industrial and research institutions reduce the need for physical prototyping
 and derive greater insight and knowledge from their data.
 
-BattMo prolongs this effort by supporting fully 3D geometry and the possibility to easily modify the underlying
-equations. We provide a library of standard parameterized battery geometries. Design optimization can also be done
-on the geometry, which is an essential part of the design.
+BattMo extends this effort by supporting fully 3D geometry and the
+possibility to easily modify the underlying equations. We provide a
+library of standard parameterized battery geometries. Design
+optimization can also be done on the geometry, which is an essential
+part of the design. BattMo is useful for both experienced battery
+designers, wanting to optimize and virtually test different designs,
+research software developers integrating simulation tools in
+electro-chemical system workflows as well as beginners wanting to
+understand fundamental processes.
 
-A challenge for physics-based models for battery and other electro-chemical systems is the difficulty to calibrate the parameters. With an adjoint-based
+A challenge for physics-based models for batteries and other electro-chemical systems is the difficulty to calibrate the parameters. With an adjoint-based
 approach, we can effectively calibrate the models from experiments in a reasonable computational time.
 
 # State of the field
@@ -82,18 +88,18 @@ Recently, a variety of
 open-source battery modelling codes have been released including PyBaMM [@sulzer2021python], cideMOD
 [@CiriaAylagas2022], LIONSIMBA [@torchio2016lionsimba], and PETLion [@Berliner_2021], among others. These
 open-source frameworks help the community reduce the cost of model development and help ensure the
[…]
