### 2026-05-22 c75762b
Use pdf version of fig
