### 2026-05-22 b0a82f8
Bump pygments version
