+++ Documentation/requirements.txt
@@ -49,7 +49,7 @@ platformdirs==4.3.6
 pybtex==0.24.0
 pybtex-docutils==1.0.3
 pycodestyle==2.12.1
-Pygments==2.18.0
+Pygments==2.20.0
 pyparsing==3.2.3
 python-dateutil==2.9.0.post0
 PyYAML==6.0.2
