+++ Electrolyser/Models/ImpedanceElectrolyser.m
@@ -1,0 +1,106 @@
+classdef ImpedanceElectrolyser < Electrolyser
+
+    properties
+
+    end
+
+    methods
+
+        function model = ImpedanceElectrolyser(inputparams)
+
+            model = model@Electrolyser(inputparams);
+
+        end
+
+
+        function model = registerVarAndPropfuncNames(model)
+
+        %% Declaration of the Dynamical Variables and Function of the model
+        % (setup of varnameList and propertyFunctionList)
+
+            model = registerVarAndPropfuncNames@Electrolyser(model);
+
+            model = model.registerVarName('omega');
+            
+            % defines shorthands for the submodels
+            inm = 'IonomerMembrane';
+            her = 'HydrogenEvolutionElectrode';
+            oer = 'OxygenEvolutionElectrode';
+
+            ctl = 'CatalystLayer';
+            exr = 'ExchangeReaction';
+            ptl = 'PorousTransportLayer';
+
+            eldes = {her, oer};
+            layers = {ctl, exr};
+            
+            fn = @() ImpedanceElectrolyser.updateIonomerImpedanceTerm;
+            inputnames = {{inm, 'H2Oceps'}, 'omega'};
+            model = model.registerPropFunction({{inm, 'H2Oaccum'}, fn, inputnames});
+
+            fn = @() ImpedanceElectrolyser.updatePorousTransportLayerimpedanceTerm;
+            % HER
+            ngas = model.(her).(ptl).gasInd.ngas;
+            inputnames = {VarName({her, ptl}, 'compGasMasses', ngas), {her, ptl, 'OHceps'}, 'omega'};
+            model = model.registerPropFunction({VarName({her, ptl}, 'compGasAccums', ngas), fn, inputnames});
+            model = model.registerPropFunction({{her, ptl, 'OHaccum'}, fn, inputnames});
+            % OER
+            ngas = model.(her).(ptl).gasInd.ngas;
+            inputnames = {VarName({oer, ptl}, 'compGasMasses', ngas), {oer, ptl, 'OHceps'}, 'omega'};
+            model = model.registerPropFunction({VarName({oer, ptl}, 'compGasAccums', ngas), fn, inputnames});
+            model = model.registerPropFunction({{oer, ptl, 'OHaccum'}, fn, inputnames});
+            
+            fn = @() ImpedanceElectrolyser.updatePorousTransportLayerLiquidImpedanceTerm;
+            % HER
+            inputnames = {{her, ptl, 'liqrhoeps'}, 'omega'};
+            model = model.registerPropFunction({{her, ptl, 'liquidAccumTerm'}, fn, inputnames});
+            % OER
+            inputnames = {{oer, ptl, 'liqrhoeps'}, 'omega'};
+            model = model.registerPropFunction({{oer, ptl, 'liquidAccumTerm'}, fn, inputnames});
+
+        end
+
+
+        function state = updateLiquideImpedanceTerm(model, state)
+            
+            ne      = 'NegativeElectrode';
+            pe      = 'PositiveElectrode';
+            co      = 'Coating';
+            am      = 'ActiveMaterial';
+            sd      = 'SolidDiffusion';
+
+            op = model.(ne).(co).(am).(sd).operators;
+
+            c     = state.(ne).(co).(am).(sd).c;
+            omega = state.Control.omega;
+            
+            state.(ne).(co).(am).(sd).massAccum = i*omega.*op.vols.*c;
+
+            
+        end
+        
+    end
+
+end
+
+
+
+%{
+Copyright 2021-2024 SINTEF Industry, Sustainable Energy Technology
+and SINTEF Digital, Mathematics & Cybernetics.
+
+This file is part of The Battery Modeling Toolbox BattMo
+
[…]
