### 2026-05-31 83d0e4a
moved method to from ImpedanceBattery to BaseModel
