### 2026-05-31 a6b80ba
wip for impedance in electrolyser
