### 2026-06-01 70a5efe
new boxing working for thevenin model
