+++ Battery/EquivalentCircuitModel/spectral_analysis/FittingEIS.m
@@ -45,7 +45,9 @@ classdef FittingEIS
         end
 
         function p_norm = unscaled2scaled(feis, p)
-            p_norm = (p-feis.params0./feis.scales)./(feis.scales*feis.params0-feis.params0/feis.scales);
+            a = feis.scales;
+            p0 = feis.params0;
+            p_norm = (p-p0./a)./(a*p0-p0/a);
         end
 
         function p = scaled2unscaled(feis, p_norm)
@@ -54,6 +56,7 @@ classdef FittingEIS
             p = (a*p0-p0/a).*p_norm +p0/a;
         end
 
+        %% Thevenin Model
 
         function [min_value, history, best_params, fitting_error] = optimizationBFGS(feis)
 
@@ -63,10 +66,16 @@ classdef FittingEIS
             % best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
             params0_norm = params0_norm(:);
            
+            a = feis.scales;
+            p03 = feis.params0(3);
+            p05 = feis.params0(5);
 
-            A_custom = [0, 0, -1, 0, 2*feis.scales(5)/feis.scales(3)];
-            b_custom = 0;
+            A_custom = [0, 0, -1, 0, 2*(a*p05 - p05/a)/(a*p03- p03/a)];
+            b_custom = (p03-2*p05)/(a^2*p03 - p03);
             
+            max_iter = 300;
+            tol_obj = 1e-7;
+            tol_grad = 1e-5;
 
             [min_value, best_params_norm, history] = unitBoxBFGS(...
                 params0_norm, ...                             
@@ -74,14 +83,35 @@ classdef FittingEIS
                 'maximize', false, ...      
                 'linIneq', struct('A', A_custom, 'b', b_custom),  ...   %A*u<=b
                 'enforceFeasible', true, ...     
-                'maxIt', 300, ...               
-                'objChangeTol', 1e-6, ...       
-                'gradTol', 1e-5, ...
+                'maxIt', max_iter, ...               
+                'objChangeTol', tol_obj, ...       
+                'gradTol', tol_grad, ...
                 'lineSearchMaxIt', 100 ...
                                                                 );
             best_params = scaled2unscaled(feis, best_params_norm(:) );
 
             fitting_error = feis.optifunc(best_params);
+
+            it_count = length(history.val) - 1; 
+            final_pg = history.pg(end);         
+            
+            % Cost function variation
+            if it_count >= 1
+                delta_v = abs(history.val(end) - history.val(end-1));
+            else
+                delta_v = inf;
+            end
+            
+            if it_count >= max_iter
+                fprintf('Stopped because %d iterations reached.\n', max_iter);
+            elseif final_pg < tol_grad
+                fprintf('Stopped because reached gradTol = %e.\n', tol_grad);
+            elseif delta_v < tol_obj
+                fprintf('Stopped because reached objChangeTol = %e.\n', tol_obj);
+            else
+                fprintf('Unexpected error or relative tolerance reached.\n');
+            end           
+
 
         end
 
@@ -136,17 +166,21 @@ classdef FittingEIS
                 J_re = [g_re_dR0, g_re_dR1, g_re_dC1, g_re_dR2, g_re_dC2];
 
                 g_im_dR0 = zeros(size(w));
[…]
