### 2026-06-02 64e0958
scaling changed to pmin and pmax, try to normalize the error (doesn't work..)
