+++ Battery/EquivalentCircuitModel/spectral_analysis/runclassfitting.m
@@ -10,11 +10,15 @@
 
 
 params0 = [2, ...
-             1, ...
-             5e2, ...
-             1, ...
-             1e2];                % initial condition: C1>2*C2
-scales = 100;
+           1, ...
+           7e3, ...
+           1, ...
+           1.5e3];  % initial condition: C1>2*C2
+
+a = 100;
+pmin = params0 / a;
+pmax = params0 * a;
+scales = [pmin, pmax];
 
 params0_w = [1e-1,1e-1,1e-1,1e-1,1e-2,1,1,1e2,1e-07];
 scales_w = params0_w*10;              %to be changed between lsq and bfgs
