### 2026-06-03 daa45b8
fitting robuste pour thevenin
