### 2026-06-03 e9e4a96
fix in doubleLayerCapacitance
