+++ Battery/EquivalentCircuitModel/spectral_analysis/pxdparams_impact/doubleLayerCapacitance_impact.m
@@ -30,13 +30,21 @@ b = logspace(-2, 2, 20);                        %change here
 frequences = logspace(-2, 4, 30); 
 
 
-for i = 1:length(b)
-    
-    jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);        %change here
+for i = 18:20
 
+    fprintf('compute impedance for doubleLayerCapacitance = %.2e ... ', b(i));  %change here
+
+    tic
+    % jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);        %change here
+    % [model, inputparams, ~] = setupModelFromJson(jsonstruct);
+
+    inputparams.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);
     impsolv = ImpedanceSolver(inputparams, 'initstate', initstate, 'computeSteadyState', false);
     
     Z = impsolv.computeImpedance(frequences);
+
+    fprintf('done in %g s\n', toc);
+    
     Z_re = real(Z);
     Z_im = imag(Z);
     curve = sprintf('doubleLayerCapacitance = %.2e', b(i));  %change here
