+++ Battery/EquivalentCircuitModel/spectral_analysis/pxdparams_impact/param_dependency.m
@@ -1,88 +1,115 @@
 function param_dependency()
     mrstModule add ad-core mrst-gui mpfa agmg linearsolvers
-
+    
+    % 1. Chargement des données (Conservez votre méthode de chargement ici)
     jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
     jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
+    jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
     
-    jsonstruct = mergeJsonStructs({jsonstruct_material, ...
-                                   jsonstruct_geometry});
+    c_ne = 29.866*mol/litre; 
+    c_pe = 17.038*mol/litre; 
     
+    % 2. Création de l'interface graphique (Fenêtre plus haute pour accueillir les curseurs)
+    fig = uifigure('Name', 'Impedance Explorer', 'Position', [100, 50, 600, 650]);
     
-    
-    includeDoubleLayer = false;
-    
-    if includeDoubleLayer
-    
-        jsonstruct.(ne).(co).(am).(itf).useDoubleLayerCapacity = true;
-        jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = 0.2;
-    
-    end
-    
-    
-    [~, inputparams, ~] = setupModelFromJson(jsonstruct);
-    
-    c_ne = 29.866*mol/litre; % initial concentration at negative electrode
-    c_pe = 17.038*mol/litre; % initial concentration at positive electrode
-    
-    % 2. Création de l'interface graphique
-    fig = uifigure('Name', 'Impedance explorer', 'Position', [100, 100, 600, 450]);
-    ax = uiaxes(fig, 'Position', [50, 150, 500, 250]);
-    title(ax, 'Diagramme de Nyquist (Chen 2020)');
-    xlabel(ax, 'Z'' (\Omega)');
-    ylabel(ax, '-Z" (\Omega)');
+    % Graphique
+    ax = uiaxes(fig, 'Position', [50, 350, 500, 250]);
+    title(ax, 'Nyquist Diagram');
+    xlabel(ax, 'Re(Z)');
+    ylabel(ax, '-Im(Z)');
     grid(ax, 'on');
+    
+    % Label de statut
+    lbl = uilabel(fig, 'Position', [50, 310, 500, 22], 'Text', 'Ready, use slider to start.');
+    lbl.FontWeight = 'bold';
 
-    % 3. Création du curseur (-1 à +1 ordre de grandeur)
-    sld = uislider(fig, ...
-        'Position', [100, 80, 400, 3], ...
-        'Limits', [-1, 1], ... 
+    % --- CURSEUR 1 : Diffusion Solide (Anode) ---
+    % Modifie le transport du Lithium dans les particules de graphite.
+    uilabel(fig, 'Position', [50, 260, 500, 40], 'WordWrap', 'on', ...
+        'Text', '1. refDiffusionCoefficient * 10^x');
+
+    sld1 = uislider(fig, 'Position', [50, 240, 500, 3], ...
+        'Limits', [-10, 10], ...           % Définit le minimum et le maximum
+        'MajorTicks', -10:1:10, ...        % Ajoute une graduation tous les 1
         'Value', 0);
     
-    lbl = uilabel(fig, 'Position', [100, 100, 400, 22], 'Text', 'Prêt. Bougez le curseur pour lancer un calcul.');
-
-    % 4. Connexion au relâchement de la souris (ValueChangedFcn)
-    % On passe la structure inputparams complète à notre fonction
[…]
