### 2026-06-16 19b7a00
Recet changes and notebook added
