+++ Battery/EquivalentCircuitModel/spectral_analysis/load_chen_data.m
@@ -23,12 +23,12 @@ function [Z_real, Z_imag, omegas] = load_chen_data()
 
 
     
-    includeDoubleLayer = false;
+    includeDoubleLayer = true;
     
     if includeDoubleLayer
     
         jsonstruct.(ne).(co).(am).(itf).useDoubleLayerCapacity = true;
-        jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = 0.2;
+        jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = 30;
     
     end
     
