+++ Battery/EquivalentCircuitModel/spectral_analysis/pxdparams_impact/param_dependency.m
@@ -5,7 +5,22 @@ function param_dependency()
     jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
     jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
     jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
+    ne = 'NegativeElectrode';
+    co = 'Coating';
+    am = 'ActiveMaterial';
+    itf = 'Interface';
     
+    includeDoubleLayer = true;
+    
+    if includeDoubleLayer
+    
+        jsonstruct.(ne).(co).(am).(itf).useDoubleLayerCapacity = true;
+        jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = 0.2;
+    
+    end
+    
+  
+
     c_ne = 29.866*mol/litre; 
     c_pe = 17.038*mol/litre; 
     
@@ -26,29 +41,29 @@ function param_dependency()
     % --- CURSEUR 1 : Diffusion Solide (Anode) ---
     % Modifie le transport du Lithium dans les particules de graphite.
     uilabel(fig, 'Position', [50, 260, 500, 40], 'WordWrap', 'on', ...
-        'Text', '1. refDiffusionCoefficient * 10^x');
+        'Text', '1. Anode refDiffusionCoefficient * 10^x (initial value: 1.3135e-15)');
 
     sld1 = uislider(fig, 'Position', [50, 240, 500, 3], ...
-        'Limits', [-10, 10], ...           % Définit le minimum et le maximum
-        'MajorTicks', -10:1:10, ...        % Ajoute une graduation tous les 1
+        'Limits', [-5, 5], ...           % Définit le minimum et le maximum
+        'MajorTicks', -5:0.5:5, ...        % Ajoute une graduation tous les 1
         'Value', 0);
     
     % --- CURSEUR 2 : Taux de réaction (Anode) ---
     % Modifie la cinétique de transfert de charge.
     uilabel(fig, 'Position', [50, 170, 500, 40], 'WordWrap', 'on', ...
-        'Text', '2. Anode reactionRateConstant * 10^x');
+        'Text', '2. Anode reactionRateConstant * 10^x   (initial value: 5.031e-11) ');
     sld2 = uislider(fig, 'Position', [50, 150, 500, 3],...
-        'Limits', [-10, 10], ...           % Définit le minimum et le maximum
-        'MajorTicks', -10:1:10, ...        % Ajoute une graduation tous les 1
+        'Limits', [-5, 5], ...           % Définit le minimum et le maximum
+        'MajorTicks', -5:0.5:5, ...        % Ajoute une graduation tous les 1
         'Value', 0);
     
     % --- CURSEUR 3 : Taux de réaction (Cathode) ---
     % Modifie la cinétique de transfert de charge.
     uilabel(fig, 'Position', [50, 80, 500, 40], 'WordWrap', 'on', ...
-        'Text', '3. Cathode reactionRateConstant * 10^x');
+        'Text', '3. doubleLayerCapacitance * x (initial value: 0.2)');
     sld3 = uislider(fig, 'Position', [50, 60, 500, 3], ...
-        'Limits', [-10, 10], ...           % Définit le minimum et le maximum
-        'MajorTicks', -10:1:10, ...        % Ajoute une graduation tous les 1
+        'Limits', [-5, 5], ...           % Définit le minimum et le maximum
+        'MajorTicks', -5:0.5:5, ...        % Ajoute une graduation tous les 1
[…]
