+++ Battery/EquivalentCircuitModel/spectral_analysis/pxdparams_impact/doubleLayerCapacitance_impact.m
@@ -6,7 +6,10 @@ jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geo
 jsonstruct = mergeJsonStructs({jsonstruct_material, ...
                                jsonstruct_geometry});
 
-
+ne = 'NegativeElectrode';
+co = 'Coating';
+am = 'ActiveMaterial';
+itf = 'Interface';
 
 includeDoubleLayer = true;
 
@@ -29,23 +32,18 @@ b = logspace(-2, 2, 20);                        %change here
 
 frequences = logspace(-2, 4, 30); 
 
-<<<<<<< HEAD
+
 figure;
+hold on;
 for i = 1:length(b)
     
-    jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);        %change here
-=======
-
-for i = 18:20
->>>>>>> e9e4a96ab3cdbc2f0ce14a0e473826216d412978
-
     fprintf('compute impedance for doubleLayerCapacitance = %.2e ... ', b(i));  %change here
 
     tic
-    % jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);        %change here
-    % [model, inputparams, ~] = setupModelFromJson(jsonstruct);
+    jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);        %change here
+    [model, inputparams, ~] = setupModelFromJson(jsonstruct);
 
-    inputparams.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);
+    % inputparams.(ne).(co).(am).(itf).doubleLayerCapacitance = b(i);
     impsolv = ImpedanceSolver(inputparams, 'initstate', initstate, 'computeSteadyState', false);
     
     Z = impsolv.computeImpedance(frequences);
