### 2026-06-16 73960d2
codes added to notebook
