### 2026-06-17 f9e6c15
switch to GenericBattery in ImpedanceSolver
