### 2026-06-17 58d4c90
rewrote option setup in ImpedanceSolver
