+++ Examples/Advanced/Impedance/runImpedanceChen.m
@@ -35,7 +35,14 @@ c_pe = 17.038*mol/litre; % initial concentration at positive electrode
 
 initstate = initStateChen2020(model, c_ne, c_pe);
 
-impsolv = ImpedanceSolver(inputparams, 'initstate', initstate, 'computeSteadyState', false);
+options = [];
+options.stateInitialization.initializationSetup = 'given state';
+options.stateInitialization.computeSteadyState  = false;
+
+extrastructs = [];
+extrastructs.initstate = initstate;
+
+impsolv = ImpedanceSolver(inputparams, options, extrastructs);
 
 %%
 
