### 2026-06-17 bde883b
rewrote impedance solver
