+++ Battery/ImpedanceSolver.m
@@ -1,11 +1,11 @@
-classdef ImpedanceSolver
+classdef ImpedanceSolver2 < handle
 
     properties (SetAccess = private)
 
         model
         inputparams
         state
-        
+
         %% options
         options
         extrastructs
@@ -15,14 +15,17 @@ classdef ImpedanceSolver
         DM
         DA
         b
+        
         indUs
+        indI
+        inds
         
     end
 
     methods
 
-        function impsolv = ImpedanceSolver(inputparams, options, extrastructs)
-
+        function impsolv = ImpedanceSolver2(inputparams, options, extrastructs)
+            
             if nargin < 3
                 extrastructs = [];
                 if nargin < 2
@@ -30,7 +33,13 @@ classdef ImpedanceSolver
                 end
             end
 
-            options = setDefaultJsonStructField(options, {'stateInitialization', 'initializationSetup'}, 'soc');  % other value are 'given state'
+            % reference dt used to compute jacobians
+            options = setDefaultJsonStructField(options, {'dt'}, 1e3);
+            
+            % other choice for state initializaztion is 'given state'
+            options = setDefaultJsonStructField(options, {'stateInitialization', 'initializationSetup'}, 'soc');  
+                                                                                                                             
+            % default soc value
             options = setDefaultJsonStructField(options, {'stateInitialization', 'soc'}, 1);
 
             switch getJsonStructField(options, {'stateInitialization', 'initializationSetup'})
@@ -49,16 +58,16 @@ classdef ImpedanceSolver
             impsolv.options = options;
 
             impsolv.extrastructs = extrastructs;
-            
+
             ctrl = 'Control';
 
             impsolv.inputparams = inputparams;
 
-            inputparams.(ctrl) = ImpedanceControlModelInputParams([]);
-            impsolv.model = ImpedanceBattery(inputparams);
+            inputparams.(ctrl) = CCDischargeControlModelInputParams([]);
+            impsolv.model = GenericBattery(inputparams);
 
-            impsolv = impsolv.setupSteadyState();
-            impsolv = impsolv.setupHelpers();
+            impsolv.setupSteadyState();
+            impsolv.setupHelpers();
             
         end
 
@@ -82,18 +91,60 @@ classdef ImpedanceSolver
             
         end
         
-        function impsolv = setupHelpers(impsolv)
+        function setupHelpers(impsolv)
 
-            state = impsolv.state;
             model = impsolv.model;
             
             ctrl = 'Control';
-
-            state.(ctrl).omega = 1;
-            state.(ctrl).I = 0;
             
-            state = model.initStateAD(state);
+            indI = model.getIndexPrimaryVariable({ctrl, 'I'});
 
+            impsolv.indI = indI;
+            
+            inds = true(numel(model.getPrimaryVariableNames), 1);
+            inds(indI) = false;
+            inds = find(inds);
+
+            impsolv.inds = inds;
+            
+            state = model.initStateAD(impsolv.state); % just needed to get AD sample
+            indUs = model.getRangePrimaryVariable(state.(ctrl).E, {ctrl, 'E'});
+
+            impsolv.indUs = indUs;
+
+            dt = impsolv.options.dt;
+
+            jac1 = impsolv.getJacobian(dt);
+            jac2 = impsolv.getJacobian(2*dt);
+            
+            impsolv.DM = 2*dt*(jac1 - jac2);
+            impsolv.DA = 2*jac2 - jac1;
+
+        end
+
+        function jac = getJacobian(impsolv, dt)
+
+            state = impsolv.state;
+            model = impsolv.model;
+            indI  = impsolv.indI;
+            inds  = impsolv.inds;
[…]
