### 2026-06-17 d9ff8ea
removed unused ImpedanceBattery model
