### 2026-06-17 176931a
small fix in ImpedanceSolver
