### 2026-06-17 790f4a5
fixes in impedance refactoring
