### 2026-06-17 30eef69
fix in BaseModel
