+++ Utilities/Various/BaseModel.m
@@ -892,44 +892,6 @@ classdef BaseModel < PhysicalModel
             end
             
         end
-
-        function isequal = compareVarName(varname1, varname2)
-            
-            if isempty(varname1) & isempty(varname2)
-                % Needed in recursion
-                isequal = true;
-                return
-            end
-            
-            if numel(varname1) == numel(varname2)
-
-                svarname1 = varname1{1};
-                svarname2 = varname2{1};
-
-                if ischar(svarname1) & ischar(svarname2)
-                    if strcmp(svarname1, svarname2)
-                        isequal = BaseModel.compareVarName(varname1(2 : end), varname2(2 : end));
-                    else
-                        isequal = false;
-                        return
-                    end
-                elseif isnumeric(svarname1) & isnumeric(svarname2)
-                    if svarname1 == svarname2
-                        isequal = BaseModel.compareVarName(varname1(2 : end), varname2(2 : end));
-                    else
-                        isequal = false;
-                        return
-                    end
-                else
-                    isequal = false;
-                    return
-                end
-            else
-                isequal = false;
-                return
-            end
-        end
-        
         
     end
     
