### 2026-06-17 401a206
use base class for impedance solver
