### 2026-06-17 93a2b43
clean-up in ImpedanceSolver base class
