### 2026-06-17 42d1a4c
more refactoring for ImpedanceSolver
