### 2026-06-17 6b21bc5
removed unused model
