### 2026-06-17 3a2cd8b
started with ElectrolyserImpedanceSolver
