### 2026-06-17 13c1605
update in ImpedanceSolver
