+++ Electrolyser/Models/ElectrolyserImpedanceSolver.m
@@ -4,6 +4,13 @@ classdef ElectrolyserImpedanceSolver < ImpedanceSolver
 
         function impsolv = ElectrolyserImpedanceSolver(inputparams, options, extrastructs)
 
+            if nargin < 3
+                extrastructs = [];
+                if nargin < 2
+                    options = [];
+                end
+            end
+
             impsolv = impsolv@ImpedanceSolver(inputparams, options, extrastructs);
             
             % other choice for state initializaztion is 'given state'
@@ -15,6 +22,38 @@ classdef ElectrolyserImpedanceSolver < ImpedanceSolver
             
         end
 
+        function setupIvalue(impsolv)
+
+            options = impsolv.options;
+            
+            switch getJsonStructField(options, {'stateInitialization', 'initializationSetup'})
+                
+              case 'given current'
+
+                impsolv.I = getJsonStructField(options, {'stateInitialization', 'I'});
+                
+              case 'given state'
+
+                initstate = impsolv.extrastructs.initstate;
+                impsolv.I = impsolv.model.getProp(initstate, impsolv.Ivarname);
+                
+              otherwise
+                
+                error('initializationSetup not recognized');
+                
+            end
+            
+        end
+        
+        function setupModel(impsolv)
+
+            model = Electrolyser(impsolv.inputparams);
+            model = model.setupBcAndInitialState();
+
+            impsolv.model = model;
+            
+        end
+        
         function setupVarNames(impsolv)
             
             oer = 'OxygenEvolutionElectrode';
@@ -29,87 +68,73 @@ classdef ElectrolyserImpedanceSolver < ImpedanceSolver
         function drivingForces = setupDrivingForces(impsolv)
             
             drivingForces = impsolv.model.getValidDrivingForces();
-                      
-        end
-        
-        function state = setupIvalue(impsolv, state)
+            drivingForces.src = @(time) impsolv.I;
             
-            %% TO BE UPDATED
-
-            options = impsolv.options
-
-            if strcmp(getStructField(options, {'stateInitialization', 'initializationSetup'}), 'given current')
-                
-                I = getStructField(options, {'stateInitialization', 'current'});
-
-                assert(isAssigned(I), 'A value for the current must be given');
-                
-                state = model.setProp(state, impsolv.Ivarname, I);
-
-            end
-
         end
 
         function setupSteadyState(impsolv)
 
-            ctrl = 'Control';
+            options = impsolv.options;
 
-            inputparams = impsolv.inputparams;
-            options     = impsolv.options;
-            
-            inputparams.(ctrl) = CCDischargeControlModelInputParams([]);;
-            inputparams.(ctrl).lowerCutoffVoltage = 3; % not used but needed for proper initialization
+            if getJsonStructField(options, {'stateInitialization', 'computeSteadyState'})
 
-            initsetup = getJsonStructField(options, {'stateInitialization', 'initializationSetup'});
-
-            switch initsetup
-
-              case 'soc'
+                model    = impsolv.model;
+                controlI = impsolv.I;
                 
-                inputparams.SOC = options.stateInitialization.soc;
-                model = GenericBattery(inputparams);
-                state = model.setupInitialState();
+                switch getJsonStructField(options, {'stateInitialization', 'initializationSetup'})
+                    
+                  case 'given current'
 
-              case 'given state'
+                    [~, initstate] = model.setupBcAndInitialState();
+                    
+                  case 'given state'
 
-                state = impsolv.extrastructs.initstate;
-                model = GenericBattery(inputparams);
+                    initstate = impsolv.extrastructs.initstate;
+                    
+                  otherwise
+                    
+                    error('initializationSetup not recognized');
+                    
+                end
                 
-              otherwise
-                
-                error('initsetup not found');
-                
[…]
