### 2026-06-17 3639843
added current as primary variable in CatalystLayer
necessary for impedance computations
