### 2026-06-17 63e684d
update in BatteryImpedanceSolver
