### 2026-06-17 f75305a
added function to compute equation name index
