### 2026-06-17 e6a4df1
we add the current equation name in ImpedanceSolver
