+++ Battery/EquivalentCircuitModel/utils/createParametersECM.m
@@ -1,4 +1,4 @@
-function p = create_parameters()
+function p = createParametersECM()
     
     p.nominalcellcapacity = 63.7;
     p.initSOC = 1.0;
@@ -14,7 +14,7 @@ function p = create_parameters()
     p.OCP.functionFormat = 'tabulated';
     p.OCP.argumentList = {'SOC'};
     p.OCP.dataX = [0.00, 0.02, 0.04, 0.06, 0.08, 0.10, 0.12, 0.14, 0.16, 0.18, 0.20, 0.22, 0.24, 0.26, 0.28, 0.30, 0.32, 0.34, 0.36, 0.38, 0.40, 0.42, 0.44, 0.46, 0.48, 0.50, 0.52, 0.54, 0.56, 0.58, 0.60, 0.62, 0.64, 0.66, 0.68, 0.70, 0.72, 0.74, 0.76, 0.78, 0.80, 0.82, 0.84, 0.86, 0.88, 0.90, 0.92, 0.94, 0.96, 0.98, 1.00];
-    p.OCP.dataY = [3.126000, 3.273000, 3.380000, 3.447000, 3.470000, 3.480000, 3.489000, 3.500000, 3.515000, 3.533000, 3.551000, 3.567000, 3.582000, 3.594000, 3.602000, 3.609000, 3.614000, 3.620000, 3.625000, 3.631000, 3.637000, 3.643000, 3.649000, 3.657000, 3.665000, 3.674000, 3.684000, 3.696000, 3.710000, 3.727000, 3.750000, 3.775000, 3.798000, 3.818000, 3.838000, 3.858000, 3.877000, 3.897000, 3.917000, 3.937000, 3.958000, 3.979000, 4.000000, 4.022000, 4.044000, 4.066000, 4.089000, 4.112000, 4.135000, 4.159000, 4.182000]
+    p.OCP.dataY = [3.126000, 3.273000, 3.380000, 3.447000, 3.470000, 3.480000, 3.489000, 3.500000, 3.515000, 3.533000, 3.551000, 3.567000, 3.582000, 3.594000, 3.602000, 3.609000, 3.614000, 3.620000, 3.625000, 3.631000, 3.637000, 3.643000, 3.649000, 3.657000, 3.665000, 3.674000, 3.684000, 3.696000, 3.710000, 3.727000, 3.750000, 3.775000, 3.798000, 3.818000, 3.838000, 3.858000, 3.877000, 3.897000, 3.917000, 3.937000, 3.958000, 3.979000, 4.000000, 4.022000, 4.044000, 4.066000, 4.089000, 4.112000, 4.135000, 4.159000, 4.182000];
     
[…]
