+++ Battery/EquivalentCircuitModel/calibration/runFittingEIS.m
@@ -1,172 +1,0 @@
-%% Fitting Script
-
-% initial values [R0, R1, C1, R2, C2]
-% params0 = [8.2e-3, 1.5e-3, 1, 1.5e-2, 200];  %  handmade best ones
-
-params0 = [3e-3, 0.00260, 100, 0.00132, 1];
-
-[Z_re_exp, Z_im_exp, omega] = load_experimental_data();
-
-% lengthZ = length(Z_im_exp);
-
-% synthetic data
-% opti_params = [0.00341, 0.00260, 5663.9, 0.00132, 0.9];     % params with
-% lsqnonlin
-% 
-% params with unitboxBFGS:
-% R0 = 0.00318 Ohms
-% R1 = 0.00247 Ohms
-% C1 = 10654.4 Farads
-% R2 = 0.00155 Ohms
-% C2 = 0.4 Farads
-% 
-
-% [Z_re_exp, Z_im_exp] = load_nyquist(opti_params, omega);
-
-%%
-
-scales = [10e-2, 10e-3, 15000, 10e-3, 5];
-
-params0_norm = params0 ./ scales; 
-
-lb_true = [1e-5, 1e-5, 1e-5, 1e-5, 1e-5];
-lb_norm = lb_true ./ scales;
-ub_norm = [];
-
-f_opt = @(p_norm) optifunc(p_norm, scales, Z_re_exp, Z_im_exp, omega) ;
-
-% best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
-params0_norm = params0_norm(:);
-
-[min_value, best_params_norm, history] = unitBoxBFGS(...
-    params0_norm, ...                             
-    f_opt, ...           
-    'maximize', false, ...      
-    'linIneq', struct('A', [0, 0, -1, 0, 2*scales(5)/scales(3)], 'b', 0), ...   
-    'enforceFeasible', true, ...     
-    'maxIt', 300, ...               
-    'objChangeTol', 1e-6, ...       
-    'gradTol', 1e-5 ...
-);
-
-
-best_params = best_params_norm(:) .* scales(:);
-
-[v_opt, g_opt] = optifunc(best_params_norm, scales, Z_re_exp, Z_im_exp, omega)
-
-fprintf('\n=== FITTING SCORE ===\n');
-fprintf('Error : %e\n', v_opt);
-
-fprintf('\n=== PARAMETERS FOUND ===\n');
-fprintf('R0 = %.5f Ohms\n', best_params(1));
-fprintf('R1 = %.5f Ohms\n', best_params(2));
-fprintf('C1 = %.1f Farads\n', best_params(3));
-fprintf('R2 = %.5f Ohms\n', best_params(4));
-fprintf('C2 = %.1f Farads\n', best_params(5));
-
-%% Graphique 
-
-
-parameters.R0 = best_params(1);
-parameters.R1 = best_params(2);
-parameters.C1 = best_params(3);
-parameters.R2 = best_params(4);
-parameters.C2 = best_params(5);
-
-[Z_re_fit, Z_im_fit] = load_nyquist(best_params, omega);  
-
-
-%%
-figure;
-subplot(3,1,1);
-semilogx(omega, Z_re_exp, 'ro', 'MarkerFaceColor', 'r');
-hold on;
-semilogx(omega, Z_re_fit, 'bo', 'LineWidth', 2);        
-% legend('Expérimental', 'Modèle (Fitted)');
-title('Fitting results');
[…]
