+++ Battery/EquivalentCircuitModel/calibration/specific_calibrations/testCalibration.m
@@ -12,6 +12,10 @@ scales = [pmin, pmax];
 
 feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega);
 
+set(0, 'defaultlinelinewidth', 3);
+set(0, 'DefaultAxesFontSize', 16);
+set(0, 'defaulttextfontsize', 18);
+
 [~, ~, best_params, fitting_error] = feis.optimizationBFGS();
 [Z_re_fit, Z_im_fit] = load_nyquist(best_params, feis.omega); 
 figure;
