+++ Battery/EquivalentCircuitModel/calibration/specific_calibrations/chenCalibration.m
@@ -12,7 +12,13 @@ feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega);
 [~, ~, best_params, fitting_error] = feis.optimizationBFGS();
 
 [Z_re_fit, Z_im_fit] = load_nyquist(best_params, feis.omega);  
-            
+
+
+set(0, 'defaultlinelinewidth', 3);
+set(0, 'DefaultAxesFontSize', 16);
+set(0, 'defaulttextfontsize', 18);
+
+
 figure;
 
 plot(feis.Z_re_exp, -feis.Z_im_exp, 'r', 'MarkerFaceColor', 'r');
