### 2026-06-26 62bdd45
uselesss function deleted
