### 2026-06-29 8093641
mapEisToEcmTable() added
