+++ Battery/EquivalentCircuitModel/calibration/runClassFitting.m
@@ -1,44 +1,45 @@
+function p = runClassFitting(Z_re_exp, Z_im_exp, omega)
+    %% Main code:
+    
+    % filename = 'C:\Users\Alexandre Fichter\Documents\stage_3A\contenu stage\data_August\ank_data\Supplementary material\02_Electrical_characterization\EIS\131-828_EIS_01_MB_CD8.txt';
+    % filename = '/home/xavier/Matlab/Projects/battmo/Data/131-828_EIS_01_MB_CD8.txt';
+    % 
+    % [Z_re_exp, Z_im_exp, omega] = load_experimental_data(filename);
+    % [Z_re_exp, Z_im_exp, omega] = load_santoni_data();
+    % [Z_re_exp, Z_im_exp, omega] = load_chen_data();
+    
+    % omega = logspace(-4, 2, 50);
+    % params = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];
+    % [Z_re_exp, Z_im_exp] = load_nyquist(params, omega);
+    
+    params0 = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];  % initial condition: C1>2*C2
+    
+    a = 1000;
+    
+    pmin = params0 / a;
+    pmax = params0 * a;
+    scales = [pmin, pmax];
+    
+    params0_w = [1e-1,1e-1,1e-1,1e-1,1e-2,1,1,1e2,1e-07];
+    scales_w = params0_w*10;              %to be changed between lsq and bfgs
+    
+    feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega);
+    
+    [~, ~, best_params, fitting_error] = feis.optimizationBFGS();
+    
+    
+    % [min_value, best_params, fitting_error] = feis.optimizationLsqnonlin();
+    %            %  Lancement avec LSQNONLIN
+    
+    
+    %% Printing results
+    if nargin == 2
+    feis.plotresults_thevenin(best_params, fitting_error);
+    
+    feis.printResults(best_params, fitting_error);          % to be changed btw Warburg and Thevenin
+    end
 
-%% Main code:
-
-% filename = 'C:\Users\Alexandre Fichter\Documents\stage_3A\contenu stage\data_August\ank_data\Supplementary material\02_Electrical_characterization\EIS\131-828_EIS_01_MB_CD8.txt';
-% filename = '/home/xavier/Matlab/Projects/battmo/Data/131-828_EIS_01_MB_CD8.txt';
-% 
-% [Z_re_exp, Z_im_exp, omega] = load_experimental_data(filename);
-% [Z_re_exp, Z_im_exp, omega] = load_santoni_data();
-[Z_re_exp, Z_im_exp, omega] = load_chen_data();
-
-% omega = logspace(-4, 2, 50);
-% params = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];
-% [Z_re_exp, Z_im_exp] = load_nyquist(params, omega);
-
-params0 = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];  % initial condition: C1>2*C2
-
-a = 1000;
-
-pmin = params0 / a;
-pmax = params0 * a;
-scales = [pmin, pmax];
-
-params0_w = [1e-1,1e-1,1e-1,1e-1,1e-2,1,1,1e2,1e-07];
-scales_w = params0_w*10;              %to be changed between lsq and bfgs
-
-feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega);
-
-[~, ~, best_params, fitting_error] = feis.optimizationBFGS();
-
-
-% [min_value, best_params, fitting_error] = feis.optimizationLsqnonlin();
[…]
