### 2026-06-30 573def5
validation program in time domain working
