+++ Battery/EquivalentCircuitModel/examples/validateEcmTime.m
@@ -1,0 +1,152 @@
+function validateEcmTime()
+    %% ====================================================================
+    %% STEP 1 : DATA LOADING AND OCV CATCHING
+    %% ====================================================================
+    
+    if exist('ecm_map_results.mat', 'file')
+        load('ecm_map_results.mat', 'ecm_table');
+    else
+        error('file not found ecm_map_results.mat. Run mapEisToEcmTable first.');
+    end
+    
+    % Cathode (Positive)
+    json_pe_path = fullfile('ParameterData','ParameterSets','Chen2020','chen2020_positive_electrode_interface.json');
+    json_pe = parseBattmoJson(json_pe_path);
+    [fn_ocp_pe, ~] = setupFunction(json_pe.openCircuitPotential);
+    
+    % Anode (Negative)
+    json_ne_path = fullfile('ParameterData','ParameterSets','Chen2020','chen2020_negative_electrode_interface.json');
+    json_ne = parseBattmoJson(json_ne_path);
+    [fn_ocp_ne, ~] = setupFunction(json_ne.openCircuitPotential);
+    
+    % 1.3. Generation of the "Ground Truth": Continuous P2D discharge (BattMo)
+    % We run a constant current discharge WITHOUT any pauses to serve as a temporal reference
+    jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
+    jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
+    jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
+    
+    [model, inputparams] = setupModelFromJson(jsonstruct);
+    state_init = setupInitialState(model);
+    
+    C_rate = 1; 
+    t_sim = 3200; % Duration in seconds, be careful not to reach the computed SOC limit (here <10%)
+    
+    inputparams.Control.controlPolicy = 'CCDischarge';
+    inputparams.Control.DRate = C_rate;
+    model.Control = model.setupControl(inputparams.Control);
+    
+    % Continuous schedule construction
+    dt_step = 10;
+    N_sim_steps = floor(t_sim / dt_step);
+    schedule_p2d = struct();
+    schedule_p2d.step.val = ones(N_sim_steps, 1) * dt_step;
+    schedule_p2d.step.control = ones(N_sim_steps, 1);
+    schedule_p2d.control.src = @(t, varargin) varargin{end};
+    schedule_p2d.control.Control = struct('stopFunction', @(model, state, state0) false);
+    
+    fprintf('Simulation of the P2D physical model (BattMo) in progress...\n');
+    [~, states_p2d, ~] = simulateScheduleAD(state_init, model, schedule_p2d);
+    
+    % Extraction of time vectors from the P2D model
+    N_points = length(states_p2d);
+    time_vec = zeros(N_points, 1);
+    V_p2d = zeros(N_points, 1);
+    I_p2d = zeros(N_points, 1);
+    
+    for k = 1:N_points
+        time_vec(k) = states_p2d{k}.time;
+        V_p2d(k) = states_p2d{k}.Control.E;
+        I_p2d(k) = states_p2d{k}.Control.I;
+    end
+    
+    % Coulomb Counting to obtain SOC
+    total_charge = trapz(time_vec, I_p2d);
+    soc_p2d = 1 - cumtrapz(time_vec, I_p2d) / total_charge;
+    
+    %% ====================================================================
[…]
