+++ Battery/EquivalentCircuitModel/calibration/mapEisToEcmTable.m
@@ -1,16 +1,21 @@
 function ecm_table = mapEisToEcmTable()
     %% 1. Initialisation
-    json_path = 'C:\Users\Alexandre Fichter\Documents\stage_3A\contenu stage\data_August\jp3-params\jp3-opt-1d-full.json';
-    jsonstruct = parseBattmoJson(json_path);
+    jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
+    jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
+    
+    jsonstruct = mergeJsonStructs({jsonstruct_material, ...
+                                   jsonstruct_geometry});
+   
+    
     [model, inputparams] = setupModelFromJson(jsonstruct);
     state_current = setupInitialState(model);
     
-    soc_list = (1.0:-0.05:0.05)';
+    soc_list = (1.0:-0.1:0.1)';
     N_steps = length(soc_list);
     
     results_matrix = zeros(N_steps, 6);
     
-
+    freq = logspace(-4, 2, 50)./(2*pi);
     C_rate = 1;
     t_5percent = 3600 * 0.05 / C_rate;
     
@@ -22,13 +27,30 @@ function ecm_table = mapEisToEcmTable()
         fprintf('==================================================\n');
         
         %% extraction de l'EIS par computeImpedance()
+        options = struct();
+
+        options.stateInitialization.computeSteadyState = false;
+
+        options.stateInitialization.initializationSetup = 'given state'; 
+        extrastructs.initstate = state_current;
         
+        impsolv = ImpedanceSolver(inputparams, options, extrastructs);
+        
+        fprintf('Linearisation...\n');
+        Z = impsolv.computeImpedance(freq);
+        
+
+        Z_re = real(Z)';
+        Z_im = imag(Z)';
+        
+        omega = (2*pi).*freq;
+
         %% extraction des paramètres ECM
         p = runClassFitting(Z_re, Z_im, omega); 
         
         % Stockage dans la matrice de résultats
         results_matrix(i, 1) = current_soc;
-        results_matrix(i, 2:6) = p; % Insère [R0, R1, C1, R2, C2]
+        results_matrix(i, 2:6) = p;
         
         %% Décharge de 5% pour passer au SOC suivant
         if i < N_steps
@@ -37,15 +59,40 @@ function ecm_table = mapEisToEcmTable()
             inputparams.Control.DRate = C_rate;
             jsonstruct.TimeStepping.tmax = t_5percent;
             
-            schedule_pulse = model.setupSchedule(jsonstruct);
-            [~, states_pulse] = simulateBattery(model, state_current, schedule_pulse);
+            model.Control = model.setupControl(inputparams.Control);
             
-            % Micro-relaxation 
-            inputparams.Control.DRate = 0;
-            jsonstruct.TimeStepping.tmax = 60; % 1 minute de repos
-            schedule_rest = model.setupSchedule(jsonstruct);
-            [~, states_rest] = simulateBattery(model, states_pulse{end}, schedule_rest);
+           % CONVERSIONS MANUELLES DU SCHEDULE (Plus de setupScheduleStep !)
+            % On découpe les 180s en 10 étapes de 18s pour aider le solveur à converger
+            N_substeps = 10; 
+            schedule_pulse = struct();
+            schedule_pulse.step.val = ones(N_substeps, 1) * (t_5percent / N_substeps);
+            schedule_pulse.step.control = ones(N_substeps, 1);
+            % Cette fonction anonyme renvoie le courant Imax calculé par BattMo
+            schedule_pulse.control.src = @(t, varargin) varargin{end}; 
             
[…]
