### 2026-07-01 c82a35c
evrything works/ table in right order/ plotEcmSoc added
