+++ Battery/EquivalentCircuitModel/examples/validateEcmTime.m
@@ -1,12 +1,19 @@
-function validateEcmTime()
+function validateEcmTime(file_name)
+    
     %% ====================================================================
     %% STEP 1 : DATA LOADING AND OCV CATCHING
     %% ====================================================================
     
-    if exist('ecm_map_results.mat', 'file')
-        load('ecm_map_results.mat', 'ecm_table');
+    % Set default file name if no argument is provided
+    if nargin < 1
+        file_name = 'ecm_map_results_10.mat'; 
+    end
+
+    % Check if the file exists and load it
+    if exist(file_name, 'file')
+        load(file_name, 'ecm_table');
     else
-        error('file not found ecm_map_results.mat. Run mapEisToEcmTable first.');
+        error(['File not found: ', file_name]);
     end
     
     % Cathode (Positive)
@@ -20,28 +27,34 @@ function validateEcmTime()
     [fn_ocp_ne, ~] = setupFunction(json_ne.openCircuitPotential);
     
     % 1.3. Generation of the "Ground Truth": Continuous P2D discharge (BattMo)
-    % We run a constant current discharge WITHOUT any pauses to serve as a temporal reference
     jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
     jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
     jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
     
     [model, inputparams] = setupModelFromJson(jsonstruct);
-    state_init = setupInitialState(model);
     
-    C_rate = 1; 
-    t_sim = 3200; % Duration in seconds, be careful not to reach the computed SOC limit (here <10%)
+    %% New time-varying current profile setup
+    custom_time    = [0, 100,  600, 800, 1500, 2000]; 
+    custom_current = [0,   5,   -2,   0,    6,    0]; % Current profile in Amperes
     
-    inputparams.Control.controlPolicy = 'CCDischarge';
-    inputparams.Control.DRate = C_rate;
+    t_sim = custom_time(end); 
+    
+    % FIX HERE: Keeping CCDischarge ensures the MRST time-stepper steps correctly
+    inputparams.Control.controlPolicy = 'CCDischarge'; 
     model.Control = model.setupControl(inputparams.Control);
     
-    % Continuous schedule construction
-    dt_step = 10;
+    % Safe initialization using the standard policy branch
+    state_init = setupInitialState(model);
+    
+    % Building the fine-grained time schedule (2-second steps)
+    dt_step = 50; 
     N_sim_steps = floor(t_sim / dt_step);
     schedule_p2d = struct();
     schedule_p2d.step.val = ones(N_sim_steps, 1) * dt_step;
     schedule_p2d.step.control = ones(N_sim_steps, 1);
-    schedule_p2d.control.src = @(t, varargin) varargin{end};
+    
+    % Dynamically feed the current profile into the simulation via interpolation
+    schedule_p2d.control.src = @(t, varargin) interp1(custom_time, custom_current, t, 'previous', 0);
     schedule_p2d.control.Control = struct('stopFunction', @(model, state, state0) false);
     
     fprintf('Simulation of the P2D physical model (BattMo) in progress...\n');
@@ -52,6 +65,14 @@ function validateEcmTime()
[…]
