### 2026-07-02 39c8ad2
added explanation in runImpedanceElectrolyser
