+++ .github/workflows/runGitHubTests.yml
@@ -5,7 +5,7 @@ name: Running runGitHubTests
 on:
   # Triggers the workflow on push or pull request events
   push:
-    branches: [ main, dev ]
+    branches: [ dev ]
   pull_request:
     
 jobs:
