+++ .github/workflows/doc.yml
@@ -2,8 +2,8 @@ name: Deploy Sphinx documentation to Pages
 
 on:
   push:
-    branches: [main] # branch to trigger deployment
-
+    tags:
+      - '*'
 jobs:
   pages:
     runs-on: ubuntu-latest
