+++ .github/workflows/createZipArchive.yml
@@ -2,8 +2,7 @@ name: Releases
 
 on:
   push:
-    tags:
-      - '*'
+    branches: [main]
 
 jobs:
   build:
