+++ .github/workflows/createZipArchive.yml
@@ -2,7 +2,8 @@ name: Releases
 
 on:
   push:
-    branches: [main]
+    tags:
+      - '*'
 
 jobs:
   build:
