+++ .github/workflows/doc.yml
@@ -1,26 +1,0 @@
-name: Deploy Sphinx documentation to Pages
-
-on:
-  push:
-    tags:
-      - '*'
-jobs:
-  pages:
-    runs-on: ubuntu-latest
-    steps:
-    - name: Install Pandoc
-      run: |
-        sudo apt-get update
-        sudo apt-get install -y pandoc      
-    - id: deployment
-      uses: sphinx-notes/pages@v3
-      with:
-        publish: false
-        documentation_path: ./Documentation
-        requirements_path: ./Documentation/requirements.txt
-    - uses: peaceiris/actions-gh-pages@v3
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: ${{ steps.deployment.outputs.artifact }}    
-
-
