### 2026-07-03 dae6734
switch syntax for struct utilities
