+++ Utilities/JsonUtils/equalizeStructFields.m
@@ -1,5 +1,5 @@
-function jsonstruct = equalizeJsonStructFields(jsonstruct, fieldnamelists, varargin)
-% same as equalizeJsonStructField (without 's' at end) but iterate over the fieldnamelists;
+function jsonstruct = equalizeStructFields(jsonstruct, fieldnamelists, varargin)
+% same as equalizeStructField (without 's' at end) but iterate over the fieldnamelists;
 
     fielnamelist1 = fieldnamelists{1};
 
@@ -7,7 +7,7 @@ function jsonstruct = equalizeJsonStructFields(jsonstruct, fieldnamelists, varar
         
         fieldnamlist2 = fieldnamelists{ifl};
 
-        jsonstruct = equalizeJsonStructField(jsonstruct, fielnamelist1, fieldnamlist2, varargin{:});
+        jsonstruct = equalizeStructField(jsonstruct, fielnamelist1, fieldnamlist2, varargin{:});
         
     end
 
