### 2026-07-05 9ef8959
update after struct changes
