+++ Utilities/JsonUtils/FlatStructViewer.m
@@ -1,28 +1,28 @@
-classdef FlatJsonViewer
+classdef FlatStructViewer
 
     properties
 
-        flatjson
+        flatstruct
         columnnames
 
     end
 
     methods
 
-        function fjv  = FlatJsonViewer(flatjson, columnnames)
+        function fsv  = FlatStructViewer(flatstruct, columnnames)
 
-            if isa(flatjson, 'table')
-                columnnames = flatjson.Properties.VariableNames;
-                flatjson = table2cell(flatjson);
+            if isa(flatstruct, 'table')
+                columnnames = flatstruct.Properties.VariableNames;
+                flatstruct = table2cell(flatstruct);
                 skip_column_name_set = true;
             else
                 skip_column_name_set = false;                
             end
             
-            fjv.flatjson = flatjson;
+            fsv.flatstruct = flatstruct;
 
             if (nargin < 2) & ~skip_column_name_set
-                ncol = size(flatjson, 2);
+                ncol = size(flatstruct, 2);
                 % default values for the column
                 columnnames{1} = 'parameter name';
                 if ncol == 2
@@ -36,17 +36,17 @@ classdef FlatJsonViewer
                 end
             end
 
-            fjv.columnnames = columnnames;
+            fsv.columnnames = columnnames;
 
         end
 
-        function T = getTable(fjv)
+        function T = getTable(fsv)
 
-            T = cell2table(fjv.flatjson, 'VariableNames', fjv.columnnames);
+            T = cell2table(fsv.flatstruct, 'VariableNames', fsv.columnnames);
             
         end
 
-        function T = print(fjv, varargin)
+        function T = print(fsv, varargin)
 
             opt = struct('filter'         , []   , ...
                          'filename'       , []   , ...
@@ -56,22 +56,22 @@ classdef FlatJsonViewer
             opt = merge_options(opt, varargin{:});
 
             if ~isempty(opt.filter)
-                fjv = fjv.filter(opt.filter);
+                fsv = fsv.filter(opt.filter);
             end
 
             if opt.trimLongEntries
                 maxlen = 50;
-                for icol = 1 : size(fjv.flatjson, 2)
-                    for irow = 1 : size(fjv.flatjson, 1)
-                        entry = fjv.flatjson{irow, icol};
+                for icol = 1 : size(fsv.flatstruct, 2)
+                    for irow = 1 : size(fsv.flatstruct, 1)
+                        entry = fsv.flatstruct{irow, icol};
                         if ischar(entry) && (numel(entry) > maxlen)
-                            fjv.flatjson{irow, icol} = [entry(1 : floor(maxlen/2)) ' ...' entry(end - floor(maxlen/2) : end)];
+                            fsv.flatstruct{irow, icol} = [entry(1 : floor(maxlen/2)) ' ...' entry(end - floor(maxlen/2) : end)];
                         end
                     end
                 end
             end
             
-            T = cell2table(fjv.flatjson, 'VariableNames', fjv.columnnames);
+            T = cell2table(fsv.flatstruct, 'VariableNames', fsv.columnnames);
 
             if opt.print
                 display(T, 'view');
@@ -84,10 +84,10 @@ classdef FlatJsonViewer
         end
 
 
-        function sortedfjv = sort(fjv, orderdesc)
+        function sortedfsv = sort(fsv, orderdesc)
 
-            columnnames = fjv.columnnames;
-            flatjson    = fjv.flatjson;
+            columnnames = fsv.columnnames;
+            flatstruct    = fsv.flatstruct;
 
             ncol = numel(columnnames);
 
@@ -104,30 +104,30 @@ classdef FlatJsonViewer
 
[…]
