### 2026-07-05 898b798
removed unused files in doc
