+++ Documentation/publishedExamples/runBatteryP2D_source.rst
@@ -1,198 +1,0 @@
-:orphan:
-
-.. _runBatteryP2D_source:
-
-Source code for runBatteryP2D
------------------------------
-
-.. code:: matlab
-
-
-  %% Pseudo-Two-Dimensional (P2D) Lithium-Ion Battery Model
-  % This example demonstrates how to setup a P2D model of a Li-ion battery
-  % and run a simple simulation.
-  
-  % Clear the workspace and close open figures
-  clear
-  close all
-  
-  %% Import the required modules from MRST
-  % load MRST modules
-  
-  mrstModule add ad-core mrst-gui mpfa agmg linearsolvers
-  
-  %% Setup the properties of Li-ion battery materials and cell design
-  % The properties and parameters of the battery cell, including the
-  % architecture and materials, are set using an instance of
-  % :class:`BatteryInputParams <Battery.BatteryInputParams>`. This class is
-  % used to initialize the simulation and it propagates all the parameters
-  % throughout the submodels. The input parameters can be set manually or
-  % provided in json format. All the parameters for the model are stored in
-  % the inputparams object.
-  
-  jsonstruct = parseBattmoJson(fullfile('ParameterData','BatteryCellParameters','LithiumIonBatteryCell','lithium_ion_battery_nmc_graphite.json'));
-  
-  % We define some shorthand names for simplicity.
-  ne      = 'NegativeElectrode';
-  pe      = 'PositiveElectrode';
-  elyte   = 'Electrolyte';
-  thermal = 'ThermalModel';
-  co      = 'Coating';
-  am      = 'ActiveMaterial';
-  itf     = 'Interface';
-  sd      = 'SolidDiffusion';
-  ctrl    = 'Control';
-  cc      = 'CurrentCollector';
-  
-  jsonstruct.use_thermal = false;
-  jsonstruct.include_current_collectors = false;
-  
-  inputparams = BatteryInputParams(jsonstruct);
-  
-  use_cccv = true;
-  if use_cccv
-      inputparams.SOC = 0;
-      cccvstruct = struct( 'controlPolicy'     , 'CCCV'       , ...
-                           'initialControl'    , 'charging', ...
-                           'numberOfCycles'    , 2            , ...
-                           'CRate'             , 1.5          , ...
-                           'DRate'             , 1            , ...
-                           'lowerCutoffVoltage', 3            , ...
-                           'upperCutoffVoltage', 4            , ...
-                           'dIdtLimit'         , 1e-2         , ...
-                           'dEdtLimit'         , 1e-4);
-      cccvinputparams = CcCvControlModelInputParams(cccvstruct);
-      inputparams.Control = cccvinputparams;
-  end
-  
-  
-  %% Setup the geometry and computational grid
-  % Here, we setup the 1D computational grid that will be used for the
-  % simulation. The required discretization parameters are already included
-  % in the class BatteryGeneratorP2D.
-  gen = BatteryGeneratorP2D();
-  
-  % Now, we update the inputparams with the properties of the grid.
-  inputparams = gen.updateBatteryInputParams(inputparams);
-  
-  
-  %%  Initialize the battery model.
-  % The battery model is initialized by sending inputparams to the Battery class
-  % constructor. see :class:`Battery <Battery.Battery>`.
-  model = GenericBattery(inputparams);
-  
-  inspectgraph = false;
-  if inspectgraph
-      cgt = model.computationalGraph;
-      return
-  end
-  
-  
-  %% Setup the schedule
-  %
-  
-  timestep.timeStepDuration = 100;
-  
[…]
