### 2026-07-05 2cb2610
some clean-up in conf.y
