+++ Documentation/unparsed/thermal.rst
@@ -29,7 +29,7 @@ model<3dgeometry>` mainly used for demonstration.
    jsonfilename = fullfile('Examples', 'JsonDataFiles', 'cc_discharge_control.json');
    jsonstruct_control = parseBattmoJson(jsonfilename);
    
-   jsonstruct = mergeJsonStructs({jsonstruct_geometry , ...
+   jsonstruct = mergeStructs({jsonstruct_geometry , ...
                                   jsonstruct_material , ...
                                   jsonstruct_control});
    
