### 2026-07-05 584ee0b
fix after struct utils change in doc files
