### 2026-07-05 b34aaaf
removed warning in conf.py
