+++ Battery/EquivalentCircuitModel/calibration/FittingTime.m
@@ -1,0 +1,288 @@
+classdef FittingTime
+
+    properties
+        params0
+        scales
+        time_vec
+        current_exp
+        voltage_exp
+        ocv_vec
+    end
+
+
+    methods        
+    
+        function ftime = FittingTime(params0, scales, time_vec, current_exp, voltage_exp, ocv_vec)
+           
+
+            if istable(params0), params0 = table2array(params0); end
+            if istable(scales), scales = table2array(scales); end
+            if istable(time_vec), time_vec = table2array(time_vec); end
+            if istable(current_exp), current_exp = table2array(current_exp); end
+            if istable(voltage_exp), voltage_exp = table2array(voltage_exp); end
+            if istable(ocv_vec), ocv_vec = table2array(ocv_vec); end
+            
+            ftime.params0     = double(params0(:));
+            ftime.scales      = double(scales(:));
+            ftime.time_vec    = double(time_vec(:));
+            ftime.current_exp = double(current_exp(:));
+            ftime.voltage_exp = double(voltage_exp(:));
+            ftime.ocv_vec     = double(ocv_vec(:));
+        end
+
+
+        function p_norm = unscaled2scaled(ftime, p)
+            pmin = ftime.scales(1:5);
+            pmax = ftime.scales(6:10);
+            p_norm = (p-pmin)./(pmax-pmin);
+        end
+
+        function p = scaled2unscaled(ftime, p_norm)
+            pmin = ftime.scales(1:5);
+            pmax = ftime.scales(6:10);
+            p = (pmax-pmin).*p_norm +pmin;
+        end
+
+        
+        function [min_value, history, best_params, fitting_error] = optimizationBFGS(ftime)
+            
+
+            f_opt = @(p_norm) ftime.optifunc(scaled2unscaled(ftime, p_norm));
+            
+            params0_norm = unscaled2scaled(ftime, ftime.params0);
+            % best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
+            params0_norm = params0_norm(:);
+           
+            pmin = ftime.scales(1:5);
+            pmax = ftime.scales(6:10);
+            p03min = pmin(3);
+            p03max = pmax(3);
+            p05min = pmin(5);
+            p05max = pmax(5);
+
+            A_custom = [0, 0, -1, 0, 2*(p05max - p05min)/(p03max- p03min)];
+            b_custom = (p03min-2*p05min)/(p03max - p03min);
+            
+            max_iter = 300;
+            tol_obj = 1e-15;
+            tol_grad = 1e-5;
+
+            [min_value, best_params_norm, history] = unitBoxBFGS(...
+                params0_norm, ...                             
+                f_opt, ...           
+                'maximize', false, ...      
+                'linIneq', struct('A', A_custom, 'b', b_custom),  ...   %A*u<=b
+                'enforceFeasible', true, ...     
+                'maxIt', max_iter, ...               
+                'objChangeTol', tol_obj, ...       
+                'gradTol', tol_grad, ...
+                'lineSearchMaxIt', 100 ...
+                                                                );
+            best_params = scaled2unscaled(ftime, best_params_norm(:) );
+
+            % explications of why it stopped
+            fitting_error = ftime.optifunc(best_params);
+
+            it_count = length(history.val) - 1; 
+            final_pg = history.pg(end);         
+            
+            % Cost function variation
+            if it_count >= 1
+                delta_v = abs(history.val(end) - history.val(end-1));
+            else
[…]
