### 2026-07-06 a06a384
update in installation.rst
