+++ _sources/publishedExamples/runGasSupply.rst.txt
@@ -32,7 +32,7 @@ The input for the initial state is loaded from :battmofile:`gas-supply-initializ
   filename = fullfile(battmoDir(), 'ProtonicMembrane', 'jsonfiles', 'gas-supply-initialization.json');
   jsonstruct_initialization = parseBattmoJson(filename);
   
-  jsonstruct = mergeJsonStructs({jsonstruct_material, ...
+  jsonstruct = mergeStructs({jsonstruct_material, ...
                                  jsonstruct_geometry, ...
                                  jsonstruct_initialization});
 
