### 2026-07-06 5661007
Fix zip exit code 18 in createZipArchive workflow
