### 2026-07-06 6c05ebe
time simulation enhanced
