+++ Battery/EquivalentCircuitModel/calibration/runClassFittingTime.m
@@ -1,8 +1,15 @@
 function p = runClassFittingTime(pr)
                                                                 
 %% Main code:
-    time_init = (0:100:1000)';
-    current_init = ones(size(time_init));[time_vec, current_exp, voltage_exp, ocv_vec] = loadP2dVoltage(time_init, current_init);
+    time_init = (0:10:1000)';
+    current_init = ones(size(time_init));
+    
+    current_init(time_init >= 0   & time_init < 200)  = 2.5;  % 1. Décharge (0 à 200s)
+    current_init(time_init >= 200 & time_init < 400)  = -1.5; % 2. Charge (200 à 400s)
+    current_init(time_init >= 400 & time_init < 700)  = 0.0;  % 3. Repos (400 à 700s)
+    current_init(time_init >= 700 & time_init <= 1000) = 3.5;  % 4. Décharge (700 à 1000s)
+
+    [time_vec, current_exp, voltage_exp, ocv_vec] = loadP2dVoltage(time_init, current_init);
 
     params0 = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];  % initial condition: C1>2*C2
     
