### 2026-07-06 87ab356
fix function names after merge
