+++ Electrolyser/Plotting/plotPrimScript.m
@@ -24,19 +24,17 @@ end
 
 close all
 
-fds = {'HydrogenEvolutionElectrode.PorousTransportLayer.H2rhoeps', ...
-       'HydrogenEvolutionElectrode.PorousTransportLayer.liqrhoeps', ...
-       'HydrogenEvolutionElectrode.PorousTransportLayer.liqeps', ...
+fds = {'HydrogenEvolutionElectrode.PorousTransportLayer.H2rhoeps'    , ...
+       'HydrogenEvolutionElectrode.PorousTransportLayer.liqeps'      , ...
        'HydrogenEvolutionElectrode.PorousTransportLayer.H2Ogasrhoeps', ...
-       'HydrogenEvolutionElectrode.PorousTransportLayer.OHceps', ...
-       'HydrogenEvolutionElectrode.PorousTransportLayer.phi', ...
-       'OxygenEvolutionElectrode.PorousTransportLayer.O2rhoeps', ...
-       'OxygenEvolutionElectrode.PorousTransportLayer.liqrhoeps', ...
-       'OxygenEvolutionElectrode.PorousTransportLayer.liqeps', ...
-       'OxygenEvolutionElectrode.PorousTransportLayer.H2Ogasrhoeps', ...
-       'OxygenEvolutionElectrode.PorousTransportLayer.OHceps', ...
-       'OxygenEvolutionElectrode.PorousTransportLayer.phi', ...
-       'IonomerMembrane.H2Oceps', ...
+       'HydrogenEvolutionElectrode.PorousTransportLayer.OHceps'      , ...
+       'HydrogenEvolutionElectrode.PorousTransportLayer.phi'         , ...
+       'OxygenEvolutionElectrode.PorousTransportLayer.O2rhoeps'      , ...
+       'OxygenEvolutionElectrode.PorousTransportLayer.liqeps'        , ...
+       'OxygenEvolutionElectrode.PorousTransportLayer.H2Ogasrhoeps'  , ...
+       'OxygenEvolutionElectrode.PorousTransportLayer.OHceps'        , ...
+       'OxygenEvolutionElectrode.PorousTransportLayer.phi'           , ...
+       'IonomerMembrane.H2Oceps'                                     , ...
        'IonomerMembrane.phi'};
 
 dosave = true;
