+++ Electrolyser/Models/PorousTransportLayer.m
@@ -144,8 +144,6 @@ classdef PorousTransportLayer < ElectronicComponent
             varnames{end + 1} = 'H2Ogasrhoeps';
             % Liquid volume fraction, without unit [-]
             varnames{end + 1} = 'liqeps';
-            % total liquid density (mass of liquid per total volume) in [kg m^-3]
-            varnames{end + 1} = 'liqrhoeps';
             % Phase pressures in [Pa]
             phasePressures = VarName({}, 'phasePressures', nph);
             varnames{end + 1} = phasePressures;
@@ -235,12 +233,14 @@ classdef PorousTransportLayer < ElectronicComponent
             varnames{end + 1} = 'liquidMassCons';
             % Residual for the conservation equation for OH in the liquid phase
             varnames{end + 1} = 'OHMassCons';
-            % Residual for the equation of state of the liquid
-            varnames{end + 1} = 'liquidStateEquation';
 
             model = model.registerVarNames(varnames);
 
-
+            %% Extra variables
+            % The concentration of OH and H2O are not used in the assembly but can be evaluated.
+            inds = setdiff((1 : nliquid)', liquidInd.OH);
+            model = model.setAsExtraVarName(VarName({}, 'concentrations', nliquid, inds));
+            
             fn = @() PorousTransportLayer.updateVolumeFractions;
             inputnames = {'liqeps'};
             model = model.registerPropFunction({volumeFractions, fn, inputnames});
@@ -267,8 +267,8 @@ classdef PorousTransportLayer < ElectronicComponent
 
             % update liquid density
             fn = @() PorousTransportLayer.updateLiquidDensity;
-            inputnames = {'liqrhoeps', ...
-                          VarName({}, 'volumeFractions', nph, phaseInd.liquid)};
+            inputnames = {'T', ...
+                          VarName({}, 'concentrations', nliquid, liquidInd.OH)};
             model = model.registerPropFunction({'liqrho', fn, inputnames});
 
             % update conductivity
@@ -408,13 +408,8 @@ classdef PorousTransportLayer < ElectronicComponent
             fn = @() PorousTransportLayer.updateLiquidAccum;
             functionCallSetupFn = @(propfunction) PropFunction.accumFuncCallSetupFn(propfunction);
             fn = {fn, functionCallSetupFn};
-            inputnames = {'liqrhoeps'};
+            inputnames = {'liqrho', 'liqeps'};
             model = model.registerPropFunction({'liquidAccumTerm', fn, inputnames});
-
-            % Assemble residual of equation of state for the liquid phase
-            fn = @() PorousTransportLayer.setupLiquidStateEquation;
-            inputnames = {concentrations};
-            model = model.registerPropFunction({'liquidStateEquation', fn, inputnames});
 
             fn = @() PorousTransportLayer.updateAccumTerms;
             functionCallSetupFn = @(propfunction) PropFunction.accumFuncCallSetupFn(propfunction);
@@ -624,7 +619,7 @@ classdef PorousTransportLayer < ElectronicComponent
 
             vols   = model.G.getVolumes();
 
-            state.liquidAccumTerm = vols.*(state.liqrhoeps - state0.liqrhoeps)/dt;
+            state.liquidAccumTerm = vols.*(state.liqrho.*state.liqeps - state0.liqrho.*state0.liqeps)/dt;
 
         end
 
@@ -717,15 +712,6 @@ classdef PorousTransportLayer < ElectronicComponent
             j = j + assembleFlux(model, cOH, coef);
 
             state.j = j;
-
-        end
-
-
-        function state = updateLiquidDensity(model, state)
-
[…]
