### 2026-07-06 70abb3b
Get python exec from system
