### 2026-07-07 f555ddf
changed function name to runBattery
