+++ Utilities/JsonUtils/runBattery.m
@@ -1,4 +1,4 @@
-function  output = runBatteryJson(jsonstruct, varargin)
+function  output = runBattery(jsonstruct, varargin)
 
     opt = struct('runSimulation'       , true , ...
                  'includeGridGenerator', false, ...
