### 2026-07-07 62e5496
update in rst files after change in function name
