+++ Documentation/output.rst
@@ -6,7 +6,7 @@ BattMo output
    :maxdepth: 2
    :hidden:
 
-   exploreOutput.nblink
+   nblinks/exploreOutput.ipynb
    staticspec
 
 Simulation output visualization and postprocessing are best illustrated in the :ref:`Tutorials <tutorials:Tutorials>`.
@@ -22,7 +22,7 @@ that do not require any simulation such as mass, capacity or nominal energy.
    .. grid-item-card::
       :padding: 2
       
-      :ref:`Simulation Output Structure<exploreOutput:Simulation Output Structure>`
+      :ref:`Simulation Output Structure<nblinks/exploreOutput:Simulation Output Structure>`
 
    .. grid-item-card::
       :padding: 2
