### 2026-07-07 51f8c72
reorganized examples
