### 2026-07-08 3528b64
fix in tutorials after name change
