+++ _sources/nblinks/runSPMplating.nblink.txt
@@ -1,0 +1,3 @@
+{
+  "path" : "../../Examples/Advanced/Plating/notebooks/runSPMplating.ipynb"
+}
