+++ _sources/runSPMplating.nblink.txt
@@ -1,3 +1,0 @@
-{
-  "path" : "../Examples/Advanced/Plating/notebooks/runSPMplating.ipynb"
-}
