+++ _sources/pynbnotebooks/tutorial_7_a_simple_p4d_model_live.ipynb.txt
@@ -1,417 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# <span style=\"color:rgb(213,80,0)\">Tutorial 7 \\- A Simple P4D Simulation</span>\n",
-    "\n",
-    "## Introduction\n",
-    "\n",
-    "In this tutorial, we will use a simple P4D simulation to explore the effects of battery cell architecture. After completing this tutorial, you should have a working knowledge of:\n",
-    "\n",
-    "- How to setup and run a P4D simulation of a single cell in BattMo\n",
-    "- Advanced usage of combining model descriptions from multiple sources\n",
-    "\n",
-    "## Construct the Model from Different Sources\n",
-    "\n",
-    "Let's say that you parameterize some cell materials and put that data in a JSON file. Then you get some description of your cell geometry, and you put that data in another JSON file. Now you want to combine those descriptions into a single model and simulate it using some pre\\-defined contol protocol and simulation settings in other files. How can we combine those easily, without having to do any recoding?\n",
-    "\n",
-    "To do that, we can simply make use of the [mergeStructs](https://github.com/BattMoTeam/BattMo/blob/main/Utilities/JsonUtils/mergeStructs.m) function in **BattMo** (an other examle is presented  [here](<https://battmoteam.github.io/BattMo/mergejsonstruct.html#merging-parameters>).\n",
-    "\n",
-    "First, let’s define our cell materials. We have provided a JSON file that contains material properties for a NMC and Graphite active materials, which we can parse as a **BattMo** structure:\n",
-    ""
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "metadata": {},
-   "source": [
-    "% parse material definitions as a BattMo structure\n",
-    "jsonfilename = 'ParameterData/BatteryCellParameters/LithiumIonBatteryCell/lithium_ion_battery_nmc_graphite.json';\n",
-    "jsonstruct_material = parseBattmoJson(jsonfilename);"
-   ],
[…]
