+++ _sources/pynbnotebooks/tutorial_9_simulate_a_cylindrical_cell_live.ipynb.txt
@@ -1,393 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# <span style=\"color:rgb(213,80,0)\">Tutorial 9 \\- Simulate a cylindrical cell</span>\n",
-    "\n",
-    "## Introduction\n",
-    "\n",
-    "In this tutorial, we simulate a multilayer pouch cell. We use the same material property as in the other tutorials\n",
-    ""
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "metadata": {},
-   "source": [
-    "jsonstruct_material = parseBattmoJson('Examples/jsondatafiles/sample_input.json');"
-   ],
-   "outputs": []
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "\n",
-    "Next, we load and parse a json file where we have chosen some parameters for the multilayer pouch domain. Note that all the parameters are described in a json schema, see [Geometry.schema.json](https://github.com/BattMoTeam/BattMo/blob/main/Utilities/JsonSchemas/Geometry.schema.json), even if the simplest way to proceed is to start with an example, in this case given by [4680\\-geometry.json](https://github.com/BattMoTeam/BattMo/blob/main/Examples/JsonDataFiles/4680-geometry.json).\n",
-    ""
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "metadata": {},
-   "source": [
-    "jsonfilename = 'Examples/JsonDataFiles/4680-geometry.json';\n",
-    "jsonstruct_geometry = parseBattmoJson(jsonfilename);"
-   ],
-   "outputs": []
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "\n",
-    "We change some parameters to get a smaller model and simulation time.\n",
-    ""
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 3,
-   "metadata": {},
-   "source": [
-    "jsonstruct_geometry.Geometry.outerRadius = jsonstruct_geometry.Geometry.innerRadius + 4*milli*meter;\n",
-    "jsonstruct_geometry.Geometry.numberOfDiscretizationCellsVertical = 2;\n",
[…]
