+++ _sources/exploreOutput.nblink.txt
@@ -1,3 +1,0 @@
-{
-  "path" : "../Examples/Documentation/notebooks/exploreOutput.ipynb"
-}
