+++ _sources/pynbnotebooks/tutorial_1_a_simple_p2d_model_live.ipynb.txt
@@ -1,528 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# <span style=\"color:rgb(213,80,0)\">Tutorial 1 \\- Your First BattMo Model</span>\n",
-    "\n",
-    "## Introduction\n",
-    "\n",
-    "Let’s make your first **BattMo** model! ?⚡?\n",
-    "\n",
-    "In this tutorial, we will build, run, and visualize a simple pseudo\\-two\\-dimensional (P2D) simulation for a Li\\-ion battery cell. We will first introduce how **BattMo** handles model parameters, then run the simulation, dashboard the results, and explore the details of the simulation output.\n",
-    "\n",
-    "## Define Parameters\n",
-    "\n",
-    "**BattMo** uses JSON to manage parameters. This allows you to easily save, document, and share complete parameter sets from specific simulations. We have used long and explicit key names for good readability. If you are new to JSON, you can learn more about it [here](https://www.w3schools.com/js/js_json_intro.asp). Details on the BattMo specification are available in the [JSON input specification](https://battmoteam.github.io/BattMo/json.html).\n",
-    "\n",
-    "For this example, we provide a sample JSON file [sample\\_input.json](https://github.com/BattMoTeam/BattMo/blob/main/Examples/JsonDataFiles/sample_input.json) that describes a simple NMC\\-Graphite cell.\n",
-    "\n",
-    "We load and parse the JSON input file into **BattMo** using the command:\n",
-    ""
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "metadata": {},
-   "source": [
-    "jsonstruct = parseBattmoJson('Examples/JsonDataFiles/sample_input.json');"
-   ],
-   "outputs": []
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "\n",
[…]
