+++ _sources/nblinks/runControlExamples.nblink.txt
@@ -1,0 +1,3 @@
+{
+  "path" : "../../Examples/Basic/notebooks/runControlExamples.ipynb"
+}
