+++ _sources/runImpedanceScript.nblink.txt
@@ -1,3 +1,0 @@
-{
-  "path" : "../Examples/Advanced/Impedance/notebooks/runImpedanceScript.ipynb"
-}
