+++ Battery/EquivalentCircuitModel/calibration/FittingEIS.m
@@ -304,130 +304,7 @@ classdef FittingEIS
 
 
 
-        %% lsq method
-
-        function residuals = optifunc_lsq(feis, p)
-            % 1. On empêche physiquement les paramètres de valoir zéro
-            p = max(p, 1e-8); 
-            
-            % 2. Chargement du modèle
-            [Z_re_param, Z_im_param] = load_nyquist(p, feis.omega);
-            if istable(Z_re_param), Z_re_param = table2array(Z_re_param); end
-            if istable(Z_im_param), Z_im_param = table2array(Z_im_param); end
-            
-            Z_re_param = double(Z_re_param(:));
-            Z_im_param = double(Z_im_param(:));
-            
-            % Si le modèle renvoie du NaN (physiquement impossible), on pénalise
-            if ~isnumeric(Z_re_param) || ~isreal(Z_re_param) || any(isnan(Z_re_param(:)))
-                residuals = 1e10 * ones(2 * length(feis.omega), 1);
-                return;
-            end
-            
-            % 3. Calcul du module pour l'erreur relative
-            Modulus_Z = sqrt(feis.Z_re_exp.^2 + feis.Z_im_exp.^2);
-            Modulus_Z = Modulus_Z(:);
-            
-            % 4. Vecteurs de résidus (Non mis au carré, lsqnonlin s'en charge)
-            err_re = (feis.Z_re_exp(:) - Z_re_param(:)) ./ Modulus_Z;
-            err_im = (feis.Z_im_exp(:) - Z_im_param(:)) ./ Modulus_Z;
-            
-            % 5. On empile les parties réelles et imaginaires en un seul vecteur colonne
-            residuals = [err_re; err_im];
-        end
-
-        function [min_value, best_params, resnorm] = optimizationLsqnonlin(feis)
-            % Fonction objectif : renvoie le vecteur de résidus
-            f_res = @(p_norm) feis.optifunc_lsq(p_norm(:) .* feis.scales(:));
-            
-            % Paramètres initiaux normés
-            params0_norm = feis.params0(:) ./ feis.scales(:); 
-            
-            % Bornes (Limites inférieures strictes > 0 pour éviter les crashs)
-            lb_norm = ones(5,1) * 1e-5; 
-            ub_norm = ones(5,1) * 1e5./ feis.scales(:); 
-            
-            % Options de lsqnonlin
-            options = optimoptions('lsqnonlin', ...
-                'Display', 'iter-detailed', ... % Affichage très détaillé
-                'Algorithm', 'trust-region-reflective', ... % Parfait pour l'EIS
-                'MaxFunctionEvaluations', 3000, ...
-                'MaxIterations', 500, ...
-                'StepTolerance', 1e-8, ...
-                'FunctionTolerance', 1e-6);
-                
-            % Lancement de l'optimiseur natif MATLAB
-            [best_params_norm, resnorm, ~, exitflag, output] = lsqnonlin(...
-                f_res, params0_norm, lb_norm, ub_norm, options);
-                
-            % Récupération des vraies valeurs (dénormalisation)
-            best_params = best_params_norm(:) .* feis.scales(:);
-            min_value = resnorm; % Erreur finale (somme des carrés des résidus)
-            
-            % Affichage du statut de fin
-            disp('=== RAISON DE FIN DE LSQNONLIN ===');
-            disp(output.message);
-        end
-
-%% lsq method warburg
-        function v = optifunclsq_warburg(feis, p) 
-            p_safe = max(p, 1e-15);
-            
[…]
