+++ Battery/EquivalentCircuitModel/calibration/classFitting.m
@@ -1,4 +1,4 @@
-function p = runClassFitting(Z_re_exp, Z_im_exp, omega, print)  %print is a boolean, if one gives it to the function it will plot the results.
+function p = classFitting(Z_re_exp, Z_im_exp, omega, print)  %print is a boolean, if one gives it to the function it will plot the results.
     %% Main code:
     
     % filename = 'C:\Users\Alexandre Fichter\Documents\stage_3A\contenu stage\data_August\ank_data\Supplementary material\02_Electrical_characterization\EIS\131-828_EIS_01_MB_CD8.txt';
