+++ Battery/EquivalentCircuitModel/calibration/runMapEisToEcmTable.m
@@ -1,0 +1,4 @@
+soc_step = 0.5;
+ecm_table = mapEisToEcmTable(soc_step);
+plotEcmSoc(ecm_table);
+
