+++ Battery/EquivalentCircuitModel/calibration/runFittingTime.m
@@ -1,40 +1,45 @@
-function p = classFittingTime(pr)
-                                                                
-%% Main code:
-    time_init = (0:10:1000)';
-    current_init = ones(size(time_init));
-    
-    current_init(time_init > = 0   & time_init < 200)  = 2.5;  % 1. Décharge (0 à 200s)
-    current_init(time_init > = 200 & time_init < 400)  = -1.5; % 2. Charge (200 à 400s)
-    current_init(time_init > = 400 & time_init < 700)  = 0.0;  % 3. Repos (400 à 700s)
-    current_init(time_init > = 700 & time_init <= 1000) = 3.5;  % 4. Décharge (700 à 1000s)
+%% Time fitting
 
-    [time_vec, current_exp, voltage_exp, ocv_vec] = loadP2dVoltage(time_init, current_init);
 
-    params0 = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];  % initial condition: C1>2*C2
-    
-    a = 1000;
-    
-    pmin = params0 / a;
-    pmax = params0 * a;
-    scales = [pmin, pmax];
-        
-    ftime = FittingTime(params0, scales, time_vec, current_exp, voltage_exp, ocv_vec);
-    
-    [~, ~, best_params, fitting_error] = ftime.optimizationBFGS();
-    
-    p = best_params;
-    
-    
-    %% Printing results
-    if nargin == 4 || nargin == 1
-        
-        ftime.plotresults_thevenin(best_params, fitting_error);
-        
-        ftime.printResults(best_params, fitting_error);          
-        
-        ftime.plottest(best_params)
-        
-    end
+%% Setup current input
+time_init = (0:10:1000)';
+current_init = ones(size(time_init));
 
+current_init(time_init >= 0   & time_init < 200)  = 2.5;  % 1. Décharge (0 à 200s)
+current_init(time_init >= 200 & time_init < 400)  = -1.5; % 2. Charge (200 à 400s)
+current_init(time_init >= 400 & time_init < 700)  = 0.0;  % 3. Repos (400 à 700s)
+current_init(time_init >= 700 & time_init <= 1000) = 3.5;  % 4. Décharge (700 à 1000s)
+
+jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
+jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
+jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
+
+[time_vec, current_exp, voltage_exp, ocv_vec] = loadP2dVoltage(jsonstruct, time_init, current_init);
+
+params0 = [0.05052, 1.12673, 59119.9, 0.03155, 11054.0];  % initial condition: C1>2*C2
+
+a = 1000;
+
+pmin = params0 / a;
+pmax = params0 * a;
+scales = [pmin, pmax];
+
+ftime = FittingTime(params0, scales, time_vec, current_exp, voltage_exp, ocv_vec);
+
+[~, ~, best_params, fitting_error] = ftime.optimizationBFGS();
+
+p = best_params;
+
+
+doprint = true;
+%% Printing results
+if doprint
+    
+    ftime.plotresults_thevenin(best_params, fitting_error);
+    
+    ftime.printResults(best_params, fitting_error);          
+    
+    ftime.plottest(best_params)
+    
 end
+
