### 2026-07-13 8c13e37
update in EquivalentCircuitModel model
