+++ Battery/EquivalentCircuitModel/calibration/FittingTime.m
@@ -115,11 +115,102 @@ classdef FittingTime
 
             % Preventing physical values to get too close from 0
             p_safe = max(p, 1e-8);
+            
+            R0 = p_safe(1);
+            R1 = p_safe(2);
+            C1 = p_safe(3);
+            R2 = p_safe(4);
+            C2 = p_safe(5);
 
-            output = runECM(p_safe, input, options)
 
-        end 
-    
+            N_points = length(ftime.time_vec);
+            
+            V_ecm = zeros(N_points, 1);
+
+            Uc1 = 0; 
+            Uc2 = 0;
+            
+            % Sensibilities initialisation 
+
+            % we do not need s_R0 as it is trivial
+            s_R1 = 0; % dUc1 / dR1
+            s_C1 = 0; % dUc1 / dC1
+            s_R2 = 0; % dUc2 / dR2
+            s_C2 = 0; % dUc2 / dC2
+            
+            g_true = zeros(5, 1);
+            
+            for k = 1:N_points
+                
+                if k == 1
+                    dt = ftime.time_vec(1);
+                else
+                    dt = ftime.time_vec(k) - ftime.time_vec(k-1);
+                end
+                
+                I = ftime.current_exp(k);           
+
+                V_ocv = ftime.ocv_vec(k); 
+                
+                % 1. Tension du modèle au pas k
+                V_ecm(k) = V_ocv - R0 * I - Uc1 - Uc2;
+                
+                % 2. Erreur instantanée
+                err = ftime.voltage_exp(k) - V_ecm(k);
+
+                if options.computeDerivatives
+                    
+                    % Dérivées partielles de V_ecm par rapport à chaque paramètre
+                    dV_dR0 = -I;
+                    dV_dR1 = -s_R1;
+                    dV_dC1 = -s_C1;
+                    dV_dR2 = -s_R2;
+                    dV_dC2 = -s_C2;
+                    
+                    % Accumulation du gradient de la fonction coût (v = sum(err^2))
+                    g_true(1) = g_true(1) - 2 * err * dV_dR0;
+                    g_true(2) = g_true(2) - 2 * err * dV_dR1;
+                    g_true(3) = g_true(3) - 2 * err * dV_dC1;
+                    g_true(4) = g_true(4) - 2 * err * dV_dR2;
+                    g_true(5) = g_true(5) - 2 * err * dV_dC2;
+                    
+                    % Mise à jour des sensibilités pour le pas suivant (Euler Explicite)
+                    s_R1 = s_R1 + (-s_R1 / (R1 * C1) + Uc1 / (R1^2 * C1)) * dt;
+                    s_C1 = s_C1 + (-s_C1 / (R1 * C1) + Uc1 / (R1 * C1^2) - I / C1^2) * dt;
+                    s_R2 = s_R2 + (-s_R2 / (R2 * C2) + Uc2 / (R2^2 * C2)) * dt;
+                    s_C2 = s_C2 + (-s_C2 / (R2 * C2) + Uc2 / (R2 * C2^2) - I / C2^2) * dt;
+                    
+                end
+                
+                % 4. Mise à jour des variables d'état électriques pour le pas suivant
+                Uc1 = Uc1 + (-Uc1 / (R1 * C1) + I / C1) * dt;
+                Uc2 = Uc2 + (-Uc2 / (R2 * C2) + I / C2) * dt;
+                
+            end
+            
+            % Valeur de la fonction objectif (Erreur quadratique totale)
+            v = sum((ftime.voltage_exp - V_ecm).^2);
+            
+
+            if isnan(v) || isinf(v) || (nargout > 1 && (any(isnan(g_true)) || any(isinf(g_true))))
+                v = 1e10; % Assigne un coût massif pour rejeter le point instable
+                if nargout > 1
[…]
