+++ Battery/EquivalentCircuitModel/calibration/FittingTime.m
@@ -5,20 +5,25 @@ classdef FittingTime
         params0     % initial guess for the parameters to be fitted
         scales      % scaling for the unit box optimization
         
-        time_vec     % time steps for current_exp and voltage_exp
+        time_vec   % time steps for current_exp and voltage_exp
+
         current_exp  % experimental current values
-        voltage_exp  % experimental voltage values
-
-        ocv_vec      % open circuit voltage values computed from a model
+        jsontruct  % input structure for P2D model which gives us the the ocv and the experimental data
         
-    end
+        %% helpers
+        
+        voltage_exp  % experimental voltage values
+        ocv_vec      % open circuit voltage values computed from jsonstruct
 
+    end
 
     methods        
     
-        function ftime = FittingTime(params0, scales, time_vec, current_exp, voltage_exp, ocv_vec)
+        function ftime = FittingTime(jsonstruct, time_vec, current_exp, params0, scales)
            
 
+            ftime.jsonstruct = jsonstruct;
+            
             if istable(params0), params0 = table2array(params0); end
             if istable(scales), scales = table2array(scales); end
             if istable(time_vec), time_vec = table2array(time_vec); end
@@ -30,10 +35,13 @@ classdef FittingTime
             ftime.scales      = double(scales(:));
             ftime.time_vec    = double(time_vec(:));
             ftime.current_exp = double(current_exp(:));
-            ftime.voltage_exp = double(voltage_exp(:));
-            ftime.ocv_vec     = double(ocv_vec(:));
-        end
 
+            [V_p2d, ocv_vec] = ftime.setup(ftime.time_vec, ftime.current_exp);
+
+            ftime.voltage_exp = V_p2d;
+            ftime.ocv_vec     = ocv_vec;
+            
+        end
 
         function p_norm = unscaled2scaled(ftime, p)
             pmin = ftime.scales(1:5);
@@ -47,7 +55,72 @@ classdef FittingTime
             p = (pmax-pmin).*p_norm +pmin;
         end
 
-        
+
+        function [V_p2d, ocv_vec] = setupOCP(ftime, time_vec, current_exp)
+
+            jsonstruct  = ftime.jsonstruct;
+            
+            [model, inputparams] = setupModelFromJson(jsonstruct);
+            
+            inputparams.Control.controlPolicy = 'CCDischarge'; 
+            model.Control = model.setupControl(inputparams.Control);
+            
+            state_init = setupInitialState(model);
+            
+            time_vec    = double(time_vec(:));
+            current_exp = double(current_exp(:));
+            
+            dt_steps = diff(time_vec); 
+            N_sim_steps = length(dt_steps);
+            
+            schedule_p2d = struct();
+            schedule_p2d.step.val = dt_steps;
+            schedule_p2d.step.control = ones(N_sim_steps, 1);
+            
+            schedule_p2d.control.src = @(t, varargin) interp1(time_vec, current_exp, t, 'linear', 0);
+            schedule_p2d.control.Control = struct('stopFunction', @(model, state, state0) false);
+            
+            [~, states_p2d, ~] = simulateScheduleAD(state_init, model, schedule_p2d);
+            
+            N_points_sim = length(states_p2d);
+            V_p2d    = zeros(N_points_sim, 1);
+            I_p2d    = zeros(N_points_sim, 1);
+            time_sim = zeros(N_points_sim, 1);
+            
+            pe_guestStoichiometry0   = model.PositiveElectrode.Coating.ActiveMaterial.Interface.guestStoichiometry0;
[…]

[…]
