### 2026-07-13 9ba4c45
update in runEquivalentCircuitModel
