+++ Battery/EquivalentCircuitModel/utils/createParametersECM.m
@@ -3,10 +3,9 @@ function jsonstruct = createParametersECM()
     % create example data set for ECM
     
     jsonstruct.nominalcellcapacity = 63.7;
-    jsonstruct.initSOC = 1.0;
-    jsonstruct.lowerVoltageCutoff = 3.0;
-    jsonstruct.initOverpotential2 = 0;
-    jsonstruct.totalTime = 4200;
+    jsonstruct.initSOC             = 1.0;
+    jsonstruct.lowerVoltageCutoff  = 3.0;
+    jsonstruct.totalTime           = 4200;
 
     jsonstruct.I.functionFormat = 'tabulated';
     jsonstruct.I.argumentList = {'time'};
@@ -19,19 +18,22 @@ function jsonstruct = createParametersECM()
     jsonstruct.OCP.dataY = [3.126000, 3.273000, 3.380000, 3.447000, 3.470000, 3.480000, 3.489000, 3.500000, 3.515000, 3.533000, 3.551000, 3.567000, 3.582000, 3.594000, 3.602000, 3.609000, 3.614000, 3.620000, 3.625000, 3.631000, 3.637000, 3.643000, 3.649000, 3.657000, 3.665000, 3.674000, 3.684000, 3.696000, 3.710000, 3.727000, 3.750000, 3.775000, 3.798000, 3.818000, 3.838000, 3.858000, 3.877000, 3.897000, 3.917000, 3.937000, 3.958000, 3.979000, 4.000000, 4.022000, 4.044000, 4.066000, 4.089000, 4.112000, 4.135000, 4.159000, 4.182000];
     
     R0_tot = [0.0074814,0.00375135,0.00199926,0.000808379,0.000529892;0.0074814,0.00375135,0.00185474,0.000738337,0.00049771;0.0074814,0.00373265,0.00151744,0.000673703,0.000468303;0.0074814,0.00342697,0.00130825,0.000635903,0.000448193;0.0074814,0.00272078,0.00117717,0.000611909,0.000432971;0.0069826,0.00229452,0.00109773,0.00058627,0.000423262;0.00612671,0.00206227,0.00105217,0.000571475,0.000417812;0.00536849,0.00193337,0.00101405,0.000562212,0.000409251;0.00480278,0.00185103,0.000983687,0.000549631,0.000410476;0.00440167,0.00178588,0.000957044,0.000543895,0.000405368;0.00411575,0.00172477,0.000934531,0.000533558,0.0003994;0.00390187,0.00167949,0.000914869,0.000526768,0.000397983;0.00374658,0.00164546,0.000905014,0.000521591,0.000396355;0.0036334,0.00160835,0.000895256,0.000517151,0.000393944;0.00353281,0.00158063,0.0008793,0.000516464,0.000391954;0.00344791,0.0015608,0.000869252,0.00051227,0.000391406;0.00338977,0.00154377,0.000859461,0.000509065,0.000390664;0.00333872,0.00152388,0.000857998,0.000507685,0.00038969;0.00329657,0.00150926,0.000852066,0.000501813,0.000385995;0.00326141,0.00149673,0.000846844,0.000505136,0.000388294;0.00322799,0.00149188,0.00084232,0.00050115,0.000384211;0.00320237,0.00148102,0.000839376,0.00050073,0.000386191;0.00318137,0.00147045,0.000836217,0.000500865,0.000386833;0.00
[…]
