### 2026-07-13 f5a326e
first fixes for fitting time, 4 arguments for setupOCP to ensure same vector sizes
