+++ Battery/EquivalentCircuitModel/calibration/runCalibration.m
@@ -1,0 +1,162 @@
+%% ECM Fitting for lithium-ion batteries
+
+jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
+jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
+jsonstruct_base     = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
+
+jsonstruct_base.NegativeElectrode.Coating.ActiveMaterial.Interface.useDoubleLayerCapacity = true;
+
+c_ne = 29.866*mol/litre; 
+c_pe = 17.038*mol/litre; 
+
+
+%% 
+% This slider allows logarithmic changes: new_val = val_init * 10^p.
+
+p_diffusion              = 0;
+p_reactionRate           = 0;
+p_doubleLayerCapacitance = 0;
+base_diff_anode          = 1.3135e-15;
+base_rate_anode          = 5.031e-11;
+base_cdl                 = 0.2;
+
+val_diff_anode = base_diff_anode * (10^p_diffusion);
+val_rate_anode = base_rate_anode * (10^p_reactionRate);
+val_cdl = base_cdl * (10^p_doubleLayerCapacitance);
+
+jsonstruct = jsonstruct_base;
+
+jsonstruct.NegativeElectrode.Coating.ActiveMaterial.SolidDiffusion.referenceDiffusionCoefficient = val_diff_anode;
+jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.reactionRateConstant = val_rate_anode;
+jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.doubleLayerCapacitance = val_cdl;
+
+
+[model, inputparams, ~] = setupModelFromJson(jsonstruct);
+initstate = initStateChen2020(model, c_ne, c_pe);
+options = [];
+options.stateInitialization.initializationSetup = 'given state';
+options.stateInitialization.computeSteadyState = false;
+
+extrastructs = [];
+extrastructs.initstate = initstate;
+
+impsolv = ImpedanceSolver(inputparams, options, extrastructs);
+frequences = logspace(-2, 3, 30); 
+
+Z = impsolv.computeImpedance(frequences);
+
+figure('Name', 'Nyquist Diagram');
+plot(real(Z), -imag(Z), '-o', 'LineWidth', 2, 'Color', '#0072BD');
+grid on;
+xlabel('Re(Z)');
+ylabel('-Im(Z)');
+title('Interactive Nyquist Diagram');
+
+disp('Calculation succeeded');
+
+
+%% 
+% 
+% The impact of diffusion is mainly seen on the tail at low frequencies. While 
+% a larger coefficient does not have much of an impact, using a very small one 
+% shows how strongly the diffusion phenomenon is linked to this tail.
+% 
+% The double-layer capacitance affects the first semicircle, as the frequency 
+% at its peak is governed by the relation $f_{\max } =\frac{1\;}{2\pi R_{\textrm{CT}} 
+% C_{\textrm{DL}} }$. A large capacitance will result in a lower $f_{\max }$ (a 
+% shift to the left, as frequencies are plotted inversely), while a small one 
+% will shift the center of the first semicircle to the right.
+% 
+% Finally, the influence of the reaction rate on the Nyquist diagram is clearly 
+% evident. A more facile reaction (i.e., a higher reaction rate) leads to a lower 
+% charge-transfer resistance. As this parameter dictates the width of the first 
[…]
