### 2026-07-13 f0913f5
added class impedance Explore
