+++ Battery/EquivalentCircuitModel/examples/pxdParamsImpact/ImpedanceExplore.m
@@ -1,0 +1,201 @@
+classdef ImpedanceExplore < handle
+
+    properties (SetAccess = private)
+
+        jsonstruct
+        baseValues
+        
+    end
+    
+    properties
+        
+        %%
+        % TODO: add option to change soc in slider
+        soc = 1
+        
+        parameters
+        parameterLegendNames
+        
+    end
+    
+    methods
+
+        function impexp = ImpedanceExplore(jsonstruct)
+
+            impexp.jsonstruct = jsonstruct;
+            impexp.setupDefaultParameters()
+            impexp.baseValues = impexp.getParameterValues();
+            
+        end
+
+        function setupDefaultParameters(impexp)
+
+            elyte   = 'Electrolyte';
+            ne      = 'NegativeElectrode';
+            pe      = 'PositiveElectrode';
+            co      = 'Coating';
+            am      = 'ActiveMaterial';
+            itf     = 'Interface';
+            sd      = 'SolidDiffusion';
+            cc      = 'CurrentCollector';
+            ctrl    = 'Control';
+            thermal = 'ThermalModel';
+ 
+            impexp.parameters = {{ne, co, am, sd,  'referenceDiffusionCoefficient'}, ...
+                                 {ne, co, am, itf, 'reactionRateConstant'}, ...
+                                 {ne, co, am, itf, 'doubleLayerCapacitance'}};
+            
+            %%
+            % TODO : setup default way to get legend
+            impexp.parameterLegendNames = {'Negative Electrode Solid Diffusion Coefficient', ...
+                                           'Negative Electrode Reaction Rate Constant', ...
+                                           'Negative Electrode Double Layer Capacitance'};
+
+        end
+        
+        function updateImpedance(impexp)
+            
+        end
+
+        function sld = setupSlider(imexp, fig, ypos, txt)
+            
+            uilabel(fig                                 , ...
+                    'Position', [50, ypos + 20, 500, 40], ...
+                    'WordWrap', 'on'                    , ...
+                    'Text', txt);
+            
+
+            sld = uislider(fig, ...
+                           'Position', [50, ypos, 500, 3], ...
+                           'Limits', [-5, 5]             , ... % Définit le minimum et le maximum
+                           'MajorTicks', -5:0.5:5        , ... % Ajoute une graduation tous les 1
+                           'Value', 0);
+                
+        end
+
+        function vals = getParameterValues(impexp, jsonstruct)
+
+            if nargin < 2
+                jsonstruct = impexp.jsonstruct;
+            end
+            
+            parameters = impexp.parameters;
+
+            for iparam = 1 : numel(parameters)
+                vals(iparam) = getJsonStructField(jsonstruct, parameters{iparam});
+            end
+            
+        end
+        
+        function jsonstruct = setParameterValues(impexp, vals, jsonstruct)
+
+            if nargin < 3
+                jsonstruct = impexp.jsonstruct;
+            end
+            
+            parameters = impexp.parameters;
+
+            for iparam = 1 : numel(parameters)
+                jsonstruct = setJsonStructField(jsonstruct, parameters{iparam}, vals(iparam), 'handleMisMatch', 'quiet');
+            end
+            
+        end
+
+        function start(impexp)
+            
+        % 2. Création de l'interface graphique (Fenêtre plus haute pour accueillir les curseurs)
+            fig = uifigure('Name', 'Impedance Explorer', 'Position', [100, 50, 600, 650]);
+            
+            % Graphique
+            ax = uiaxes(fig, 'Position', [50, 350, 500, 250]);
+            title(ax, 'Nyquist Diagram');
+            xlabel(ax, 'Re(Z)');
[…]
