### 2026-07-13 dfd5e83
minor changes in runCalibration
