+++ Battery/EquivalentCircuitModel/calibration/runCalibration.m
@@ -1,4 +1,8 @@
 %% ECM Fitting for lithium-ion batteries
+%
+
+%% Impedance computation
+% We setup a P2D model and compute the impedance response
 
 jsonstruct_material = parseBattmoJson(fullfile('ParameterData','ParameterSets','Chen2020','chen2020_lithium_ion_battery.json'));
 jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
@@ -23,23 +27,32 @@ val_cdl = base_cdl * (10^p_doubleLayerCapacitance);
 jsonstruct = jsonstruct_base;
 
 jsonstruct.NegativeElectrode.Coating.ActiveMaterial.SolidDiffusion.referenceDiffusionCoefficient = val_diff_anode;
-jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.reactionRateConstant = val_rate_anode;
-jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.doubleLayerCapacitance = val_cdl;
+jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.reactionRateConstant               = val_rate_anode;
+jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.doubleLayerCapacitance             = val_cdl;
 
-
+%%
+% Setup P2D model from the parameter set |jsonstruct|
 [model, inputparams, ~] = setupModelFromJson(jsonstruct);
+
+%%
+% We use a dedicated function to compute the initial state for the given electrode concentrations.
 initstate = initStateChen2020(model, c_ne, c_pe);
 options = [];
 options.stateInitialization.initializationSetup = 'given state';
 options.stateInitialization.computeSteadyState = false;
 
+%%
+% setup input for |ImpedanceSolver|
 extrastructs = [];
 extrastructs.initstate = initstate;
 
-impsolv = ImpedanceSolver(inputparams, options, extrastructs);
-frequences = logspace(-2, 3, 30); 
 
-Z = impsolv.computeImpedance(frequences);
+%%
+% setup impedance solver and compute the Impedance for a given set of frequencies
+impsolv = ImpedanceSolver(inputparams, options, extrastructs);
+frequencies = logspace(-2, 3, 30); 
+
+Z = impsolv.computeImpedance(frequencies);
 
 figure('Name', 'Nyquist Diagram');
 plot(real(Z), -imag(Z), '-o', 'LineWidth', 2, 'Color', '#0072BD');
@@ -47,8 +60,6 @@ grid on;
 xlabel('Re(Z)');
 ylabel('-Im(Z)');
 title('Interactive Nyquist Diagram');
-
-disp('Calculation succeeded');
 
 
 %% 
@@ -70,8 +81,6 @@ disp('Calculation succeeded');
 % at lower ones.
 
 
-
-
 %% Calibration of ECM parameters
 % Impedance is the opposition to alternating current presented by the combined 
 % effect of resistance and reactance in a circuit. It is obtained with the ratio 
@@ -83,15 +92,12 @@ disp('Calculation succeeded');
 % of the study is to obtain the set of parameters as close as possible to the 
 % real impedance curve with the 2-RC Impedance function $Z = R_0 + \frac{R_1}{1 
 % + j R_1 C_1 \omega} + \frac{R_2}{1 + j R_2 C_2 \omega}$.
-% 
-% We chose unitBoxBFGS as optimization algorithm, which allows other constraints. 
-% As the two RC parallel cells are symetric, we imposed $C_1 > 2*C_2$ to ensure 
-% robustness and injectivity of the optimization. Thanks this criteria, a particular 
[…]
