### 2026-07-14 bb43806
fixes in fittingTime, everything works
