### 2026-07-14 4f8b01b
added function to list the parameters
