+++ Documentation/ecm.rst
@@ -1,0 +1,294 @@
+=================================
+EIS and Equivalent Circuit Models
+=================================
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   runImpedanceScript.nblink
+   runCalibration.nblink
+
+ECM fitting for lithium-ion batteries
+-------------------------------------
+
+Abstract
+^^^^^^^^
+
+Batteries become a central piece of electrification of our energy consumption,
+especially for transports. To improve their performances, innovation is getting
+faster and research is very active. But before building a battery, it is
+important to have an idea of how it will react in different conditions, and
+modeling appears useful. It is necessary to work on complex models that fit the
+best reality, but also to have powerful models that are a lot simpler. This
+model could give important information such as State of Charge or State of
+Health, without spending a lot of time, with the minimum complexity.
+
+We present a work on Equivalent Circuit Model to simulate a complete battery
+through only a few parameters. The battery management system needs the value of
+those parameters to be efficient. ECM are now the most used model for battery
+management systems.
+
+Getting the impedance from PXD Model
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+ECM parameterisation from PXD model using automatic differentiation.
+
+Obtaining the impedance is a major part of the work, as we try to fit the
+impedance given by our parameters to the synthetic impedance from the PXD
+model.
+
+We work with the state vector :math:`\mathcal{X}(t, x)` containing all the
+variables from the PXD model:
+
+.. math::
+
+   \mathcal{X}(t, x) =
+   [c_{\mathrm{ne}}(t, x), \phi_{\mathrm{ne}}(t, x),
+   c_{\mathrm{elyte}}(t, x), \phi_{\mathrm{elyte}}(t, x), \ldots]
+
+The generic equation of the system is
+
+.. math::
+
+   \frac{\partial \mathcal{M}(\mathcal{X})}{\partial t} + \mathcal{A}(\mathcal{X}) = 0.
+
+The operators :math:`\mathcal{M}` and :math:`\mathcal{A}` are non-linear, but
+we discretize them and :math:`\mathcal{X}(t, x)` becomes :math:`X(t)`
+belonging to :math:`\mathbb{R}^{N}`.
+
+The equation then becomes a non-linear system of *ordinary differential
+equations*:
+
+.. math::
+
+   \frac{\partial M(X)}{\partial t} + A(X) = I(t)b(X).
+
+Here :math:`I(t)` is the current and :math:`b` is mostly empty. An implicit
+method is then used for the time discretization. One finds :math:`X^{n+1}`
+from :math:`X^{n}` at time step :math:`n` by solving the equation for the
+unknown :math:`X`:
+
+.. math::
+
+   \frac{M(X) - M(X^{n})}{\Delta t} + A(X) = b(X)I(t^{n+1}).
+
+Then the system is linearized around a state given by :math:`X_{0}` with a
+small perturbation :math:`\delta X = X - X_{0}`. The term :math:`A(X)` becomes
+
+.. math::
+
+   A(X) = A(X_{0}) + D_{A}(X_{0})\delta X,
+
+with :math:`D_{A}` the Jacobian of :math:`A`. Applying the same process for
+:math:`M` and :math:`b` (:math:`D_{M}` and :math:`D_{b}` being their
+Jacobians) gives the following linear system of ordinary differential
+equations:
+
+.. math::
+
+   D_{M}\frac{\partial}{\partial t}\delta X + A(X_{0}) + D_{A}\delta X
+   = I(t)b(X_{0}) + I(t)D_{b}\delta X.
+
+:math:`X_{0}` is a solution to the steady-state problem :math:`I(t) = 0`, which
[…]
