+++ Battery/EquivalentCircuitModel/calibration/FittingEIS.m
@@ -1,15 +1,18 @@
-classdef FittingEIS
+classdef FittingEIS < handle
 
     properties
-        params0
+
+        % Observed impedances (array of same lengths)
+        omega 
         Z_re_exp
         Z_im_exp
-        omega
+
+        % initial parameters
+        params0
+
+        % scales for unit box search algorithm
         scales
         
-        % best_params_found = [] 
-        % fitting_error = []
-
     end
 
 
@@ -17,53 +20,63 @@ classdef FittingEIS
     methods        
 
 
-        function feis = FittingEIS(params0, scales, Z_re_exp, Z_im_exp, omega)
+        function feis = FittingEIS(eisdata, params0, scales)
             
-            % feis.params0  = params0;
-            % feis.scales   = scales;
-            % feis.Z_re_exp = Z_re_exp;
-            % feis.Z_im_exp = Z_im_exp;
-            % feis.omega    = omega;
-
-
-            if istable(Z_re_exp), Z_re_exp = table2array(Z_re_exp); end
-            if istable(Z_im_exp), Z_im_exp = table2array(Z_im_exp); end
-            if istable(omega), omega = table2array(omega); end
-            if istable(params0), params0 = table2array(params0); end
-            if istable(scales), scales = table2array(scales); end
-
-            % Everything has to be "double" and colon (:)
-            feis.params0 = double(params0(:));
-            feis.scales  = double(scales(:));
-            feis.Z_re_exp= double(Z_re_exp(:));
-            feis.Z_im_exp= double(Z_im_exp(:));
-            feis.omega   = double(omega(:));
-
-            %pb here because the way I take exp data changes a lot the
-            %result...
+            feis.Z_re_exp = eisdata.Z_re_exp;
+            feis.Z_im_exp = eisdata.Z_im_exp;
+            feis.omega    = eisdata.omega;
+            feis.params0  = params0;
+            
+            if nargin < 3 || isempty(scales)
+                feis.setupDefaultScales();
+            end
 
         end
 
+        function setupDefaultScales(feis)
+
+            params0 = feis.params0;
+
+            a = 1000;
+            
+            pmin = params0 / a;
+            pmax = params0 * a;
+            scales = [pmin; pmax];
+
+            feis.scales = scales;
+            
+        end
+        
         function p_norm = unscaled2scaled(feis, p)
+            
             pmin = feis.scales(1:5);
             pmax = feis.scales(6:10);
             p_norm = (p-pmin)./(pmax-pmin);
+            
         end
 
         function p = scaled2unscaled(feis, p_norm)
+            
             pmin = feis.scales(1:5);
             pmax = feis.scales(6:10);
             p = (pmax-pmin).*p_norm +pmin;
+            
         end
 
+
+        function [optparams, fitting_error] = run(feis)
+            
+            [~, ~, optparams, fitting_error] = feis.optimizationBFGS();
+            
+        end
+        
         %% Thevenin Model
 
         function [min_value, history, best_params, fitting_error] = optimizationBFGS(feis)
-            
 
             f_opt = @(p_norm) feis.optifunc(scaled2unscaled(feis, p_norm));
             
-            params0_norm = unscaled2scaled(feis, feis.params0);
+            params0_norm = feis.unscaled2scaled(feis.params0);
             % best_params_norm = lsqnonlin(deltagap, params0_norm, lb_norm, ub_norm, [0,0,-1,0,10], 0);
             params0_norm = params0_norm(:);
            
[…]
