+++ Battery/EquivalentCircuitModel/calibration/WarburgFittingEIS.m
@@ -1,0 +1,207 @@
+classdef  WarburgFittingEIS < FittingEIS
+
+    methods        
+
+        function feis = WarburgFittingEIS(params0, eisdata, scales)
+
+            if nargin < 3
+                scales = [];
+            end
+            
+            feis = feis@FittingEIS(params0, eisdata, scales);
+
+        end
+
+        function [min_value, history, best_params, fitting_error] = optimizationBFGS(feis)
+
+            %% Warburg Model
+            error('not updated....');
+            
+            f_opt = @(p_norm) feis.optifunc(p_norm.* feis.scales) ;
+            
+            params0_norm = feis.params0 ./ feis.scales; 
+            params0_norm = params0_norm(:);
+          
+            
+
+            [min_value, best_params_norm, history] = unitBoxBFGS(...
+                params0_norm, ...                             
+                f_opt, ...           
+                'maximize', false, ...      
+                'enforceFeasible', true, ...     
+                'maxIt', 300, ...               
+                'objChangeTol', 1e-6, ...       
+                'gradTol', 1e-5, ...
+                'lineSearchMaxIt', 200 ...
+                                           );
+            best_params = best_params_norm(:) .* feis.scales(:);
+
+            fitting_error = feis.optifunc(best_params);
+
+        end
+
+        function [v, g_norm] = optifunc(feis, p)
+        
+            % Preventing physical values to get too close from 0
+            p_safe = max(p, 1e-7);
+
+
+            w  = feis.omega(:);
+            R0 = p_safe(1);
+            R1 = p_safe(2);
+            Q1 = p_safe(3);
+            a1 = p_safe(4);
+            R2 = p_safe(5);
+            Q2 = p_safe(6);
+            a2 = p_safe(7);
+            Q  = p_safe(8);
+            L  = p_safe(9);
+
+            s = 1i .* w;
+            s = s(:);
+            
+            sa1 = s.^a1;
+            sa2 = s.^a2;
+
+            Z_param_cplx = R0 + s.*L + R1 ./ (1 + R1.*Q1.*sa1) + R2 ./ (1 + R2.*Q2.*sa2) + 1 ./ (Q.*(s.^0.5));
+    
+            Z_re_param = real(Z_param_cplx);
+            Z_im_param = imag(Z_param_cplx);
+
+            if istable(Z_re_param), Z_re_param = table2array(Z_re_param); end
+            if istable(Z_im_param), Z_im_param = table2array(Z_im_param); end
+            
+           
+            Z_re_param = double(Z_re_param(:));
+            Z_im_param = double(Z_im_param(:));
+
+            Modulus_Z = sqrt(Z_re_param.^2 + Z_im_param.^2);
+            Modulus_Z = Modulus_Z(:);
+            
+            err_re = (feis.Z_re_exp(:) - Z_re_param(:)) ./ Modulus_Z(:);
+            err_im = (feis.Z_im_exp(:) - Z_im_param(:)) ./ Modulus_Z(:);
+            
+            v = sum(err_re.^2 + err_im.^2);
+             
+            
+            if nargout > 1
+
+            den1 = (1 + R1 .* Q1 .* sa1).^2;
+            den2 = (1 + R2 .* Q2 .* sa2).^2;
+            
+            
+            dZ_dR0 = ones(size(s));                     %  R0
+            dZ_dR1 = 1 ./ den1;                         %  R1
+            dZ_dR2 = 1 ./ den2;                         %  R2
+            
+            dZ_dQ1 = - (R1^2 .* sa1) ./ den1;           % Q1
[…]
