### 2026-07-16 06e05d5
Impedance explorer ready for notebook
