+++ Battery/EquivalentCircuitModel/examples/pxdParamsImpact/ImpedanceExplore.m
@@ -1,7 +1,5 @@
 classdef ImpedanceExplore < handle
-
     properties (SetAccess = immutable)
-
         jsonstruct
         model
         
@@ -9,8 +7,8 @@ classdef ImpedanceExplore < handle
     
     properties
         
-        %%
-        % TODO: add option to change soc in slider
+        
+        % done : add option to change soc in slider
         soc = 1
         
         parameters
@@ -18,9 +16,26 @@ classdef ImpedanceExplore < handle
     end
     
     methods
-
+        %%
+        % Constructor method that initializes the application, auto-loading the physical P2D Chen2020 battery dataset from BattMo files if no input JSON structure is provided.
         function impexp = ImpedanceExplore(jsonstruct)
-
+            if nargin < 1 
+                fprintf('Automatic initialization of Chen2020 configuration (BattMo)...\n');
+                try
+                    jsonstruct_material = parseBattmoJson(fullfile('ParameterData', 'ParameterSets', 'Chen2020', 'chen2020_lithium_ion_battery.json'));
+                    jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
+                    
+                    jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
+                    if ~isfield(jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface, 'doubleLayerCapacitance')
+                        jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.doubleLayerCapacitance = 0.2; 
+                        fprintf('-> default doubleLayerCapacitance, value = 0.2 F/m^2\n');
+                    end
+                    
+                catch ME
+                    error('ImpedanceExplore:AutoLoadFailed', ...
+                        ME.message);
+                end
+            end
             model = setupModelFromJson(jsonstruct);
             impexp.jsonstruct = jsonstruct;
             impexp.model      = model;
@@ -28,17 +43,19 @@ classdef ImpedanceExplore < handle
             impexp.setupDefaultParameters()
             
         end
-
+        
+        %%
+        % Utility method to flatten the active JSON structure and print all available scalar parameters in the model to the command window.
         function viewOptions(impexp)
-
             fjv = flattenJsonStruct(impexp.jsonstruct, 'doprint', false);
             fprintf('\n\n*** List of all the scalar parameters in the model ***\n');
             fjv.print('filter', {'value', @(val) (isnumeric(val) && isscalar(val))});
             
         end
-            
+        
+        %%
+        % Configures the default internal hierarchical paths of the active material parameters to be observed and modified.
         function setupDefaultParameters(impexp)
-
             elyte   = 'Electrolyte';
             ne      = 'NegativeElectrode';
             pe      = 'PositiveElectrode';
@@ -53,11 +70,11 @@ classdef ImpedanceExplore < handle
             impexp.parameters = {{ne, co, am, sd,  'referenceDiffusionCoefficient'}, ...
                                  {ne, co, am, itf, 'reactionRateConstant'}, ...
                                  {ne, co, am, itf, 'doubleLayerCapacitance'}};
-
         end
-
+        
+        %%
+        % Helper method that generates human-readable parameter names by joining the individual sub-keys of their nested JSON paths.
         function pnames = setupParameterLegendNames(impexp)
-
             params = impexp.parameters;
             for iparam = 1 : numel(params)
                 pnames{iparam} = strjoin(params{iparam});
@@ -65,145 +82,175 @@ classdef ImpedanceExplore < handle
             
         end
         
-        function sld = setupSlider(imexp, fig, ypos, txt)
+        %%
[…]
