### 2026-07-16 19d4013
runImpedanceexplore modified to suit a notebook
