+++ Battery/EquivalentCircuitModel/examples/pxdParamsImpact/ImpedanceExplore.m
@@ -1,90 +1,71 @@
 classdef ImpedanceExplore < handle
     properties (SetAccess = immutable)
-        jsonstruct
-        model
-        
+        jsonstruct % Holds the base P2D model configuration structure
+        model      % Initialized BattMo model object
     end
     
     properties
-        
-        
-        % done : add option to change soc in slider
-        soc = 1
-        
-        parameters
-        
+        soc = 1 % Default State of Charge (0 to 1)
+        parameters % Cell array storing the hierarchical JSON paths of observed parameters
     end
     
     methods
-        %%
-        % Constructor method that initializes the application, auto-loading the physical P2D Chen2020 battery dataset from BattMo files if no input JSON structure is provided.
         function impexp = ImpedanceExplore(jsonstruct)
-            if nargin < 1 
-                fprintf('Automatic initialization of Chen2020 configuration (BattMo)...\n');
+            % Constructor: auto-loads Chen2020 P2D model if no structure is provided
+            if nargin < 1 || isempty(jsonstruct)
+                fprintf('Initializing default Chen2020 P2D model...\n');
                 try
                     jsonstruct_material = parseBattmoJson(fullfile('ParameterData', 'ParameterSets', 'Chen2020', 'chen2020_lithium_ion_battery.json'));
                     jsonstruct_geometry = parseBattmoJson(fullfile('Examples', 'JsonDataFiles', 'geometryChen.json'));
-                    
                     jsonstruct = mergeJsonStructs({jsonstruct_material, jsonstruct_geometry});
-                    if ~isfield(jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface, 'doubleLayerCapacitance')
-                        jsonstruct.NegativeElectrode.Coating.ActiveMaterial.Interface.doubleLayerCapacitance = 0.2; 
-                        fprintf('-> default doubleLayerCapacitance, value = 0.2 F/m^2\n');
-                    end
                     
+                    % Safe default fallback for capacitance
+                    ne = 'NegativeElectrode'; co = 'Coating'; am = 'ActiveMaterial'; itf = 'Interface';
+                    if ~isfield(jsonstruct.(ne).(co).(am).(itf), 'doubleLayerCapacitance')
+                        jsonstruct.(ne).(co).(am).(itf).useDoubleLayerCapacity = true;
+                        jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = 0.2;
+                    end
                 catch ME
-                    error('ImpedanceExplore:AutoLoadFailed', ...
-                        ME.message);
+                    error('ImpedanceExplore:AutoLoadFailed', 'Failed to auto-load default files: %s', ME.message);
                 end
             end
-            model = setupModelFromJson(jsonstruct);
+            
             impexp.jsonstruct = jsonstruct;
-            impexp.model      = model;
-            
-            impexp.setupDefaultParameters()
-            
+            impexp.model      = setupModelFromJson(jsonstruct);
+            impexp.setupDefaultParameters();
         end
         
-        %%
-        % Utility method to flatten the active JSON structure and print all available scalar parameters in the model to the command window.
         function viewOptions(impexp)
+            % Helper to print all available scalar properties in the model JSON
             fjv = flattenJsonStruct(impexp.jsonstruct, 'doprint', false);
             fprintf('\n\n*** List of all the scalar parameters in the model ***\n');
             fjv.print('filter', {'value', @(val) (isnumeric(val) && isscalar(val))});
-            
         end
-        
-        %%
-        % Configures the default internal hierarchical paths of the active material parameters to be observed and modified.
+            
         function setupDefaultParameters(impexp)
-            elyte   = 'Electrolyte';
[…]
