### 2026-07-17 c29a218
calibrateEcmFromP2D now returns a jsonfile
