+++ Battery/EquivalentCircuitModel/calibration/calibrateEcmFromP2D.m
@@ -1,111 +1,203 @@
 function jsonstruct_ecm = calibrateEcmFromP2D(jsonstruct_p2d, soc_step)
-
-    % soc_step : soc increment between 0 and 1. Start from soc = 1.
+    % CALIBRATEECMFROMP2D Calibrates a 2-RC Equivalent Circuit Model (ECM)
+    % from a physics-based P2D model across a range of State of Charge (SOC) points.
+    %
+    % Inputs:
+    %   jsonstruct_p2d - P2D model configuration structure
+    %   soc_step       - SOC decrement interval (default: 0.1)
+    %
+    % Output:
+    %   jsonstruct_ecm - Standardized BattMo ECM configuration structure
     
-    %% 1. Initialisation
-
-    [model, inputparams] = setupModelFromJson(jsonstruct_p2d);
-    state_current = setupInitialState(model);
-    if nargin < 1
+    %% 1. Initialization
+    if nargin < 2
         soc_step = 0.1; 
     end
+    
+    [model, inputparams] = setupModelFromJson(jsonstruct_p2d);
+    state_current = setupInitialState(model);
+    
     soc_list = (1.0:-soc_step:0.1)';
     N_steps = length(soc_list);
     
-    results_matrix = zeros(N_steps, 6);
+    results_matrix = zeros(N_steps, 6); % Stores [SOC, R0, R1, C1, R2, C2]
+    ocv_list = zeros(N_steps, 1);       % Stores Open Circuit Voltage for each SOC
     
     freq = logspace(-4, 2, 50)./(2*pi);
     C_rate = 1;
     t_xpercent = 3600 * soc_step / C_rate;
     
-    %% 2. Main loop
+    %% 2. Main Loop
     for i = 1:N_steps
         current_soc = soc_list(i);
         fprintf('\n==================================================\n');
         fprintf('  EIS fitting for SOC = %.2f (%d/%d)\n', current_soc, i, N_steps);
         fprintf('==================================================\n');
         
-        %% extraction de l'EIS par computeImpedance()
+        % Robustness fallback: ensure the initial state has evaluated terminal voltage (.Control.E)
+        if i == 1 && (~isfield(state_current, 'Control') || ~isfield(state_current.Control, 'E'))
+            inputparams.Control.DRate = 0;
+            model.Control = model.setupControl(inputparams.Control);
+            schedule_init = struct();
+            schedule_init.step.val = 0.001;
+            schedule_init.step.control = 1;
+            schedule_init.control.src = @(t, varargin) 0;
+            schedule_init.control.Control = struct('stopFunction', @(model, state, state0) false);
+            [~, states_init, ~] = simulateScheduleAD(state_current, model, schedule_init);
+            state_current = states_init{end};
+        end
+        
+        % Capture the relaxed terminal voltage as the Open Circuit Voltage (OCV)
+        ocv_list(i) = state_current.Control.E;
+        fprintf('  Measured Open Circuit Voltage (OCV): %.4f V\n', ocv_list(i));
+        
+        % Run the Impedance Solver
         options = struct();
         options.stateInitialization.computeSteadyState = false;
-
         options.stateInitialization.initializationSetup = 'given state'; 
         extrastructs.initstate = state_current;
         
         impsolv = ImpedanceSolver(inputparams, options, extrastructs);
         
-        fprintf('Linearisation...\n');
[…]
