### 2026-08-07 c958adf
Add clear and close to generic step control examples
