+++ Examples/Advanced/runGenericStepControlCycle.m
@@ -4,6 +4,9 @@
 % <https://github.com/BattMoTeam/BattMo/blob/main/Utilities/JsonSchemas/GenericControl.schema.json json schema
 % documentation>.
 
+clear all
+close all
+
 %% Input setup
 %
 
@@ -44,7 +47,7 @@ jsonstruct_material = removeStructFields(jsonstruct_material              , ...
 
 %%
 % We merge all the input structures
-% 
+%
 
 jsonstruct = mergeStructs({jsonstruct_material, ...
                                jsonstruct_geometry, ...
@@ -65,7 +68,7 @@ jsonstruct.include_current_collectors = false;
 output = runBattery(jsonstruct);
 
 %% Plotting
-% 
+%
 
 states = output.states;
 
@@ -84,4 +87,3 @@ plot(time/hour, I, '*-', 'linewidth', 3);
 grid on
 xlabel 'time  / h';
 ylabel 'Current  / A';
-
