### 2026-08-07 5e10766
Create setupScheduleStep function such that a proper schedule is created
