### 2026-08-07 c22c5c0
use tighter estimate
