+++ Control/GenericControlModel.m
@@ -1,29 +1,29 @@
 classdef GenericControlModel < ControlModel
 
     properties
-        
+
         controlsteps % cell array, each cell describing a control (see GenericControlModel.schema.json)
-        
+
     end
-    
+
     methods
-        
+
         function model = GenericControlModel(inputparams)
-            
+
             model = model@ControlModel(inputparams);
-            
+
             fdnames = {'controlsteps'};
             model = dispatchParams(model, inputparams, fdnames);
 
             model = model.parseControlSteps();
             model = model.setupTerminationFunctions();
-            
+
         end
-        
+
         function model = registerVarAndPropfuncNames(model)
-            
+
             model = registerVarAndPropfuncNames@ControlModel(model);
-            
+
             varnames = {};
             % Control type : string that can take following value
             % - 'rest'
@@ -38,49 +38,49 @@ classdef GenericControlModel < ControlModel
             varnames{end + 1} = 'ctrlSwitchTime';
             % Time spanned since beginning
             varnames{end + 1} = 'time';
-            
+
             model = model.registerVarNames(varnames);
-            
+
             model = model.setAsStaticVarNames({'ctrlStepIndex', 'ctrlSwitchTime'});
-            
+
             % Register the functions
             fn = @GenericControlModel.updateControlType;
             model = model.registerPropFunction({'ctrlType', fn, {'ctrlStepIndex'}});
-            
+
             fn = @GenericControlModel.updateControlTime;
             model = model.registerPropFunction({'ctrlTime', fn, {'time', 'ctrlSwitchTime'}});
 
             fn = @GenericControlModel.updateControlEquation;
             model = model.registerPropFunction({'controlEquation', fn, {'ctrlType', 'E', 'I', 'ctrlTime'}});
-            
+
         end
-        
+
         function cleanState = addStaticVariables(model, cleanState, state)
-            
+
             cleanState.ctrlStepIndex  = state.ctrlStepIndex;
             cleanState.ctrlSwitchTime = state.ctrlSwitchTime;
 
         end
-        
+
         function newstate = addVariablesAfterConvergence(model, newstate, state)
 
             newstate.ctrlTime = state.ctrlTime;
-            
+
         end
 
 
         function model = parseControlSteps(model)
 
             model.controlsteps = GenericControlModel.unfoldCycle(model.controlsteps);
-            
+
         end
 
-        
+
         function model = setupTerminationFunctions(model)
-            
+
         % For each control step termination, setup the function that determines if the termination condition is fulfilled. This
         % function is stored under controlstep.termination.function.
-            
+
             controlsteps = model.controlsteps;
 
             for ictrl = 1 : numel(controlsteps)
@@ -89,9 +89,9 @@ classdef GenericControlModel < ControlModel
             end
 
             model.controlsteps = controlsteps;
-            
+
         end
-        
+
         function schedule = setupSchedule(model, jsonstruct)
 
             if nargin > 1 && isfield(jsonstruct, 'TimeStepping')
@@ -115,12 +115,15 @@ classdef GenericControlModel < ControlModel
             else
                 params = [];
             end
+            params = model.parseTimeSteppingStruct(params);
 
             controlsteps = model.controlsteps;
             dts = cellfun(@(ctrlstep) ctrlstep.timeStepSize, controlsteps);
 
             if isAssigned(params, {'totalTime'})
                 totalTime = params.totalTime;
+                % With a global totalTime, future generic control switches are not known when the schedule is
+                % constructed. Take the min(dts) for safety.
                 dt = min(dts);
                 nsteps = ceil(totalTime/dt);
                 vals = repmat(dt, nsteps, 1);
[…]
