### 2026-08-07 2b35478
Allow for setting a bgfactor that multiplies electrolyte region Bruggeman coeffs
