### 2026-08-10 2308b05
make the bgfactor proportional to the elyte conductivity and diffusion
