### 2026-08-11 #11114 @editorialbot
[REVIEW]: BattMo - Battery Modelling Toolbox
**Submitting author:** @augustjohansson (<a href="http://orcid.org/0000-0001-6950-6016">August Johansson</a>)
**Repository:** https://github.com/BattMoTeam/BattMo
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.0
