### 2026-08-12 2d34f23
started with co2 capture
