+++ CO2capture/CO2capture.m
@@ -1,0 +1,246 @@
+classdef CO2capture < BaseModel
+    
+    properties
+
+        Feed
+        Permeate
+        
+        constants
+        
+        nGas   % Number of gas (each of them will have a partial pressure). Only needed when gasSupplyType == 'coupled'
+        gasInd % Structure whose fieldname give index number of the corresponding gas component.
+
+        permeabilities % structure with permability for each gas
+        thickness
+
+    end
+    
+    properties (SetAccess = immutable)
+
+        %% helpers
+        permeabilityValues % vector computed from permeabilities
+
+    end
+    
+    methods
+        
+        function model = CO2capture(inputparams)
+            
+            model = model@BaseModel();
+            
+            fdnames = {'permeabilities', ...
+                       'thickness'};
+
+            model = dispatchParams(model, inputparams, fdnames);
+            
+            model.Feed     = CO2captureChannel(inputparams.Feed);
+            model.Permeate = CO2captureChannel(inputparams.Permeate);
+            
+            model.constants = PhysicalConstants();
+
+            model = CO2capture.setupGasStructures(model);
+
+            nGas   = model.nGas;
+            gasInd = model.gasInd;
+            
+            fdnames = fieldnames(model.permeabilities);
+            assert(numel(fdnames) == nGas, 'problem in setup');
+            
+            for igas = 1 : nGas
+                fdname = fdnames{igas};
+                permValues(gasInd.(fdname)) = model.permeabilities.(fdname);
+            end
+
+            model.permeabilityValues = permValues;
+            
+        end
+        
+        function model = registerVarAndPropfuncNames(model)
+            
+            model = registerVarAndPropfuncNames@BaseModel(model);
+
+            nGas = model.nGas;
+            
+            varnames = {};
+            % transfer fluxes [mol/s/m]
+            varnames{end + 1} = VarName({}, 'crossFluxes', nGas);
+            
+            model = model.registerVarNames(varnames);
+
+            channels = {'Feed', 'Permeate'};
+
+            for igas = 1 : nGas
+                
+                for  ichannel = 1 : numel(channels)
+
+                    channel = channels{ichannel};
+                    
+                    fn = @CO2capture.updateMassSources;
+                    inputvarnames = {VarName({}, 'crossFluxes', nGas, igas), ...
+                                     VarName({channel}, 'bcSources', nGas, igas)};
+                    outputvarname = VarName({channel}, 'massSources', nGas, igas);
+                    model = model.registerPropFunction({outputvarname, fn, inputvarnames});
+
+                end
+                
+                fn = @CO2capture.updateTransferRates;
+                inputvarnames = {VarName({'Feed'}, 'pressures', nGas, igas), ...
+                                 VarName({'Permeate'}, 'pressures', nGas, igas)};
+                outputvarname = VarName({}, 'crossFluxes', nGas, igas);
+                model = model.registerPropFunction({outputvarname, fn, inputvarnames});                
+
+                
+            end
+
+        end
+
+        function initstate = setupInitialState(model)
+
+            initstate.Feed     = model.Feed.setupInitialState();
+            initstate.Permeate = model.Permeate.setupInitialState();
+            
+        end
+
+        function state = addVariables(model, state)
+
+        % Given a state where only the primary variables are defined, this
+        % functions add all the additional variables that are computed in the assembly process and have some physical
+        % interpretation.
+        %
+        % To do so, we use getEquations function and sends dummy variable for state0, dt and drivingForces
+
+            dt            = 1;
+            state0        = state;
+            drivingForces = model.getValidDrivingForces();
+
+            [~, state] = getEquations(model, state0, state, dt, drivingForces, 'ResOnly', true);
+
+        end
+
[…]
