### 2026-08-12 7f4633c
update in startupBattMo
