+++ startupBattMo.m
@@ -47,7 +47,8 @@ dirnames = {'Battery'         , ...
             'Electrolyser'    , ...
             'SeaWater'        , ...
             'JuliaBridge'     , ...
-            'ProtonicMembrane'};
+            'ProtonicMembrane', ...
+            'CO2capture'};
 
 for ind = 1 : numel(dirnames)
     dirname = fullfile(rootdirname, dirnames{ind});
