### 2026-08-12 b1b00ea
work in progress for CO2 capture
