+++ CO2capture/CO2captureChannelControl.m
@@ -1,0 +1,19 @@
+classdef CO2captureChannelControl < BaseModel
+
+    properties
+
+    end
+
+    methods
+        
+        function model = CO2captureChannelControl(inputparams)
+            
+            model = model@BaseModel();
+            
+        end
+
+                
+    end
+
+end
+    
