+++ CO2capture/CO2captureControl.m
@@ -1,0 +1,22 @@
+classdef CO2captureControl < BaseModel
+
+    properties
+
+    end
+
+    methods
+        
+        function model = CO2captureChannelControl(inputparams)
+            
+            model = model@BaseModel();
+            
+        end
+
+        function model = setupSchedule()
+
+        end
+                
+    end
+
+end
+    
