### 2026-08-13 70f1ebb
Bump the documentation-python group in /Documentation with 3 updates
Bumps the documentation-python group in /Documentation with 3 updates: [docutils](https://github.com/rtfd/recommonmark), [packaging](https://github.com/pypa/packaging) and [traitlets](https://github.com/ipython/traitlets).


Updates `docutils` from 0.22.4 to 0.23
- [Changelog](https://github.com/readthedocs/recommonmark/blob/master/CHANGELOG.md)
- [Commits](https://github.com/rtfd/recommonmark/commits)

Updates `packaging` from 26.2 to 26.3
- [Release notes](https://github.com/pypa/packaging/releases)
- [Changelog](https://gi
[…]
