+++ Documentation/requirements.txt
@@ -11,7 +11,7 @@ click==8.4.2
 contourpy==1.3.3
 cycler==0.12.1
 defusedxml==0.7.1
-docutils==0.22.4
+docutils==0.23
 fastjsonschema==2.22.1
 fonttools==4.63.0
 h11==0.16.0
@@ -40,7 +40,7 @@ nblink==0.1.12
 nbsphinx==0.9.8
 nbsphinx-link==1.4.1
 numpy==2.5.1
-packaging==26.2
+packaging==26.3
 pandocfilters==1.5.1
 parso==0.8.7
 pathspec==1.1.1
@@ -79,7 +79,7 @@ sphinxcontrib-youtube==1.5.0
 tinycss2==1.5.1
 tornado==6.5.7
 tqdm==4.70.0
-traitlets==5.15.1
+traitlets==5.16.1
 typing_extensions==4.16.0
 urllib3==2.7.0
 webencodings==0.5.1
