### 2026-08-16 #11114 @yuefan98
Hi @augustjohansson, this is really nice and well-documented work for battery modeling. I will keep exploring the core functionalities, but several small issues catch my attention at this point.
1. BattMoApp is not accessible through the link in the paper and on the website. Can you fix that?
2. The [Change Control Parameters](https://battmo.org/BattMo/basicusage.html#change-control-parameters) section is not reproducible. I think it is because "Crate" is used here rather than "Drate". I feel this can be hard to track with many different simulations. Have you thought about using signed current for charge and discharge implementation?
3. In [Change Materials Parameters](https://battmo.org/BattMo/basicusage.html#change-material-parameters), the LFP.json is not available, which needs to be updated with the latest file name.
4. crameri is not working for some reason. As a result, plotBatteryGrid() function is not working. The error message is as follows `Unknown colormap name 'vik'. Try typing crameri with no inputs to check the options and try again.`
5. In [this example](https://battmo.org/BattMo/nblinks/runTimeControlExample.html#We-finalize-the-setup-of-the-jsonstruct-input), runBatteryJson is no longer available, which should be updated.
6. The CCCV control overshoots the voltage pretty significantly (~ 40 mV for the runBatteryP2D example) as compared to other software packages like COMSOL and PyBaMM. Have the authors observed this behavior as well? Is this a known limitation of the current CCCV implementation, or might there be some design considerations that I haven't caught.
