### 2026-08-17 26b5438
Add sphinx copybutton
