+++ Documentation/requirements.txt
@@ -15,9 +15,11 @@ docutils==0.22.4
 fastjsonschema==2.22.1
 fonttools==4.63.0
 h11==0.16.0
-habanero>=2.9.2
+habanero==2.9.2
 httpcore==1.0.9
+httpcore2==2.10.0
 httpx==0.28.1
+httpx2==2.10.0
 idna==3.18
 imagesize==2.0.0
 jedi==0.20.0
@@ -32,7 +34,7 @@ latexcodec==3.0.1
 MarkupSafe==3.0.3
 matplotlib==3.11.1
 mistune==3.3.4
-mypy-extensions==1.1.0
+mypy_extensions==1.1.0
 nbclient==0.11.0
 nbconvert==7.17.1
 nbformat==5.10.4
@@ -56,6 +58,7 @@ PyYAML==6.0.3
 pyzmq==27.1.0
 referencing==0.37.0
 requests==2.34.2
+roman-numerals==4.1.0
 rpds-py==2026.6.3
 setuptools==83.0.0
 six==1.17.0
@@ -63,9 +66,10 @@ sniffio==1.3.1
 snowballstemmer==3.1.1
 soupsieve==2.9.1
 Sphinx==9.1.0
-sphinx-rtd-theme==3.1.0
+sphinx-copybutton==0.5.2
 sphinx_collapse==0.1.4
 sphinx_design==0.7.0
+sphinx_rtd_theme==3.1.0
 sphinxcontrib-applehelp==2.0.0
 sphinxcontrib-bibtex==2.7.0
 sphinxcontrib-devhelp==2.0.0
@@ -80,6 +84,7 @@ tinycss2==1.5.1
 tornado==6.5.7
 tqdm==4.70.0
 traitlets==5.15.1
+truststore==0.10.4
 typing_extensions==4.16.0
 urllib3==2.7.0
 webencodings==0.5.1
