### 2026-08-17 #11114 @augustjohansson
Dear yuefan98,

Thank you very much for reviewing!

Regarding 1. there must be a temporary problem with the web server. We're looking into it.
For 2--5, these may be caused by differences in branch versions, since the documentation is for the main branch which might have deviated from the joss branch. I will merge main to joss and make sure everything is up to date and works.
For 6, we will look into this and get back to you shortly.

Best regards
August
