### 2026-08-18 #11114 @yuefan98
Hi @augustjohansson, thanks for the prompt response. I double-checked and confirmed that I pulled the latest version from the main branch on GitHub. For 2-5, I think there might be some misalignment between the documentation and the actual software syntax following the recent updates. It might be worth rerunning those examples to make sure they still work as documented.
