### 2026-08-18 177553f
change default value in CCCV switch tolerance
