### 2026-08-18 #11114 @xavierr
Hi @yuefan98

Thank you very much for taking the time to review our code.

The switching time, as for example in a CCCV control, is a quantity that is computed numerically and with a certain
level of error. The error is controlled by the option under `switchTolerances`, see [here](https://github.com/BattMoTeam/BattMo/blob/main/Utilities/JsonSchemas/ControlModel.schema.json#L93). We have changed the default
value to a lower threshold (see [commit](https://github.com/BattMoTeam/BattMo/commit/177553ff426367c4ed1cf85b632f1d933caf6397)). This could be changed for a specific simulation by modifying the input struct, for example

    jsonstruct.Control.switchTolerances.CC_charge1 = 1e-3

Note that lowering the tolerances may induce higher computational cost, as the solver spends
more time improving the value of the switching time.

This can be seen when we compare the cases with `CC_charge1` relative tolerance value of 1e-3 and 1e-4 (picture below, we zoom at the first CC-CV switch). With the lowest tolerance value, we manage to can capture precisely the switching time, at the cost of more discretization
points.

Best regards

Xavier.
