### 2026-08-18 3d8a980
Update basic discharge doc example
