+++ Documentation/basicusage.rst
@@ -189,16 +189,16 @@ in the jsonstruct and re-run the simulation.
 
 .. code:: matlab
 
-   CRates = [0.5, 1, 2];
+   DRates = [0.8, 1, 2];
    figure()
-   for i = 1 : numel(CRates)
-       jsonstruct.Control.CRate = CRates(i);
+   for i = 1 : numel(DRates)
+       jsonstruct.Control.DRate = DRates(i);
        output = runBattery(jsonstruct);
 
        states = output.states;
        time = cellfun(@(state) state.time, states);
        voltage = cellfun(@(state) state.Control.E, states);
-       plot((time/hour), voltage, '-', 'linewidth', 3)
+       plot(time/hour, voltage, '-', 'linewidth', 3)
        hold on
    end
    hold off
