### 2026-08-18 1d1e3e1
Update LFP json reference
