### 2026-08-18 bbb7fde
Fix bug in readme.rst
