### 2026-08-18 #11114 @augustjohansson
Hi again and thank you for your feedback. I've tried to address them here:

> Hi [@augustjohansson](https://github.com/augustjohansson), this is really nice and well-documented work for battery modeling. I will keep exploring the core functionalities, but several small issues catch my attention at this point.

Thank you for your kind words.

>     1. BattMoApp is not accessible through the link in the paper and on the website. Can you fix that?

The app is now up and running at [https://app.battmo.org](https://app.battmo.org).

>     2. The [Change Control Parameters](https://battmo.org/BattMo/basicusage.html#change-control-parameters) section is not reproducible. I think it is because "Crate" is used here rather than "Drate". I feel this can be hard to track with many different simulations. Have you thought about using signed current for charge and discharge implementation?

You're absolutely correct in that changing from C to D fixes this issue. We have considered having different signs for charge and discharge, but felt that for entry-level users it might be easier with letters rather than signs.

>     3. In [Change Materials Parameters](https://battmo.org/BattMo/basicusage.html#change-material-parameters), the LFP.json is not available, which needs to be updated with the latest file name.

Thanks, the filename should have been LFP_Xu2015.json. This have been corrected.

>     4. crameri is not working for some reason. As a result, plotBatteryGrid() function is not working. The error message is as follows `Unknown colormap name 'vik'. Try typing crameri with no inputs to check the options and try again.`

I think that this is due to the fact that BattMo requires git lfs enabled. We have mentioned this in the documentation, but I have now made it clearer.

>     5. In [this example](https://battmo.org/BattMo/nblinks/runTimeControlExample.html#We-finalize-the-setup-of-the-jsonstruct-input), runBatteryJson is no longer available, which should be updated.

Indeed, thank you for finding this. This has now been corrected.

>     6. The CCCV control overshoots the voltage pretty significantly (~ 40 mV for the runBatteryP2D example) as compared to other software packages like COMSOL and PyBaMM. Have the authors observed this behavior as well? Is this a known limitation of the current CCCV implementation, or might there be some design considerations that I haven't caught.

Please see Xavier's reply above: using a stricter switching tolerance yields less overshoot. Thank you for finding this issue. The value has now been updated.

Just let us know if you find more issues or have any questions.

All the best
August
