### 2026-08-20 aad9e81
Bump the documentation-python group in /Documentation with 8 updates
Bumps the documentation-python group in /Documentation with 8 updates:

| Package | From | To |
| --- | --- | --- |
| [charset-normalizer](https://github.com/jawah/charset_normalizer) | `3.4.9` | `3.5.0` |
| [docutils](https://github.com/rtfd/recommonmark) | `0.22.4` | `0.23` |
| [nbformat](https://github.com/jupyter/nbformat) | `5.10.4` | `5.11.0` |
| [numpy](https://github.com/numpy/numpy) | `2.5.1` | `2.5.2` |
| [platformdirs](https://github.com/tox-dev/platformdirs) | `4.11.0` | `4.11.2` |
| [setuptools](https://github.co
[…]
