+++ Documentation/requirements.txt
@@ -6,12 +6,12 @@ babel==2.18.0
 beautifulsoup4==4.15.0
 bleach==6.4.0
 certifi==2026.7.22
-charset-normalizer==3.4.9
+charset-normalizer==3.5.0
 click==8.4.2
 contourpy==1.3.3
 cycler==0.12.1
 defusedxml==0.7.1
-docutils==0.22.4
+docutils==0.23
 fastjsonschema==2.22.1
 fonttools==4.63.0
 h11==0.16.0
@@ -37,17 +37,17 @@ mistune==3.3.4
 mypy_extensions==1.1.0
 nbclient==0.11.0
 nbconvert==7.17.1
-nbformat==5.10.4
+nbformat==5.11.0
 nblink==0.1.12
 nbsphinx==0.9.8
 nbsphinx-link==1.4.1
-numpy==2.5.1
+numpy==2.5.2
 packaging==26.3
 pandocfilters==1.5.1
 parso==0.8.7
 pathspec==1.1.1
 pillow==12.3.0
-platformdirs==4.11.0
+platformdirs==4.11.2
 pybtex==0.26.1
 pybtex-docutils==1.0.3
 pycodestyle==2.14.0
@@ -60,11 +60,11 @@ referencing==0.37.0
 requests==2.34.2
 roman-numerals==4.1.0
 rpds-py==2026.6.3
-setuptools==83.0.0
+setuptools==84.0.0
 six==1.17.0
 sniffio==1.3.1
 snowballstemmer==3.1.1
-soupsieve==2.9.1
+soupsieve==2.9.2
 Sphinx==9.1.0
 sphinx-copybutton==0.5.2
 sphinx_collapse==0.1.4
@@ -81,7 +81,7 @@ sphinxcontrib-qthelp==2.0.0
 sphinxcontrib-serializinghtml==2.0.0
 sphinxcontrib-youtube==1.5.0
 tinycss2==1.5.1
-tornado==6.5.7
+tornado==6.5.8
 tqdm==4.70.0
 traitlets==5.16.1
 typing_extensions==4.16.0
