### 2026-08-20 4a581aa
added ClassMethodFunction
