### 2026-08-20 bd23349
setup for class method function
