### 2026-08-20 fb4f009
update in ClassMethodFunction
