+++ Utilities/FunctionInterface/ClassMethodFunction.m
@@ -13,6 +13,7 @@ classdef ClassMethodFunction < Function
         %% helper
         %
         functionHandler
+        isstatic
         
     end
 
@@ -38,41 +39,42 @@ classdef ClassMethodFunction < Function
             if isempty(fn.className)
                 error('ClassMethodFunction:MissingClassName','No className specified');
             end
+
             if isempty(fn.methodName)
                 error('ClassMethodFunction:MissingMethodName','No methodName specified');
             end
-            % Try to instantiate object using supplied parameters (if any).
-            % try
-                if ~isempty(fn.parameters)
-                    fn.obj = feval(fn.className, fn.parameters);
-                else
+            
+            if ~isempty(fn.parameters)
+                
+                fn.obj = feval(fn.className, fn.parameters);
+                
+            else
+                
+                mc = meta.class.fromName(fn.className);
+                idx = find(strcmp({mc.MethodList.Name}, fn.methodName), 1);
+                isStatic = ~isempty(idx) && mc.MethodList(idx).Static;
+
+                if ~isStatic
                     fn.obj = feval(fn.className);
+                else
+                    fn.obj [];
                 end
-            % catch
-            %     % If construction fails, assume the target is a static/class method.
-            %     fn.obj = [];
-            % end
-            % Set the function handler to call the method on the object (instance method)
-            % or a class/static method.
+                
+            end
+
             if ~isempty(fn.obj)
                 fn.functionHandler = @(varargin) fn.obj.(fn.methodName)(varargin{:});
             else
                 % static/class method handle like @() ClassName.methodName(...)
                 fn.functionHandler = str2func([fn.className '.' fn.methodName]);
             end
+            
         end
         
         function y = eval(fn, varargin)
-            % Use the prepared function handler if available.
-            if isempty(fn.functionHandler)
-                fn = fn.setupFunctionHandler();
-            end
-            if isa(fn.functionHandler, 'function_handle')
-                y = fn.functionHandler(varargin{:});
-            else
-                % Fallback to calling a method on this object (legacy behavior).
-                y = fn.(fn.methodName)(varargin{:});
-            end
+
+            y = fn.functionHandler(varargin{:});
+
         end
         
     end
