### 2026-08-20 c428713
added setupParameters function
