### 2026-08-20 68e4931
added example for classmethod function interface
