### 2026-08-20 71a5ea7
more in setupParameters
