+++ Utilities/FunctionInterface/setupParameters.m
@@ -1,7 +1,16 @@
 function fn = setupParameters(fn, parameters)
 
-    if isstruct(parameters) && numel(parameters) > 1
-        parameters = num2cell(parameters);
+    if isempty(parameters)
+        return
+    end
+    
+    if isstruct(parameters)
+        if numel(parameters) > 1
+            parameters = num2cell(parameters);
+        else
+            parameters = {parameters};
+            fn = setupParameters(fn, parameters);
+        end
     end
     
     for iparam = 1 : numel(parameters)
