### 2026-08-20 7106ce6
fix and clean-up in ClassMethodFunction
