### 2026-08-20 075e970
Make evalObj BattMo update schedule
