### 2026-08-21 d40e0db
Correct description of the bgfactor
