+++ Electrochemistry/ElectrolyteInputParams.m
@@ -5,38 +5,38 @@ classdef ElectrolyteInputParams < ComponentInputParams
     properties
 
         species % Structure given properties of the main ion (Lithium)
-        
+
         %% Standard parameters
 
         density              % the mass density of the material (symbol: rho)
         ionicConductivity    % a function to determine the ionic conductivity of the electrolyte under given conditions (symbol: kappa)
-        diffusionCoefficient % a function to determine the diffusion coefficient of a molecule in the electrolyte under given conditions (symbol: D)        
+        diffusionCoefficient % a function to determine the diffusion coefficient of a molecule in the electrolyte under given conditions (symbol: D)
         bruggemanCoefficient % the coefficient for determining effective transport parameters in porous media (symbol: beta)
-        
+
         % Region support for bruggeman coefficient
         useRegionBruggemanCoefficients % Set to true if the electrolye region for each component should get a specific
                                        % Bruggeman coefficient, as given by regionBruggemanCoefficients. Default value is false
-        bgfactor % Constant multiplicative factor applied to each region Bruggeman coefficient
+        bgfactor % Constant multiplicative factor applied linearly to ionic conductivity and diffusivity
         regionBruggemanCoefficients % Bruggeman coefficients for each region, given as a structure with fields
-                                    % - NegativeElectrode 
-                                    % - PositiveElectrode 
-                                    % - Separator 
+                                    % - NegativeElectrode
+                                    % - PositiveElectrode
+                                    % - Separator
         regionTags % Tags that identifies each grid cell for the electrolyte (negative and positive electrodes,
                    % separator). It is used only when useRegionBruggemanCoefficients is true. This property is typically setup by the grid constructor.
-        
+
         thermalConductivity  % Intrinsic Thermal conductivity of the electrolyte
         specificHeatCapacity % Specific Heat capacity of the electrolyte
 
         nominalEthyleneCarbonateConcentration % only used if a SEI layer is included in the model
-        
+
         %% Advanced parameters
 
         volumeFraction
         effectiveThermalConductivity    % (account for volume fraction)
         effectiveVolumetricHeatCapacity % (account for volume fraction and density)
-        
+
         use_thermal
-        
+
     end
 
     methods
@@ -46,7 +46,7 @@ classdef ElectrolyteInputParams < ComponentInputParams
             jsonstruct = setDefaultStructField(jsonstruct, 'useRegionBruggemanCoefficients', false);
             jsonstruct = setDefaultStructField(jsonstruct, 'bgfactor', 1);
             jsonstruct = setDefaultStructField(jsonstruct, {'species', 'chargeNumber'}, 1);
-            
+
             inputparams = inputparams@ComponentInputParams(jsonstruct);
 
         end
