### 2026-08-21 4d8ec6c
added support for generic reaction rates
