### 2026-08-21 c967687
added example with micro kinetics in electrolyser
