+++ Electrolyser/Examples/runElectrolyserWithKinetic.m
@@ -1,0 +1,168 @@
+%% Alkaline Membrane Electrolyser With Micro-Kinetics
+%
+% This script is essentially the same as |runElectrolyser| except that we use the kinetic model
+
+%% Setup input
+% Setup the physical properties for the electrolyser using json input file :battmofile:`alkalineElectrolyser.json<Electrolyser/Parameters/alkalineElectrolyser.json>`
+
+jsonstruct_material = parseBattmoJson('Electrolyser/Parameters/alkalineElectrolyser.json');
+jsonstruct_geometry = parseBattmoJson('Electrolyser/Parameters/electrolysergeometry1d.json');
+
+jsonstruct = mergeStructs({jsonstruct_material, ...
+                           jsonstruct_geometry});
+
+%%
+% We define shortcuts for the different submodels.
+inm = 'IonomerMembrane';
+her = 'HydrogenEvolutionElectrode';
+oer = 'OxygenEvolutionElectrode';
+ptl = 'PorousTransportLayer';
+exr = 'ExchangeReaction';
+ctl = 'CatalystLayer';
+
+%%
+% We add the kinetic functions to the json structure. The kinetic functions are defined in a separate json file
+% :battmofile:`kineticfunctions.json<Electrolyser/Parameters/kineticfunctions.json>`
+
+jsonstruct_kinetics  = parseBattmoJson('Electrolyser/Parameters/kineticfunctions.json');
+
+jsonstruct.(her).(ctl).genericRateFunctions = jsonstruct_kinetics;
+
+simsetup = setupElectrolyserSimulation(jsonstruct);
+
+%%
+% We use the function :battmo:`rampupControl` to increase the current linearly in time
+
+total = 10*hour;
+
+n   = 100;
+dt  = total/n;
+dts = rampupTimesteps(total, dt, 10);
+
+controlI = -3*ampere/(centi*meter)^2; % if negative, O2 and H2 are produced
+tup      = total;
+srcfunc  = @(time) rampupControl(time, tup, controlI, 'rampupcase', 'linear');
+control  = struct('src', srcfunc);
+
+step = struct('val', dts, 'control', ones(numel(dts), 1));
+schedule = struct('control', control, 'step', step);
+
+simsetup.schedule = schedule;
+
+simsetup.model.verbose = true;
+
+%% Run the simulation
+
+states = simsetup.run();
+
+%% Visualize the results
+%
+% The results contain only the primary variables of the system (the unknwons that descrive the state of the system). We
+% use the method :code:`addVariables` to add all the intermediate quantities that are computed to solve the equations
+% but not stored automatically in the result.
+
+model = simsetup.model;
+
+for istate = 1 : numel(states)
+    states{istate} = model.addVariables(states{istate});
+end
+
+
+
+%%
+% We extract the time, voltage and current values for each time step
+
+time = cellfun(@(state) state.time, states);
+E    = cellfun(@(state) state.(oer).(ptl).E, states);
+I    = cellfun(@(state) state.(oer).(ctl).I, states);
+
+%%
+% We plot the results for the voltage and current
+
+set(0, 'defaultlinelinewidth', 3)
+set(0, 'defaultaxesfontsize', 15)
+
+figure
+subplot(2, 1, 1)
+plot(time/hour, E)
+xlabel('time [hour]');
+ylabel('voltage');
+title('Polarisation curve');
+
+subplot(2, 1, 2)
+plot(time/hour, -I/(1/(centi*meter)^2));
+xlabel('time [hour]');
+ylabel('Current [A/cm^2]');
+title('Input current')
+
+%% pH distribution plot
+%
[…]
