### 2026-08-21 eb5dc03
added example in test suite
