+++ Tests/TestExamples/TestRunExamples.m
@@ -31,6 +31,7 @@ classdef TestRunExamples < matlab.unittest.TestCase
                     'runBattery1DOptimize'                            , ...
                     'runParameterIdentification'                      , ...
                     'runElectrolyser'                                 , ...
+                    'runElectrolyserWithKinetics'                     , ...
                     'runConductorBlockExample'                        , ...
                     'runRegionBruggemanExample'                       , ...
                     'runOnlyThermal'                                  , ...
