### 2026-08-21 12bcfd7
added support in IridiumCatalystLayer for generic kinetics
