+++ Utilities/FunctionInterface/setupParameters.m
@@ -3,21 +3,12 @@ function fn = setupParameters(fn, parameters)
     if isempty(parameters)
         return
     end
-    
-    if isstruct(parameters)
-        if numel(parameters) > 1
-            parameters = num2cell(parameters);
-        else
-            parameters = {parameters};
-            fn = setupParameters(fn, parameters);
-        end
-    end
-    
-    for iparam = 1 : numel(parameters)
 
-        param = parameters{iparam};
-        
-        fn.(param.name) = param.value;
+    fdnames = fieldnames(parameters);
+    for iparam = 1 : numel(fdnames)
+
+        fdname = fdnames{iparam};
+        fn.(fdname) = parameters.(fdname);
         
     end
 
