### 2026-08-24 9f4e501
Update geometry terminology
