+++ Battery/BatteryGeometry/setupBatteryGridFromJson.m
@@ -28,15 +28,15 @@ function [inputparams, gridGenerator] = setupBatteryGridFromJson(inputparams, js
 
         gen.xlength = xlength;
 
-        gen.sepnx  = jsonstruct.Separator.N;
-        gen.nenx   = jsonstruct.NegativeElectrode.Coating.N;
-        gen.penx   = jsonstruct.PositiveElectrode.Coating.N;
+        gen.sepnx  = jsonstruct.Separator.numberOfDiscreteCells;
+        gen.nenx   = jsonstruct.NegativeElectrode.Coating.numberOfDiscreteCells;
+        gen.penx   = jsonstruct.PositiveElectrode.Coating.numberOfDiscreteCells;
 
         if inputparams.NegativeElectrode.include_current_collectors
-            gen.ccnenx = jsonstruct.NegativeElectrode.CurrentCollector.N;
+            gen.ccnenx = jsonstruct.NegativeElectrode.CurrentCollector.numberOfDiscreteCells;
         end
         if inputparams.PositiveElectrode.include_current_collectors
-            gen.ccpenx = jsonstruct.PositiveElectrode.CurrentCollector.N;
+            gen.ccpenx = jsonstruct.PositiveElectrode.CurrentCollector.numberOfDiscreteCells;
         end
 
         if isfield(jsonstruct.Geometry, 'faceArea')
@@ -66,11 +66,11 @@ function [inputparams, gridGenerator] = setupBatteryGridFromJson(inputparams, js
                                    jsonstruct.(pe).(co).thickness ; ...
                                    jsonstruct.(pe).(cc).thickness];
 
-        gen.sep_nz   = jsonstruct.(sep).N;
-        gen.ne_co_nz = jsonstruct.(ne).(co).N;
-        gen.pe_co_nz = jsonstruct.(pe).(co).N;
-        gen.ne_cc_nz = jsonstruct.(ne).(cc).N;
-        gen.pe_cc_nz = jsonstruct.(pe).(cc).N;
+        gen.sep_nz   = jsonstruct.(sep).numberOfDiscreteCells;
+        gen.ne_co_nz = jsonstruct.(ne).(co).numberOfDiscreteCells;
+        gen.pe_co_nz = jsonstruct.(pe).(co).numberOfDiscreteCells;
+        gen.ne_cc_nz = jsonstruct.(ne).(cc).numberOfDiscreteCells;
+        gen.pe_cc_nz = jsonstruct.(pe).(cc).numberOfDiscreteCells;
 
 
         gen.tab_width     = jsonstruct.(geom).tab.width;
@@ -107,11 +107,11 @@ function [inputparams, gridGenerator] = setupBatteryGridFromJson(inputparams, js
         zlength(5) = jsonstruct.PositiveElectrode.CurrentCollector.thickness;
         gen.zlength = zlength;
 
-        gen.sep_nz   = jsonstruct.Separator.N;
-        gen.ne_co_nz = jsonstruct.NegativeElectrode.Coating.N;
-        gen.pe_co_nz = jsonstruct.PositiveElectrode.Coating.N;
-        gen.ne_cc_nz = jsonstruct.NegativeElectrode.CurrentCollector.N;
-        gen.pe_cc_nz = jsonstruct.PositiveElectrode.CurrentCollector.N;
+        gen.sep_nz   = jsonstruct.Separator.numberOfDiscreteCells;
+        gen.ne_co_nz = jsonstruct.NegativeElectrode.Coating.numberOfDiscreteCells;
+        gen.pe_co_nz = jsonstruct.PositiveElectrode.Coating.numberOfDiscreteCells;
[…]
