+++ Documentation/utils/publishExamplesXML.m
@@ -100,7 +100,7 @@ function run_publish(filename, opt)
 end
 
 %{
-Copyright 2009-2021 SINTEF Digital, Mathematics & Cybernetics.
+Copyright 2009-2026 SINTEF Digital, Mathematics & Cybernetics.
 
 This file is part of The MATLAB Reservoir Simulation Toolbox (MRST).
 
