### 2026-08-25 #11114 @augustjohansson
> Hi [@augustjohansson](https://github.com/augustjohansson),
>
> Thanks for the detailed responses.
>
>     * I can confirm that 1 and 6 are fixed.

Great!

>     * For 2, 3, 5, I think a documentation refresh is needed. The public available documentation is not reflecting the changes.

Thanks. This slipped through. The documentation is updated now.

>     * I followed the instructions to install the git lfs, but it does not seem to solve the problem. I am not sure if this is due to the setup of my laptop. If the other reviewers are not encountering the same issue, it may be specific to my local environment.

The reason for me suspecting git lfs is because the crameri is provided as a mat file which is tracked by git lfs. If there's of any help, I have
```
$ ls -l `find . -type f -name CrameriColourMaps7.0.mat`
-rw-rw-r-- 1 august august 202057 Nov 13  2024 ./BattMo/Utilities/Visualization/Colormaps/crameri/CrameriColourMaps7.0.mat
```
What's the error you're getting?
