### 2026-08-27 #11114 @yuefan98
Hi @augustjohansson,

The code you shared is really helpful for me to troubleshoot. It turns out I just need to run `git lfs pull` to download the actual crameri file since I didn't have git lfs install initially. I will keep exploring the software to see if there are any other problems, but so far I think everything is great with the revision.
