### 2026-08-27 #11114 @augustjohansson
Hi,

Thanks! I'm happy you got it working. I will add a note on `git lfs pull`  to the readme. I will also create a CONTRIBUTING.md and mention this in the readme.

August
