### 2026-08-27 #11114 @augustjohansson
Now the readme is updated with a link to the contributing.rst file [here](https://github.com/BattMoTeam/BattMo/blob/main/contributing.rst).
