### 2026-09-01 #11114 @yuefan98
Hi @augustjohansson,

I tested most of the example notebooks, and they worked really well. However, I noticed that the documentation does not always align with the latest implementation. For example, the ⁠[current documentation](https://battmo.org/BattMo/nblinks/runImpedanceScript.html#Impedance-computation) for EIS simulations seems to be outdated. The documented code gives me the following error:

`Error using ImpedanceSolver
Too many input arguments.

Error in test (line 24)
    impsolv = ImpedanceSolver(inputparams, 'soc', soc, 'computeSteadyState', false);`

This issue is resolved when using the example notebook included with the software (It seems there is a recent syntax update).

It can be difficult for new users to keep track of all the latest changes. Therefore, I would suggest another review of the documentation to make sure it is up to date and consistent with the latest version of the software.

Other than that, I think everything looks good from my side.
