### 2026-09-06 BattMoTeam/BattMo#88 @ElektrikAkar
JOSS paper comments
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

In this issue I would like to list the minor issues or typos in the paper:

1. **Front matter**: affiliations could have city information (e.g. Oslo).
2. **Summary (L56)**:  "**S**ilicon"→ *s*ilicon  (not capitilised).
3. **Summary (L59)**: "**relationship**" → *relationships*.
4. **Summary (L62)**: "The solver in BattMo uses automatic differentiation and **support** adjoint computation" → *supports*.
5. **State of the field (L91)**: could the authors soften "PyBaMM … has just recently (2025) allowed for simulations in 2D and 3D"? PyBaMM has long offered 1+1D/2+1D thermal pouch models; but 3D is new as authors claimed.
6. **State of the field (L91)**: "BattMo's foundation is designed for coupled electro-chemical-thermal simulations in 3D, **as well calibration** and optimization" → *as well as calibration*.
7. **Functionality overview (L115)**: "Material database for NMC, **NCA**, LFP, LNMO, SiGr, electrolytes, etc",  I could not find the NCA parameter set in `ParameterData/`
8. **Battery format library (L125) and Graph based model development (L133)**: Figures 1 and 2 are never referenced from the text.
9. **BattMo family (L141)**: "The following software **include** the BattMo family" → are *included in* the BattMo family (or *comprise*).
10. **Acknowledgements (L173)**: "BattMo has ~~been~~ mainly been developed and used in projects".
11. **References**: unused bibliography entries `dandeliion`, `comsol`, `MRST:2025` are in `paper.bib` but never cited. DandeLiion  and COMSOL are good competitors so they should be cited. Also [SLIDE](https://github.com/Battery-Intelligence-Lab/SLIDE) doesn't have 3D simulation but probably it will have soon; however, it is not really related at this point.
12. **References**:  You can write {Ah}, {Li}, {Python}, so the rendered references do not lose capitalisation.

Also in both `readme.rst` and `json.rst`  → you have "that that", so there is extra that.
