### 2026-09-06 BattMoTeam/BattMo#89 @ElektrikAkar
Some files named NMC are parameterised with LCO
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

When I look at "lithium_ion_battery_nmc_*.json" files, I see some LCO parameters for positive [active material](https://github.com/BattMoTeam/BattMo/blob/c67173b449a012e1f5c147c521a4a64b99949a6f/ParameterData/BatteryCellParameters/LithiumIonBatteryCell/lithium_ion_battery_nmc_graphite.json#L74). Or [computeOCP_LCO](https://github.com/BattMoTeam/BattMo/commit/a8e4c6b9454d5a52c8be328acf39fae968e2b950#diff-4f6984a9cc18839aa7ad6bbb40d146ed1ee4843126f247661275775c50409a71)

Could authors clarify whether this is a mislabelling or a deliberate placeholder (where you are only importing some common parameters)?
