### 2026-09-06 BattMoTeam/BattMo#90 @ElektrikAkar
LFS for battmo.zip
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

`battmo.zip` experiences the same issue @yuefan98 mentioned. `.github/workflows/createZipArchive.yml` checks out with `submodules: recursive` but without `lfs: true`. It might also be worth having `startupBattMo` detect this and emit a clear warning, since users who clone without `git lfs` experience the same issue.
