### 2026-09-06 BattMoTeam/BattMo#91 @ElektrikAkar
Installation docs: missing Git LFS and supported MATLAB versions
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

Small gaps in https://battmo.org/BattMo/installation.html

1. The Git LFS prerequisite is mentioned only in the GitHub README, not on the documentation installation page.
2. I could not find a statement of the minimum/supported MATLAB version.
3. Contributing statement could be accessible from the website.
