### 2026-09-06 BattMoTeam/BattMo#92 @ElektrikAkar
TestBatteryP2D and TestBatteryPND are broken and excluded from CI
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

`TestBatteryP2D.m` hardcodes a developer path:
https://github.com/BattMoTeam/BattMo/blob/c67173b449a012e1f5c147c521a4a64b99949a6f/Tests/TestExamples/TestBatteryP2D.m#L19

- `TestBatteryPND.m` errors with `Unrecognized property 'EffectiveElectricalConductivity' for class 'CurrentCollectorInputParams'`

-  `runImpedanceChen.m` can have an `assert` statement with tolerance such as `assert(norm(pb - battmo) / norm(pb) < 1e-3);` in `runChen2020.m`
