### 2026-09-06 BattMoTeam/BattMo#93 @ElektrikAkar
startupBattMo path list is stale
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

In `startupBattMo.m`, the `dirnames` list includes `Physics` and `JuliaBridge`, neither of which exists at the repository root (the Julia bridge lives under `Utilities/JuliaBridge`, which is already covered by the `Utilities` entry), while `OxideElectrolyte/` is missing from the list. The MRST module list also contains `'solvers'` twice.

Hardcoded developer paths in `Documentation/utils/exportMLXnotebooks.m`, `Examples/Advanced/runNonLinearSolver.m`, `OxideElectrolyte/testoxidepm.m`, `OxideElectrolyte/testoxidegraphpm.m`, and `ProtonicMembrane/utils/testgraphpm.m`.

Authors could check other instances of hardcoded paths. And/or write some automated tests to check for possible hardcoded paths.
