### 2026-09-06 BattMoTeam/BattMo#94 @ElektrikAkar
Citation standardisation
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

In some parameter files I see proper citations, for example `LiC6_Xu.json` has

`  "description" : "Data from Xu et al (2015) doi:10.1016/j.energy.2014.11.073",
`

While some others don't. It would be nice to have a standardised way/entry for citing the sources.

`computeOCP_Graphite_Torchio.m` and `computeOCP_silicon.m` cite "[1]"/"[ref 4]" that are never defined in the file, and `computeOCP_silicon.m` notes that "coeff2 are the one for graphite".
