### 2026-09-06 BattMoTeam/BattMo#95 @ElektrikAkar
Residual comments
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

There are commented out code blocks/unused parameters in some files. Such as  `computeOCP_silicon.m`:

```
    %OCP = refOCP + (T - Tref) .* dUdT;
    dUdT = 0;
    OCP = refOCP;
```

As git already keeps history of changes, I would suggest deleting residuals instead of commenting them out.
