### 2026-09-06 BattMoTeam/BattMo#96 @ElektrikAkar
Other minor issues
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

Other minor issues I would like to write here:

- Naming of variables: For example in `BatteryGeneratorP2D.m` I see following variables.

```
        sepnx  = 10; % discretization number for negative current collector (default = 10)
        nenx   = 10; % discretization number for negative active material (default = 10)
        penx   = 10; % discretization number for separator (default = 10)
        ccnenx = 10; % discretization number for positive current collector (default = 10)
        ccpenx = 10; % discretization number for positive active material (default = 10)
```
Why `penx` is discretization for seperator instead of positive active material as `nenx`? Also why `sepnx` is negative current collector? `sep` feels like seperator. This could be misleading for the new users.

----------------

- API typos: `ComputationalGraph.adjencyMatrix` (→ adjacency), `BaseModel.getComputationalGrapInteractiveTool` (missing "h").
- BATMAX grant number: `readme.rst` (and the JOSS paper) say 101104013, `Documentation/index.rst` says 101104031
- Some files also have MRST licence block (e.g. `Tests/TestExamples/TestBatteryP2D.m`, `Utilities/Adjoint/evalObjectiveBattmo.m`); can authors clarify if this is intentional?
- All 8 files in `ParameterData/MaterialProperties/OrganicLiPF6Solutions/` end in an empty `%{ %}` block.
- 3 committed `.ipynb_checkpoints` files.
- "An web app" (`programminginterfaces.rst`).
- There are some comments in French: https://github.com/BattMoTeam/BattMo/blob/92a670e832b348d35e95b47c19498f9fa43630a5/Battery/EquivalentCircuitModel/calibration/FittingTime.m#L265
https://github.com/BattMoTeam/BattMo/blob/92a670e832b348d35e95b47c19498f9fa43630a5/Battery/EquivalentCircuitModel/calibration/FittingTime.m#L268
https://github.com/BattMoTeam/BattMo/blob/92a670e832b348d35e95b47c19498f9fa43630a5/Battery/EquivalentCircuitModel/calibration/FittingTime.m#L243
- %change here markers could be a parametrised script.
- Authors may consider adding a `CITATION.cff` so GitHub renders a cite box matching the Zenodo DOI.
