### 2026-09-06 BattMoTeam/BattMo#97 @ElektrikAkar
Warning statements may hint issues
Hi authors,

I am opening this issue for JOSS review [openjournals/joss-reviews#11114](https://github.com/openjournals/joss-reviews/issues/11114).

There are many lines of code that cause warnings. Could authors check all the warnings, then manually suppress unintended warnings.

For example

```
includeDoubleLayer = false;

if includeDoubleLayer

    jsonstruct.(ne).(co).(am).(itf).useDoubleLayerCapacity = true;
    jsonstruct.(ne).(co).(am).(itf).doubleLayerCapacitance = 0.2;

end

```

Inside if statement is not reachable.

Or residual `keyboard` comments:

https://github.com/BattMoTeam/BattMo/blob/92a670e832b348d35e95b47c19498f9fa43630a5/Examples/Experimental/runJellyRollLinearSolve.m#L114

Or too few arguments for this `error` command:

https://github.com/BattMoTeam/BattMo/blob/92a670e832b348d35e95b47c19498f9fa43630a5/Examples/Experimental/runJellyRollLinearSolve.m#L188

Most importantly overridden variables like in `TestBatteryP2D.m`

```
            validate = serial & has_python;
            validate = false;
```

If some of these warnings are deliberate then they should accompany `%OK<....>` statements to suppress the warning. Otherwise, they should be addressed.
